% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{ahrqComorbidAll}
\alias{ahrqComorbidAll}
\title{AHRQ comorbidities, with HTN, CHF and renal failure subgroups}
\format{list of character vectors, each named by co-morbidity}
\source{
\url{http://www.hcup-us.ahrq.gov/toolssoftware/comorbidity/comorbidity.jsp}
}
\description{
This mapping of comorbidities to ICD-9 codes is dervied directly
  from SAS code provided by AHRQ, and translated into this R data structure.
  Beyond ahrqComorbid, this includes all the HTN, CHF and renal subgroups,
  not rolled into their parent categories.
}
\keyword{datasets}

