% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-test.R
\name{all_identical}
\alias{all_identical}
\title{allow microbenchmark to compare multiple results}
\usage{
all_identical(x)
}
\arguments{
\item{x}{list of values to compare for identity, e.g. results from evaluated
expression in \code{microbenchmark::microbenchmark}}
}
\description{
allow microbenchmark to compare multiple results
}
\keyword{internal}
