% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/util.R
\name{asCharacterNoWarn}
\alias{asCharacterNoWarn}
\alias{as_char_no_warn}
\title{convert to character vector without warning}
\usage{
asCharacterNoWarn(x)

as_char_no_warn(x)
}
\arguments{
\item{x}{vector, typically numeric or a factor}
}
\value{
character vector
}
\description{
convert to character vector without warning
}
\keyword{internal}

