% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{attr_decimal_diag}
\alias{attr_decimal_diag}
\title{Set ICD short-form diagnosis code attribute}
\usage{
attr_decimal_diag(x, value = TRUE)
}
\description{
Doing this in an R function doesn't work for 'void' equivalent, and does a
copy if the updated object is returned.
}
\examples{
j <- "100"
attr(j, "icd_short_diag") <- TRUE
j
attr(j, "icd_short_diag") <- FALSE
j
icd:::attr_decimal_diag(j)
as.icd_decimal_diag(j)
# if pryr is installed, use address and refs to see what is going on
}
\keyword{attribute}
\keyword{internal}
