% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{attr_short_diag}
\alias{attr_short_diag}
\title{Set short diagnosis flag in C++}
\usage{
attr_short_diag(x, value = TRUE)
}
\arguments{
\item{x}{Any R object}

\item{value}{\code{TRUE} or \code{FALSE}}
}
\description{
Set short diagnosis flag in C++
}
\keyword{attribute}
\keyword{internal}
