% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{chapter_to_desc_range}
\alias{chapter_to_desc_range}
\title{Parse (sub)chapter description with range}
\usage{
chapter_to_desc_range(x, re_major)
}
\arguments{
\item{x}{vector of descriptions followed by ICD code ranges}
}
\value{
list of two-element character vectors, the elements being named
  'start' and 'end'.
}
\description{
Parse (sub)chapter description with range
}
\keyword{internal}
\keyword{manip}

