% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children.R
\name{children_defined}
\alias{children_defined}
\alias{children_defined.icd10cm}
\title{defined children of ICD codes}
\usage{
children_defined(x)

\method{children_defined}{icd10cm}(x, short_code = guess_short(x),
  warn = FALSE)
}
\arguments{
\item{warn}{single logical value, if \code{TRUE} will generate warnings when
some input codes are not known ICD-10-CM codes}

\item{use_cpp}{single logical flag, whether to use C++ version}
}
\description{
Find defined ICD-10 children based on 2016 ICD-10-CM list. "defined" may be a
three digit code, or a leaf node. This is distinct from 'billable'.
}
\section{Methods (by class)}{
\itemize{
\item \code{icd10cm}: Internal function to get the children of
ICD-10 code(s)
}}

\keyword{internal}
