% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbid.R
\name{comorbid_common}
\alias{comorbid_common}
\title{Internal function to calculate co-morbidities.}
\usage{
comorbid_common(x, map, visit_name = NULL, icd_name, return_df = FALSE,
  comorbid_fun = comorbidMatMul, ...)
}
\arguments{
\item{comorbid_fun}{function i.e. the function (not character string) to be
called to do the comorbidity calculation}
}
\description{
This common comorbidity calculation code does not depend on ICD type. There
is some type conversion so the map and input codes are all in 'short' format,
fast factor generation, then fast comorbidity assignment.
}
\keyword{internal}
