% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explaintable.R
\name{condense_explain_table_to_majors}
\alias{condense_explain_table_to_majors}
\title{condense to only major codes}
\usage{
condense_explain_table_to_majors(x, hierarchy)
}
\description{
condense to major codes, even when major not present in the input data
}
\keyword{internal}

