% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain-table.R
\name{condense_explain_table_worker}
\alias{condense_explain_table_worker}
\title{generate condensed code and condensed number columns}
\usage{
condense_explain_table_worker(x)
}
\value{
details for condesable rows
}
\description{
generate condensed code and condensed number columns
}
\keyword{internal}

