% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/comorbid.R
\name{cr}
\alias{cr}
\title{get sequence of column indices of comorbidities}
\usage{
cr(cbd)
}
\arguments{
\item{cbd}{matrix or data.frame of comorbidities}
}
\description{
get sequence of column indices of comorbidities
}
\keyword{internal}

