% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R, R/z-deprecated.R
\name{decimal_to_parts}
\alias{decimal_to_parts}
\alias{decimal_to_parts.icd9}
\alias{decimal_to_parts.character}
\alias{decimal_to_parts.icd10}
\alias{icd_decimal_to_parts.icd9}
\alias{icd_decimal_to_parts}
\title{Convert decimal ICD codes to component parts}
\usage{
decimal_to_parts(x, mnr_empty = "")

\method{decimal_to_parts}{icd9}(x, mnr_empty = "")

\method{decimal_to_parts}{character}(x, mnr_empty = "")

\method{decimal_to_parts}{icd10}(x, mnr_empty = "")

icd_decimal_to_parts.icd9(...)

icd_decimal_to_parts(...)
}
\description{
Convert decimal ICD codes to component parts
}
\section{Methods (by class)}{
\itemize{
\item \code{icd9}: Convert decimal ICD-9 code to parts

\item \code{character}: Convert decimal ICD code to parts, guessing
ICD version

\item \code{icd10}: Convert decimal ICD-10 code to parts. This
shares almost 100% code with the ICD-9 version: someday combine the common
code.
}}

\keyword{internal}
\keyword{manip}
