% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-test.R
\name{do_extra_tests}
\alias{do_extra_tests}
\title{Set system environment to do extra tests}
\usage{
do_extra_tests(value = "true")
}
\description{
These extra tests are not run except for one CI job, and locally as needed.
The frequently run tests cover the core code, whereas the extra tests cover
things like parsing the ICD and comorbidity definitions.
}
\keyword{debugging}
\keyword{internal}
