% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.make_icd9cm_parse_leaf_fun}
\alias{.make_icd9cm_parse_leaf_fun}
\title{Return a function which returns specific ICD data}
\usage{
.make_icd9cm_parse_leaf_fun(year)
}
\description{
This is a good use of functional programming, although more hidden for people
exploring \CRANpkg{icd} code. The alternative would be pre-generating
(voluminous) extra code. Adding an additional year of ICD-10-CM data, for
example, becomes as simple as adding the appropriate URLs and file names to
the list structure \code{icd:::.icd10cm_sources}, found in
\code{R/icd-cm-sources.R}.
}
\keyword{internal}
