% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{download_icd_data}
\alias{download_icd_data}
\title{Download all the additional data at once}
\usage{
download_icd_data()
}
\description{
This may take ten minutes on a broadband connection. It will download and
parse WHO ICD-10, French, and Belgian codes and descriptions. It will also
get years 2014, 2015, 2017, and 2018 for ICD-10-CM (diagnostic codes), and
2014--2019 procedure codes. 2016 and 2019 diagnostic codes are included in
the package data.
}
\examples{
\dontrun{
setup_icd_data()
download_icd_data()
}
}
\seealso{
\code{\link{setup_icd_data}}
}
