% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_to_vec_flip}
\alias{env_to_vec_flip}
\title{return a new environment with names and values swapped}
\usage{
env_to_vec_flip(env)
}
\arguments{
\item{env}{environment with values being sequence numbers used to fill
returned vector}
}
\description{
return a new environment with names and values swapped
}
\keyword{internal}

