% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{env_var_is_false}
\alias{env_var_is_false}
\alias{env_var_is_true}
\alias{opt_env_is_false}
\alias{opt_env_is_true}
\alias{env_opt_is_false}
\alias{env_opt_is_true}
\title{Does a character environment variable have logical value?}
\usage{
env_var_is_false(x, unset = FALSE)

env_var_is_true(x, unset = FALSE)

opt_env_is_false(x, unset = FALSE)

opt_env_is_true(x, unset = FALSE)

env_opt_is_false(x, unset = FALSE)

env_opt_is_true(x, unset = FALSE)
}
\description{
If unset, then it is not false, so returns FALSE.
}
\section{Functions}{
\itemize{
\item \code{env_var_is_true}: Whether an environment variable has true-like
value

\item \code{opt_env_is_false}: Try \R option first for falsity, then
environment variable, before returning the \code{unset} value

\item \code{opt_env_is_true}: Try \R option first for truth, then environment
variable, before returning the \code{unset} value.

\item \code{env_opt_is_false}: Try environment variable then \R option,
before returning the \code{unset} value.

\item \code{env_opt_is_true}: Try environment variable then \R option,
before returning the \code{unset} value.
}}

\examples{
# exists, usually unset or "1"
icd:::env_var_is_true("RSTUDIO")
icd:::env_var_is_false("RSTUDIO")
!icd:::env_var_is_true("RSTUDIO")
!icd:::env_var_is_false("RSTUDIO")
# exists, but not true or false:
icd:::env_var_is_true("R_HOME")
icd:::env_var_is_false("R_HOME")
!icd:::env_var_is_true("R_HOME")
!icd:::env_var_is_false("R_HOME")
# try an option, fall back on an environment variable
!icd:::opt_env_is_false("R_HOME")
icd:::env_opt_is_true("doesn't exist")
getOption("verbose",
  default = env_var_is_true("CLICOLOR_FORCE")
)
getOption("verbose",
  default = env_var_is_true("ENV_VAR_DOES_NOT_EXIST")
)
}
\keyword{internal}
