% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{expand_range.icd10cm}
\alias{expand_range.icd10cm}
\title{Expand range of ICD-10 codes returning only defined codes in ICD-10-CM}
\usage{
\method{expand_range}{icd10cm}(start, end,
  short_code = guess_short(c(start, end)), defined, ...)
}
\arguments{
\item{start}{character vector of length one containing an ICD code}

\item{end}{character vector of length one containing an ICD code}
}
\description{
This will need generalizing to any list of officially defined codes, e.g. WHO
or other. There are so many permutations of alphanumeric characters after the
decimal place that it would be easy to generate many millions with simple
ranges, the vast majority of which would be undefined.
}
\keyword{internal}
