% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{expandRangeWorker}
\alias{expandRangeWorker}
\title{expand range worker function}
\usage{
expandRangeWorker(start, end, lookup, onlyReal, excludeAmbiguousStart,
  excludeAmbiguousEnd)
}
\description{
expand range worker function
}
\keyword{internal}

