% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-test.R
\name{expect_chap_present}
\alias{expect_chap_present}
\alias{expect_icd10_chap_present}
\alias{expect_icd10_sub_chap_present}
\alias{expect_icd9_chap_present}
\alias{expect_icd9_sub_chap_present}
\title{expect that a chapter with given title exists, case-insensitive}
\usage{
expect_chap_present(x, ver_chaps, info = NULL, label = NULL, ...)

expect_icd9_sub_chap_present(x, info = NULL, label = NULL, ...)

expect_icd9_chap_present(x, ...)

expect_icd10_sub_chap_present(x, ...)

expect_icd10_chap_present(x, ...)
}
\description{
expect that a chapter with given title exists, case-insensitive
}
\keyword{internal}

