% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-test.R
\name{expect_equal_no_icd}
\alias{expect_equal_no_icd}
\title{expect equal, ignoring any ICD classes}
\usage{
expect_equal_no_icd(object, expected, ...)
}
\description{
Strips any \code{icd} classes (but not others) before making comparison
}
\keyword{debugging}
\keyword{internal}

