% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_funcs.R
\name{explain_ccs}
\alias{explain_ccs}
\title{Get description of ccs codes}
\usage{
explain_ccs(ccs_list)
}
\arguments{
\item{ccs_list}{A vector of CCS values}
}
\description{
`explain_ccs` retuns a tibble of descriptions for a vector of
of CCS categories
}
\examples{
explain_ccs(122)
explain_ccs(c(1,246))
}
