% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{factor_fast}
\alias{factor_fast}
\title{fast factor generation test}
\usage{
factor_fast(x)
}
\description{
fast factor generation test
}
\examples{
\dontrun{
codes <- unname(unlist(icd9_map_ahrq))
codes <- c(codes, icd:::randomMajorCpp(1e7))
microbenchmark::microbenchmark(
  icd:::factor_fast(codes),
  icd:::factor_nosort(codes),
  factor(codes),
  times = 10
  )
# the R factor_nosort is about as fast as Rcpp version
}
}
\keyword{internal}
