% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{factor_nosort_rcpp}
\alias{factor_nosort_rcpp}
\title{Fast factor generation with hashing, no sorting}
\usage{
factor_nosort_rcpp(x, levels)
}
\arguments{
\item{x}{vector of strings}

\item{levels}{vector of levels}
}
\description{
Rcpp minimal factor creation
}
\keyword{internal}
\keyword{manip}
