% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{factor_split_na}
\alias{factor_split_na}
\title{factor quickly and split into matched and unmatched elements}
\usage{
factor_split_na(x, levels, factor_fun = factor_nosort_rcpp)
}
\description{
factor quickly and split into matched and unmatched elements
}
\examples{
icd:::factor_nosort_rcpp(c("1", NA)) # NA becomes a level
icd:::factor_nosort_rcpp(c("1", "2"), "1") # NA not a level, just dropped!
icd:::factor_nosort_rcpp(c("1", "2"), c("1", NA)) # NA IS a level
suppressWarnings(
  print(icd:::factor_nosort_rcpp(c("1", "2"), c("1", NA, NA)))
) # Two NA levels

x <- c("A", "B", "C", "d", "A", "C")
levels <- c("A", "B")
stopifnot(
  identical(factor_split_na(factor(x), levels),
            factor_split_na(x, levels))
)
y <- c("A", NA, "B", "A", NA)
yf <- factor(y)
yf_na <- factor(y, levels = c("A", NA, "B"), exclude = NULL)
stopifnot(
  identical(factor_split_na(y, "A"),
            factor_split_na(yf, "A"))
)
stopifnot(
  identical(factor_split_na(y, "A"),
            factor_split_na(yf_na, "A"))
)
}
\keyword{internal}
\keyword{manip}
