% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{factor_split_rcpp}
\alias{factor_split_rcpp}
\title{Split a factor into two based on desired levels}
\usage{
factor_split_rcpp(df, levels, visit_name, code_name)
}
\description{
Using C++ because I also want to quickly return the visits
  corresponding to the NA or non-NA factor elements.
}
\examples{
  v <- c("X", "Y")
  df <- data.frame(visit_id = factor(c("visit1", "visit2", "visit1")),
                   icd_code = factor(c("410", "0010", "E999")))
  icd:::factor_split_rcpp(df, "410", "visit_id", "icd_code")
}
\keyword{internal}
\keyword{manip}
