% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{factorSplit}
\alias{factorSplit}
\title{Split a factor into two based on desired levels}
\usage{
factorSplit(x, levels)
}
\description{
Using C++ because I also want to quickly return the visits
  corresponding to the NA or non-NA factor elements.
}
\examples{
  icd:::factorSplit(factor(c("A", "B")), c("A"))
  icd:::factorSplit(factor(c("A", "B")), c("A", "B"))
  icd:::factorSplit(factor(c("A", "B")), c("B"))
  icd:::factorSplit(factor(c("A", "B")), c("C"))
  icd:::factorSplit(factor(c("A", "B")), c(NA_STRING_))
  icd:::factorSplit(factor(c("A", "B")), character())
  icd:::factorSplit(factor(c("A", "B")), c("C"))
}
\keyword{internal}
\keyword{manip}
