% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastIntToStringRcpp}
\alias{fastIntToStringRcpp}
\title{Convert integers to strings as quickly as possible}
\usage{
fastIntToStringRcpp(x)
}
\arguments{
\item{x}{Vector of integers}
}
\value{
Vector of characters
}
\description{
Have tried R, sprintf with Rcpp and C++ standard library.
}
\keyword{internal}
\keyword{manip}
