% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-comorbid.R
\name{fetch_ahrq_sas}
\alias{fetch_ahrq_sas}
\title{get the SAS code from AHRQ}
\usage{
fetch_ahrq_sas(offline)
}
\description{
Get the SAS code from AHRQ and save in data-raw if not already there.
}
\keyword{internal}

