% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-file.R
\name{fetch_icd10cm_all}
\alias{fetch_icd10cm_all}
\alias{fetch_icd10cm_year}
\title{Fetch ICD-10-CM data from the CMS web site}
\usage{
fetch_icd10cm_all(verbose = FALSE, ...)

fetch_icd10cm_year(year = "2018", type = c("dx", "pcs"), ...)
}
\description{
YEAR-ICD10-Code-Descriptions has flat files, YEAR-ICD10-Code-Tables-Index has
XML
}
\keyword{internal}
