% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-rtf.R
\name{fetch_rtf_year}
\alias{fetch_rtf_year}
\title{Fetch RTF for a given year}
\usage{
fetch_rtf_year(year, offline = TRUE)
}
\arguments{
\item{year}{character vector of length one, e.g. "2011"}

\item{offline}{single logical value}
}
\description{
Will return NULL if offline and not available
}
\keyword{internal}

