% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-rtf.R
\name{fix_unicode}
\alias{fix_unicode}
\title{Fix unicode characters in RTF}
\usage{
fix_unicode(filtered)
}
\description{
fix ASCII/CP1252/Unicode horror: of course, some char defs are split over
lines... This needs care in Windows, or course. Maybe Mac, too?
}
\keyword{internal}
\keyword{manip}

