% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{fixSubchapterNa}
\alias{fixSubchapterNa}
\title{Fix NA sub-chapters in RTF parsing}
\usage{
fixSubchapterNa(x, start, end)
}
\description{
Fixes a couple of corner cases in parsing the 2011 ICD-9-CM RTF
}
\keyword{datagen}
\keyword{internal}
