% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_deprecated_data.R
\name{generate_deprecated_data}
\alias{generate_deprecated_data}
\title{Generate deprecated data}
\usage{
generate_deprecated_data(save_data = TRUE)
}
\arguments{
\item{save_data}{save the data in \code{data/deprecated.RData}, default is
\code{TRUE}}
}
\description{
Unfortunately, this is easiest (and has a chance of passing CRAN checks) if I
just duplicate data rather than trying to change the namespace
}
\keyword{internal}

