% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{generate_everything}
\alias{generate_everything}
\title{generate all package data}
\usage{
generate_everything()
}
\description{
Parses (and downloads if necessary) CDC annual revisions of
  ICD-9-CM to get the 'billable' codes. Also parses the AHRQ and Quan/Deyo
  comorbidity mappings from the source SAS data. Elixhauser and
  Quan/Elixhauser mappings are generated from transcribed codes.
}
\keyword{internal}

