% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sysdata.R
\name{generate_icd10_sources}
\alias{generate_icd10_sources}
\title{Generate list of data source information for ICD-10-CM diagnosis and
procedure codes}
\usage{
generate_icd10_sources(save_data = TRUE)
}
\description{
Generate list of data source information for ICD-10-CM diagnosis and
procedure codes
}
\seealso{
\url{https://www.cms.gov/Medicare/Coding/ICD10/}
}
\keyword{internal}
