% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sysdata.R
\name{generate_icd_chapters}
\alias{generate_icd_chapters}
\title{Generate ICD-9 and ICD-10 Chapter names and number ranges, transcribed from
the official definitons from WHO and extended by US CMS.}
\usage{
generate_icd_chapters(save_data = TRUE)
}
\arguments{
\item{save_data}{single logical value, if \code{TRUE} will save the generated
data}
}
\description{
Generate ICD-9 and ICD-10 Chapter names and number ranges, transcribed from
the official definitons from WHO and extended by US CMS.
}
\keyword{datasets}
\keyword{internal}
