% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-mappings-icd10.R
\name{generate_map_quan_charlson}
\alias{generate_map_quan_charlson}
\title{Generate Quan mapping for Charlson categories of ICD-10 codes
 Based on Quan's SAS lists, transcribed by @wmurphyrd
 @export}
\usage{
generate_map_quan_charlson()
}
\description{
Generate Quan mapping for Charlson categories of ICD-10 codes
 Based on Quan's SAS lists, transcribed by @wmurphyrd
 @export
}

