% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-mappings-icd10.R
\name{generate_map_quan_elix}
\alias{generate_map_quan_elix}
\title{generate ICD-10 Quan/Elixhauser maping}
\usage{
generate_map_quan_elix(save = TRUE)
}
\description{
generate ICD-10 Quan/Elixhauser maping
}
\details{
start with Quan's SAS code (in \code{data-raw}):
}

