% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-test.R
\name{generate_random_short_icd9}
\alias{generate_random_short_icd9}
\title{generate random ICD-9 codes}
\usage{
generate_random_short_icd9(n = 50000)
}
\description{
generate random ICD-9 codes
}
\keyword{datagen}
\keyword{debugging}
\keyword{internal}
