% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_spelling.R
\name{generate_spelling}
\alias{generate_spelling}
\title{Generate spelling exceptions}
\usage{
generate_spelling(out = ".aspell/words.pws", save_data = TRUE)
}
\arguments{
\item{out}{character scalar with path to save the \code{aspell} word list}

\item{save_data}{single logical value, if \code{TRUE} will save the generated
data}
}
\description{
This is a bit painful, but defining known spellings of names and technical
terms makes the \code{aspell} checks useful (and removes a NOTE in the
\code{R CMD check} which at least some of the CRAN people use.
}
\keyword{internal}

