% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mappings.R
\name{generate_uranium_pathology}
\alias{generate_uranium_pathology}
\title{generate uranium pathology data}
\usage{
generate_uranium_pathology(save_data = FALSE)
}
\arguments{
\item{save_data}{logical whether to save the result in the source tree.
Defaults to \code{FALSE}.}

\item{condense}{Deprecated. The map is not condensed by default. If required,
it can reduced to a smaller set of codes using \code{icd_condense}.}
}
\value{
invisibly returns the list of vectors, where a co-morbidity name is
  associated with a character vector of ICD-9 codes.
}
\description{
This is downloaded from WSU where it appears to be provided in the public
domain.
}
\keyword{internal}

