% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{generateSysData}
\alias{generateSysData}
\title{Generate sysdata.rda}
\usage{
generateSysData(sysdata.path = file.path("R", "sysdata.rda"), save = TRUE)
}
\description{
Generate correctly ordered look-up tables of numeric-only, V and E codes. This is
quick, but much too slow when it appears many times in a loop.
}
\keyword{internal}

