% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{get_cim10fr2019}
\alias{get_cim10fr2019}
\title{Localised synonym for \code{\link{get_icd10fr2019}}, with French column names}
\usage{
get_cim10fr2019()
}
\description{
Localised synonym for \code{\link{get_icd10fr2019}}, with French column names
}
\seealso{
\code{\link{get_icd10fr2019}}
}
