% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.R
\name{get_defined}
\alias{get_defined}
\title{Select only defined ICD codes}
\usage{
get_defined(x, short_code = guess_short(x), billable = FALSE, leaf = billable)
}
\arguments{
\item{x}{input vector or factor, possibly with an ICD class}

\item{short_code}{logical value, whether short-form ICD code}

\item{billable}{single logical value, identical to 'leaf'. Leaf is preferred
as most adaptations of WHO ICD codes are not oriented around money.}

\item{leaf}{single logical value, whether to limit return codes also by
whether they are billable, i.e. leaf nodes. This is really only designed
for use with ICD-9-CM, ICD-10-CM etc, since the WHO versions are not
designed for billing, but for public health and death reporting.}
}
\description{
Return only those codes which are heading or leaf (billable), specifying
whether codes are all short-form or all decimal-form
}
