% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{get_icd10be2014}
\alias{get_icd10be2014}
\title{Belgian ICD-10-BE}
\source{
\url{https://www.health.belgium.be/en/node/30433}
\url{https://www.health.belgium.be/sites/default/files/uploads/fields/fpshealth_theme_file/fy2017_reflist_icd-10-be.xlsx_last_updatet_28-07-2017_1.xlsx}
\url{https://www.health.belgium.be/fr/sante/organisation-des-soins-de-sante/hopitaux/systemes-denregistrement/icd-10-be}
\url{https://www.health.belgium.be/fr/fy2014reflisticd-10-bexlsx}
}
\description{
This is based heavily on ICD-10-CM. 2014 is identical, with translations for
most of the codes into Dutch and French. 2017 has about a hundred additional
code definitions over ICD-10-CM 2017. The 2014 data also has the interesting
fields for gender specificity of a given code, and whether it is permissible
as Present-on-Arrival (POA).
}
\seealso{
\code{\link{get_icd10be2017}} \code{\link{get_icd10be2014_pc}}
  \code{\link{get_icd10be2017_pc}}
}
\keyword{datasets}
