% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{get_icd10be2017}
\alias{get_icd10be2017}
\title{ICD-10-BE 2017}
\description{
ICD-10-BE 2017
}
\seealso{
\code{\link{get_icd10be2014}} \code{\link{get_icd10be2014_pc}}
  \code{\link{get_icd10be2017_pc}}
}
\keyword{datasets}
