% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{get_icd10be2017_pc}
\alias{get_icd10be2017_pc}
\title{ICD-10-BE 2017 procedure codes}
\description{
ICD-10-BE 2017 procedure codes
}
\seealso{
\code{\link{get_icd10be2014}} \code{\link{get_icd10be2014_pc}}
  \code{\link{get_icd10be2017}}
}
\keyword{datasets}
