% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd-cm-ver.R
\name{get_icd10cm2016}
\alias{get_icd10cm2016}
\title{Get the 2019 ICD-10-CM codes and descriptions}
\usage{
get_icd10cm2016()
}
\description{
The most recent available data is also available as the package data,
currently \code{icd10cm2019}
}
