% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{get_icd10cm2018_pc}
\alias{get_icd10cm2018_pc}
\title{ICD-10-CM Procedure codes for 2018}
\description{
ICD-10-CM Procedure codes for 2018
}
\seealso{
\code{get_icd10cm2014_pc} \code{get_icd10cm2015_pc}
  \code{get_icd10cm2016_pc} \code{get_icd10cm2017_pc}
  \code{get_icd10cm2019_pc}
}
\keyword{datasets}
