% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\docType{data}
\name{get_icd10cm2019}
\alias{get_icd10cm2019}
\title{ICD-10-CM 2018}
\usage{
get_icd10cm2019(alt = NULL, must_work = is.null(alt),
  msg = paste("Unable to find", var_name))
}
\description{
ICD-10-CM 2018
}
\seealso{
\code{\link{icd10cm2019}}
}
\keyword{datasets}
