% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R, R/icd-cm-ver.R
\docType{data}
\name{get_icd10cm_latest}
\alias{get_icd10cm_latest}
\title{The latest available version of ICD-10-CM in this package}
\usage{
get_icd10cm_latest()

get_icd10cm_latest()
}
\description{
The latest available version of ICD-10-CM in this package

The latest available ICD-10-CM data in this package
}
\details{
This is an active binding, so is exported explicitly

This is an active binding, so is exported explicitly
}
\examples{
a <- get_icd10cm_latest()
identical(a, icd10cm2026)
}
\keyword{datasets}
