% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{get_icd10who2016}
\alias{get_icd10who2016}
\title{2016 WHO ICD-10 data}
\source{
\url{http://www.who.int}
}
\description{
This data must be downloaded on a per-user basis. A prompt is given when the
data is first attempted to be accessed.
}
\keyword{datasets}
