% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{get_icd9cm2014_leaf}
\alias{get_icd9cm2014_leaf}
\alias{iget_cd9cm2005_leaf}
\alias{get_icd9cm2006_leaf}
\alias{get_icd9cm2007_leaf}
\alias{get_icd9cm2008_leaf}
\alias{get_icd9cm2009_leaf}
\alias{get_icd9cm2010_leaf}
\alias{get_icd9cm2011_leaf}
\alias{get_icd9cm2012_leaf}
\alias{get_icd9cm2013_leaf}
\alias{get_icd9cm2005_leaf}
\title{ICD-9-CM, just billable/leaf codes}
\format{data frames with columns \code{code}, \code{short_desc}, and
  \code{long_desc}.}
\source{
\url{http://www.cms.gov/Medicare/Coding/ICD9ProviderDiagnosticCodes/codes.html}
}
\description{
ICD-9-CM, just billable/leaf codes
}
\details{
These are derived from the final CMS published version 32 for 2014,
  which was unchanged since 2011. The short descriptions are in ASCII with no
  special characters, whereas the long descriptions contain accented
  characters which are stored as Unicode, \code{latin-1} or \code{cp1252}.

  This all done during package creation, but can be repeated by package
  users, including pulling the data from the web pages directly. Despite my
  best efforts, current locale can give different results, but this packaged
  data is correct, with some \code{UTF-8} encoded strings.
  \code{icd9cm_billable} has been removed, and is replaced by now includes
  only the latest version (32).
}
\keyword{datasets}
