% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-icd-data.R
\name{get_icd_data}
\alias{get_icd_data}
\title{Internal use only: get data from the icd.data package, without relying on it
being attached}
\usage{
get_icd_data(data_name, alt = NULL)
}
\arguments{
\item{alt}{If the data cannot be found, this value is returned. Default is
\code{NULL}.}
}
\description{
Some data is hidden in active bindings, so it may be downloaded on demand,
and some is lazy-loaded. This will work for regular package members, active
bindings, and lazy data, whether or not the package is attached or loaded.
}
\details{
This is really just needed for the transition from icd 3.3 to icd 4.0, and
icd.data > 1.0
}
\keyword{internal}
