% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_icd_pc_name}
\alias{get_icd_pc_name}
\title{Guess the columns which contain ICD-9 or ICD10-CM procedure codes.}
\usage{
get_icd_pc_name(x, icd_name = NULL)
}
\description{
Will also guess procedure codes from other national ICD versions, when supported.
 @keywords internal
 @noRd
}
