% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redbook_funcs.R
\name{get_ndc_from_name}
\alias{get_ndc_from_name}
\title{Get NDC code from medication name}
\usage{
get_ndc_from_name(term, return_all = FALSE)
}
\arguments{
\item{term}{A medication name}
}
\description{
`get_ndc_from_name` returns the ndc code, or the corresponding redbook
 entries, for medications of a given name. The name provided is matched to
 the variables "genme" and "prodnme" in the redbook file.
}
\examples{
get_ndc_from_name("omeprazole")
get_ndc_from_name("omeprazole",TRUE)
}
