% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_non_ASCII}
\alias{get_encodings}
\alias{get_non_ASCII}
\alias{is_non_ASCII}
\title{mimic the R CMD check test}
\usage{
get_non_ASCII(x)

is_non_ASCII(x)

get_encodings(x)
}
\description{
mimic the R CMD check test
}
\examples{
sapply(icd9cm_hierarchy, get_non_ASCII)
get_encodings(icd9cm_hierarchy)
sapply(icd9cm_billable, get_non_ASCII)
sapply(icd9cm_billable, get_encodings)
}
\keyword{internal}

