% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-file.R
\name{get_pkg_name}
\alias{get_pkg_name}
\title{get name of this package}
\usage{
get_pkg_name()
}
\description{
so that I can change to another package name when needed.
\code{getPackageName} gives globalenv if running interactively.
}
\keyword{internal}

