% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{get_pos_short_decimal_class}
\alias{get_pos_short_decimal_class}
\title{get position to set short or decimal class}
\usage{
get_pos_short_decimal_class(x)
}
\description{
prefer immediately after \code{icd9cm}, etc., if not, place before system
classes at end, or at the very end if no system classes
}
\keyword{internal}

