% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{\%eine\%}
\alias{\%eine\%}
\title{in/match equivalent for two \code{Environment} arguments}
\usage{
x \%eine\% table
}
\description{
\code{x} and \code{table} are identical to match. Lookup is done based on
environment element names; contents are ignored.
}
\keyword{internal}

