% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{\%fin\%}
\alias{\%fin\%}
\title{Fast find which \code{x} are \emph{not} in \code{table}}
\usage{
x \%fin\% table
}
\description{
Uses \code{\link[fastmatch]{fmatch}} which creates hash table, and re-uses if
\code{table} is re-used.
}

