% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd-package.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\usage{
lhs \%>\% rhs
}
\description{
Re-export the pipe-like function to allow chaining of 'icd' commands without
explicitly loading the 'magrittr' library
}
\keyword{manip}

