% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{guess_icd_col_by_name}
\alias{guess_icd_col_by_name}
\title{get candidate column(s) from wide or long data frame frame, using hints}
\usage{
guess_icd_col_by_name(x, valid_codes = TRUE, defined_codes = FALSE)
}
\description{
get candidate column(s) from wide or long data frame frame, using hints
}
\examples{
wide_df <- data.frame(a = letters,
                      dx0 = icd9_map_elix$CHF[1:26],
                      dx1 = icd9_map_elix$PVD[1:26],
                      dx2 = icd9_map_elix$HTN[1:26])
icd:::guess_icd_col_by_name(wide_df)
}
\keyword{internal}
