% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess.R
\name{guess_pair_version}
\alias{guess_pair_version}
\title{Guess the ICD version (9 or 10) from a pair of codes}
\usage{
guess_pair_version(start, end, short_code = NULL)
}
\arguments{
\item{start}{ICD code}

\item{end}{ICD code}

\item{short_code}{single logical value which determines whether the ICD-9
code provided is in short (\code{TRUE}) or decimal (\code{FALSE}) form.
Where reasonable, this is guessed from the input data.}
}
\value{
Type of codes as single character \code{"icd9"} or \code{"icd10"}, or
  error if conflicting results
}
\description{
Designed for guessing version of codes given to delineate a range
}
\keyword{internal}
