% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{guess_short_cpp}
\alias{guess_short_cpp}
\title{Quickly guess whether codes are short form, as opposed to decimal}
\usage{
guess_short_cpp(x_, n = 100L)
}
\description{
Very quick heuristic, ploughs through ten million codes in less than one second
and will stop more quickly if it finds a '.'.
}
\details{
TODO: a factor version would be faster
}
\examples{
\dontrun{
codes <- generate_random_short_icd9(1e7)
codes_factor <- factor_nosort(codes)
system.time(guessShortCpp(codes, 1e7))
codes[1] <- "100.1"
system.time(guessShortCpp(codes, 1e7))
microbenchmark::microbenchmark(icd:::icd_guess_short(codes),
                               icd:::guessShortCpp(codes),
                               icd:::guessShortPlusFactorCpp(codes),
                               # icd:::icd_guess_short_fast(codes_factor),
                               icd:::icd_guess_short_fast(codes),
                               times = 100L)
}
}
\keyword{internal}

