% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess.R
\name{guess_version}
\alias{guess_version}
\title{Guess version of ICD codes}
\usage{
guess_version(x, short_code, ...)
}
\arguments{
\item{x}{input data}

\item{short_code}{single logical value which determines whether the ICD-9
code provided is in short (\code{TRUE}) or decimal (\code{FALSE}) form.
Where reasonable, this is guessed from the input data.}

\item{n}{number of elements or rows to sample}
}
\description{
The guess is indeed a guess and can be wrong. There are some codes which
could be either ICD-9 or ICD-10. The current implementation doesn't check
whether the codes exist in any definitions (ICD-9 CM or WHO, for example),
just whether they are valid.
}
\details{
Currently, ambiguous codes are guessed true or false, with no indication of
uncertainty. Possible solutions are adding an attribute, warning, or
optionally throwing an error.
}
\keyword{internal}
