% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children.R
\name{icd10_children_possible_short}
\alias{icd10_children_possible_short}
\title{Generate possible children of given ICD-10 codes}
\usage{
icd10_children_possible_short(x)
}
\arguments{
\item{x}{character vector of ICD-10 codes}
}
\description{
ultimately, this will have to be coding-scheme dependent, e.g. ICD-10-CM for
USA, vs various national or international schemes
}
\details{
This is inefficient due to the large number of combinations of
  possible codes. I've already limited the scope to letters which appear at
  certain positions in ICD-10-CM 2016. Maybe best just to limit to 'defined'
  codes, and then, when a user gives a squiffy code, e.g. a known code with
  an additional unknown value, we can: first account for all known codes in a
  list, then for unknown codes, partial match from left for known codes, and
  if we match more than one, then we take the longest matching known code.

  presumption is that we start with at least a major level code. We don't
  extrapolate from A10 to A10-19 for example. regex pattern for ICD-10 is
  [[:alpha:]][[:digit:]][[:alnum]]{1,5} (not case sensitive)

  can have x as placeholder: ignoring it will have same effect if making
  children

  there are very many alphanumeric values possible in last 4 digits (26^4 =
  456976) however, hardly any alphas are used.
}

