% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10_codes}
\alias{icd10_codes}
\title{ICD10 Codes}
\format{A data frame with 3 variables}
\usage{
data(icd10_codes)
}
\description{
ICD-10 is the 10th revision of the International Statistical Classification of Diseases
and Related Health Problems (ICD), a medical classification list by the World Health Organization (WHO). It contains codes for diseases, signs and symptoms, abnormal findings, complaints, social circumstances, and external causes of injury or diseases.
}
\details{
\itemize{
  \item code character the ICD-10 code
  \item desc character description of the code
  \item section integer section number for this code
#' }
}
\keyword{datasets}

