% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbid_alt.R
\name{icd10_comorbid_matmul}
\alias{icd10_comorbid_matmul}
\title{Comorbidities from ICD-10 codes using matrix multiplication}
\usage{
icd10_comorbid_matmul(x, map, visit_name = NULL, icd_name = get_icd_name(x),
  short_code = icd_guess_short(x, icd_name = icd_name),
  short_map = icd_guess_short(map), return_df = FALSE, ...)
}
\description{
Use the fact that we only need to look up comorbidities for those codes which
are in the input data frame, and map. Difference with ICD-9 codes is that we
only have 'parent' codes in the map.
}
\details{
1. get list of unique ICD-10 codes in the source data

2. make a new comorbidity map containing those codes where those codes'
parents appear in the given map

3. use the new common comorbidity code using openmp eigen matrix multiply
with that data.
}
