% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-ahrq-ccs.R
\name{icd10_fetch_ahrq_ccs}
\alias{icd10_fetch_ahrq_ccs}
\title{Download AHRQ CCS ICD-10 definitions}
\usage{
icd10_fetch_ahrq_ccs(version = "2018.1", ...)
}
\arguments{
\item{version}{Default \code{2018.1}}

\item{...}{arguments passed on to other functions}
}
\description{
Download AHRQ CCS ICD-10 definitions
}
\keyword{internal}
