% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mappings.R
\name{icd10_generate_map_quan_charlson}
\alias{icd10_generate_map_quan_charlson}
\title{Generate Quan mapping for Charlson categories of ICD-10 codes
 Based on Quan's SAS lists, transcribed by wmurphyrd
 @export}
\usage{
icd10_generate_map_quan_charlson()
}
\description{
Generate Quan mapping for Charlson categories of ICD-10 codes
 Based on Quan's SAS lists, transcribed by wmurphyrd
 @export
}

