% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-icd10-who.R
\name{icd10_get_who_from_cdc}
\alias{icd10_get_who_from_cdc}
\title{Get ICD-10 (not ICD-10-CM) as published by CDC}
\usage{
icd10_get_who_from_cdc()
}
\description{
Get ICD-10 (not ICD-10-CM) as published by CDC
}
\details{
There is no copyright notice, and, as I understand it, by default US
  government publications are public domain
  ftp://ftp.cdc.gov/pub/Health_Statistics/NCHS/Publications/ICD10/ and thus
  this or derivative data can be included in the package distribution
}
\keyword{internal}

