% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-comorbid.R
\name{icd10_parse_cc}
\alias{icd10_parse_cc}
\title{Import the ICD10 to CC crosswalks}
\usage{
icd10_parse_cc(save_data = FALSE)
}
\arguments{
\item{save_data}{logical whether to save the result in the source tree.
Defaults to \code{FALSE}.}

\item{condense}{Deprecated. The map is not condensed by default. If required,
it can reduced to a smaller set of codes using \code{icd_condense}.}
}
\value{
invisibly returns the list of vectors, where a co-morbidity name is
  associated with a character vector of ICD-9 codes.
}
\description{
Import the ICD10 to CC crosswalks
}
\keyword{internal}

