% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges-icd10.R
\name{icd10ChildrenRealShort}
\alias{icd10ChildrenRealShort}
\title{real icd10 children based on 2016 list}
\usage{
icd10ChildrenRealShort(icd10Short)
}
\description{
real icd10 children based on 2016 list
}

