% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-icd10.R
\name{icd10cm_extract_sub_chapters}
\alias{icd10cm_extract_sub_chapters}
\title{Get subchapters from the 2016 XML for ICD-10-CM}
\usage{
icd10cm_extract_sub_chapters(save_data = FALSE)
}
\arguments{
\item{save_data}{single logical}
}
\description{
This is a superset of ICD-10 sub-chapters, I think.
}
\keyword{internal}

