% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-icd10.R
\name{icd10cm_get_all_defined}
\alias{icd10cm_get_all_defined}
\title{get all ICD-10-CM codes}
\source{
\url{http://www.cdc.gov/nchs/data/icd/icd10cm/2016/ICD10CM_FY2016_code_descriptions.zip}.
}
\usage{
icd10cm_get_all_defined(save_data = FALSE)
}
\description{
Gets all ICD-10-CM codes from an archive on the CDC web site at Initially,
this just grabs 2016.
}
\details{
The factor generation uses \code{sort.default} which is locale dependent.
This meant a lot of time debugging a problem when whitespace was ignored for
sorting on some platforms, but not others (e.g. travis and wercker).
}
\references{
https://www.cms.gov/Medicare/Coding/ICD10/downloads/icd-10quickrefer.pdf
}
\keyword{internal}

