% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd10cm_order_rcpp}
\alias{icd10cm_order_rcpp}
\title{Order ICD-10-CM codes}
\usage{
icd10cm_order_rcpp(x)
}
\description{
currently required for C7A, C7B (which fall after C80), and
  D3A, which falls after D48. C4A M1A Z3A are also problems within
  sub-chapters.
}
\keyword{internal}
