% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges-icd10.R
\name{icd10ExpandRangeRealShort}
\alias{icd10ExpandRangeRealShort}
\title{expand range of ICD-10 codes returning only defined codes in ICD-10-CM}
\usage{
icd10ExpandRangeRealShort(start, end)
}
\arguments{
\item{start}{character vector of length one containing a real code}

\item{end}{character vector of length one containing a real code}
}
\description{
This will need generalizing to any list of 'real' codes, e.g. WHO or other
}
\keyword{internal}

