% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-mappings-icd10.R
\name{icd10GenerateMappingQuanElix}
\alias{icd10GenerateMappingQuanElix}
\title{generate ICD-10 Quan/Elixhauser maping}
\usage{
icd10GenerateMappingQuanElix(save = TRUE)
}
\description{
generate ICD-10 Quan/Elixhauser maping
}
\details{
start with Quan's SAS code (in \code{data-raw}):
}

