% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9_add_leading_zeroes_alt_cpp}
\alias{icd9_add_leading_zeroes_alt_cpp}
\title{directly apply \code{icd9AddLeadingZeroesShortSingle} to each code
without separating into parts}
\usage{
icd9_add_leading_zeroes_alt_cpp(x, short_code)
}
\description{
Unclear which is faster yet: working with parts or whole.
}
\keyword{internal}
\keyword{manip}
