% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbid_pccc.R
\name{icd9_comorbid_pccc_dx}
\alias{icd9_comorbid_pccc_dx}
\title{Calculate pediatric complex chronic conditions (PCCC) comorbidities}
\usage{
icd9_comorbid_pccc_dx(x, visit_name = NULL, icd_name = NULL,
  short_code = guess_short(x, icd_name = icd_name), return_df = FALSE, ...)
}
\description{
Calculate pediatric complex chronic conditions (PCCC) comorbidities
}
