% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.R
\name{icd9_drop_leading_zeroes}
\alias{icd9_drop_leading_zeroes}
\alias{icd9_drop_leading_zeroes.default}
\alias{icd9_drop_leading_zeroes.icd_decimal_code}
\alias{icd9_drop_leading_zeroes.icd_short_code}
\alias{icd9_drop_leading_zeroes_major}
\title{drop zero padding from decimal ICD-9 code.}
\usage{
icd9_drop_leading_zeroes(x, short_code = NULL)

\method{icd9_drop_leading_zeroes}{icd_decimal_code}(x, short_code = NULL)

\method{icd9_drop_leading_zeroes}{icd_short_code}(x, short_code = NULL)

\method{icd9_drop_leading_zeroes}{default}(x, short_code = NULL)

icd9_drop_leading_zeroes_major(major)
}
\arguments{
\item{icd9}{is
a character vector or factor of ICD-9 codes. If fewer than five characters
is given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010".}
}
\value{
character vector of ICD-9 codes with extra zeroes dropped from major
  part
}
\description{
decimal form ICD-9 codes are not ambiguous if the leading zeroes
  are dropped. Some short-form ICD-9 codes would not be ambiguous, e.g. "1"
  but many would be problematic. This is the inverse of
  \code{icd9_add_leading_zeroes}.
}
\details{
Invalid codes have no guaranteed result, and may give NA, or a (possibly
  valid) code in repsonse.
}
\section{Methods (by class)}{
\itemize{
\item \code{icd_decimal_code}: Drop leading zeroes from a decimal
format ICD-9 code.

\item \code{icd_short_code}: Drop leading zeroes from a short format
ICD-9 code

\item \code{default}: Drop leading zeroes from ICD-9 codes of
specified or unknown type.
}}
\seealso{
Other ICD-9 convert: \code{\link{icd9PartsToShort}},
  \code{\link{icd9_chapters_to_map}},
  \code{\link{icd_long_to_wide}},
  \code{\link{icd_wide_to_long}}
}
\keyword{internal}
\keyword{manip}

