% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{icd9_expand_range_worker}
\alias{icd9_expand_range_worker}
\title{expand range worker function}
\usage{
icd9_expand_range_worker(start, end, lookup, defined, ex_ambig_start,
  ex_ambig_end)
}
\description{
expand range worker function
}
\examples{
\dontrun{
microbenchmark::micmicrobenchmark(
  icd9_expand_range_worker_alt_base("100", "114", icd9_short_n, TRUE, TRUE, TRUE),
  icd9_expand_range_worker("100", "114", icd9_short_n, TRUE, TRUE, TRUE),
  times = 5
)
}
}
\keyword{internal}

