% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-ahrq-ccs.R
\name{icd9_fetch_ahrq_ccs}
\alias{icd9_fetch_ahrq_ccs}
\title{Download ahrq-css-icd-09 definition}
\usage{
icd9_fetch_ahrq_ccs(type, offline)
}
\description{
Download ahrq-css-icd-09 definition
}
\keyword{internal}
