% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-comorbid.R
\name{icd9_fetch_ahrq_sas}
\alias{icd9_fetch_ahrq_sas}
\title{get the SAS code from AHRQ}
\usage{
icd9_fetch_ahrq_sas(offline)
}
\description{
Get the SAS code from AHRQ and save in raw data directory, if not already
there.
}
\keyword{internal}
