% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sysdata.R
\name{icd9_generate_all_}
\alias{icd9_generate_all_}
\title{generate lookup data for each class of ICD-9 code}
\usage{
icd9_generate_all_(major_fun, short_code = TRUE, env = new.env(hash =
  TRUE, baseenv()))
}
\value{
list with two items, the first being an environment, the second being
  a vector. The environment has short ICD-9 codes as the names, and the
  sequence number as the contents. The vector contains the codes in order.
}
\description{
This is a little painful but the data is small enough, with huge speed gains
in common operations
}
\keyword{internal}
