% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{icd9_generate_chapters_hierarchy}
\alias{icd9_generate_chapters_hierarchy}
\title{generate ICD-9-CM hierarchy}
\usage{
icd9_generate_chapters_hierarchy(save_data = FALSE, verbose = FALSE)
}
\description{
For each row of billing code, give the chapter, sub-chapter, major code and
description, and short and long descriptions. Currently this is specifically
for the 2011 ICD-9-CM after which there have been minimal changes.
}
\keyword{internal}

