% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mappings.R
\name{icd9_generate_map_elix}
\alias{icd9_generate_map_elix}
\alias{icd10_generate_map_elix}
\title{Generate Elixhauser comorbidities}
\usage{
icd9_generate_map_elix(save_data = TRUE)

icd10_generate_map_elix(save_data = TRUE)
}
\arguments{
\item{save_data}{logical whether to save the result in the source tree.
Defaults to \code{FALSE}.}

\item{condense}{Deprecated. The map is not condensed by default. If required,
it can reduced to a smaller set of codes using \code{icd_condense}.}
}
\value{
invisibly returns the list of vectors, where a co-morbidity name is
  associated with a character vector of ICD-9 codes.
}
\description{
This function uses the \code{\%i9d\%} operator, so cannot be done as an R
file in the \code{data} directory. The data is documented in
\code{datadocs.R}.
}
\keyword{internal}
