% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sysdata.R
\name{icd9_generate_sources}
\alias{icd9_generate_sources}
\title{generate data for finding source data for ICD-9-CM}
\source{
\url{http://www.cms.gov/Medicare/Coding/ICD9ProviderDiagnosticCodes/codes.html}
}
\usage{
icd9_generate_sources(save_data = FALSE)
}
\description{
generate data for finding source data for ICD-9-CM
}
\keyword{internal}

