% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{icd9_get_chapters}
\alias{icd9_get_chapters}
\title{get ICD-9 Chapters from vector of ICD-9 codes}
\usage{
icd9_get_chapters(x, short_code = guess_short(x), verbose = FALSE)
}
\arguments{
\item{x}{vector of ICD-9 codes}

\item{short_code}{single logical value which determines whether the ICD-9
code provided is in short (\code{TRUE}) or decimal (\code{FALSE}) form.
Where reasonable, this is guessed from the input data.}

\item{verbose}{single logical value, defaults to \code{FALSE} in most functions.}
}
\description{
This runs quite slowly. Used too rarely to be worth optimizing
  now. This is used to build a master list of ICD-9 codes with their
  respective chapters, sub-chapters, etc..
}
\keyword{internal}
