% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{icd9_is_n}
\alias{icd9_is_n}
\alias{icd9_is_v}
\alias{icd9_is_e}
\title{do ICD-9 codes belong to numeric, V or E sub-types?}
\usage{
icd9_is_n(x)

icd9_is_v(x)

icd9_is_e(x)
}
\arguments{
\item{x}{vector of strings or factor to test}
}
\value{
logical vector
}
\description{
For each code, return \code{TRUE} if numeric or \code{FALSE} if a
  V or E code.
}
\section{Functions}{
\itemize{
\item \code{icd9_is_v}: are the given codes V type?

\item \code{icd9_is_e}: are the given codes E type?
}}

