% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdocs.R
\docType{data}
\name{icd9_map_single_ccs}
\alias{icd9_map_single_ccs}
\alias{icd9_map_multi_ccs}
\alias{icd10_map_ccs}
\title{Clinical Classifications Software (CCS) for ICD9/10-CM}
\format{
list of character vectors, each numbered by co-morbidity
}
\description{
The Clinical Classifications Software (CCS) for ICD-9/10-CM is one in
a family of databases and software tools developed as part of the
Healthcare Cost and Utilization Project (HCUP),a
Federal-State-Industry partnership sponsored by the Agency for
Healthcare Research and Quality. HCUP databases, tools, and software
inform decision making at the national, State, and community levels.
The software contains two different mappings. One is a single level
mapping and one is a multi level classification. This data set
contains the numeric representations of all of the codes.
}
\details{
This file contains the updated ICD9 version that includes categories
for Mental Health and Substance Abuse. More information on the
ICD-9-CM data set can be found
\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccs/ccs.jsp}.

The file \code{icd10_map_ccs} contains the 2018.1 ICD10 Version of
the mapping. More information on the ICD10 code set can be found at
\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccs10/ccs10.jsp}
}
\seealso{
\code{\link{comorbid_ccs}} \code{\link{icd9_comorbid_ccs}}
  \code{\link{icd10_comorbid_ccs}}

Other comorbidity maps:
\code{\link{icd10_map_ahrq_pcs}},
\code{\link{icd9_map_ahrq}},
\code{\link{icd9_map_elix}},
\code{\link{icd9_map_hcc}},
\code{\link{icd9_map_pccc}},
\code{\link{icd9_map_quan_deyo}},
\code{\link{icd9_map_quan_elix}}

Other comorbidities:
\code{\link[=comorbid]{comorbid()}},
\code{\link[=comorbid_hcc]{comorbid_hcc()}},
\code{\link{icd10_map_ahrq_pcs}},
\code{\link{icd9_map_ahrq}},
\code{\link{icd9_map_elix}},
\code{\link{icd9_map_hcc}},
\code{\link{icd9_map_pccc}},
\code{\link{icd9_map_quan_deyo}},
\code{\link{icd9_map_quan_elix}}
}
\concept{comorbidities}
\concept{comorbidity maps}
\keyword{datasets}
