% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-comorbid.R
\name{icd9_parse_cc}
\alias{icd9_parse_cc}
\title{Generate ICD to HCC Crosswalks from CMS}
\usage{
icd9_parse_cc(save_data = FALSE)
}
\arguments{
\item{save_data}{logical whether to save the result in the source tree.
Defaults to \code{FALSE}.}

\item{condense}{Deprecated. The map is not condensed by default. If required,
it can reduced to a smaller set of codes using \code{icd_condense}.}
}
\value{
invisibly returns the list of vectors, where a co-morbidity name is
  associated with a character vector of ICD-9 codes.
}
\description{
The ICD/HCC mappings were obtained from CMS
\url{https://www.cms.gov/Medicare/Health-Plans/MedicareAdvtgSpecRateStats/Risk-Adjustors.html}.
Due to the complex file structure of the original data (many nested zip files), they have
been organized in the folder data/icd_hcc_rawdata/.
This function creates an .RData file containing icd9/10 to CC crosswalks.
Import the ICD9 to CC crosswalks
}
\keyword{internal}

