% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{icd9_sources}
\alias{icd9_sources}
\alias{icd10_sources}
\title{ICD-9 and ICD-10 data sources}
\description{
List of ICD-9 data sources for different versions of ICD-9-CM
}
\keyword{datasets}
\keyword{internal}
