% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_funcs.R
\name{icd9_to_icd10}
\alias{icd9_to_icd10}
\title{Convert icd-9 values to icd-10}
\usage{
icd9_to_icd10(icd_list)
}
\arguments{
\item{icd_list}{A vector of ICD-9 values}
}
\description{
`icd9_to_icd10` converts a vector of ICD-9 values to possible ICD-10 values.
Uses the NBER crosswalk file.
}
\examples{
icd9_to_icd10("00845")
}
