% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9AddLeadingZeroes_alt_ShortSingle}
\alias{icd9AddLeadingZeroes_alt_ShortSingle}
\alias{icd9_add_leading_zeroes_alt_cpp}
\title{Decompose a 'short' ICD code and insert the leading zeroes as needed.}
\usage{
icd9AddLeadingZeroes_alt_ShortSingle(x)

icd9_add_leading_zeroes_alt_cpp(x, short_code)
}
\description{
This should add leading zeroes when there is definitely no ambiguity,
e.g. V1. However V10 should not be altered, because V010 is a different
code. The goal is for this to be faster, but must be correct! Example in
\code{manip.cpp} has the benchmark code.
}
\section{Functions}{
\itemize{
\item \code{icd9_add_leading_zeroes_alt_cpp}: Directly apply
\code{icd9AddLeadingZeroesShortSingle} to each code without separating into
parts
}}

\keyword{internal}
\keyword{manip}
