% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9AddLeadingZeroesMajorSingle}
\alias{icd9AddLeadingZeroesMajorSingle}
\title{Simpler add leading zeroes without converting to parts and back}
\usage{
icd9AddLeadingZeroesMajorSingle(mjr)
}
\description{
Simpler add leading zeroes without converting to parts and back
}
\keyword{internal}
\keyword{manip}
