% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9AddLeadingZeroesShortSingle}
\alias{icd9AddLeadingZeroesShortSingle}
\title{Decompose a 'short' ICD code and insert the leading zeroes as needed.}
\usage{
icd9AddLeadingZeroesShortSingle(x)
}
\description{
This should add leading zeroes when there is definitely no ambiguity,
e.g. V1. However V10 should not be altered, because V010 is a different
code. The goal is for this to be faster, but must be correct! Example in manip.cpp
has the benchmark code.
}
\keyword{internal}
\keyword{manip}
