% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9AppendMinors}
\alias{icd9AppendMinors}
\title{append minor to major using std}
\usage{
icd9AppendMinors(m, mnr, isShort)
}
\description{
benefits from having reserve string size of 5
}
\keyword{internal}

