% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/benchmark.R
\name{icd9Benchmark}
\alias{icd9Benchmark}
\title{benchmark and profile major functions with larger data sets}
\usage{
icd9Benchmark()
}
\description{
\code{icd9} package is intended to be used with large data sets, with
millions or rows. Performance of core functions is therefore of some
importance, after correctness. R package test code is for correctness,
whereas this script stresses the core functions, and looks for bottlenecks.
}

