% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convert.R
\name{icd9ChaptersToMap}
\alias{icd9ChaptersToMap}
\title{convert the chapter headings to lists of codes}
\usage{
icd9ChaptersToMap(x)
}
\arguments{
\item{x}{Either a chapter list itself, or the name of one, e.g.
icd9ChaptersSub}
}
\description{
the chapter headings can be converted into the full set of their
  children, and then used to look-up which chapter, sub-chapter, or 'major' a
  given code belongs. Always returns a map with short-form icd-9 codes. These
  can be converted en masse with \code{lapply} and \code{icd9ShortToDecimal}.
}
\keyword{internal}
\keyword{manip}

