% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{icd9Charlson}
\alias{icd9Charlson}
\alias{icd9Charlson.data.frame}
\alias{icd9CharlsonComorbid}
\title{Deprecated: calculate ICD-9 Charlson score}
\usage{
icd9Charlson(x, visitId = NULL, scoringSystem = c("original", "charlson",
  "quan"), return.df = FALSE,
  stringsAsFactors = getOption("stringsAsFactors"), ...)

\method{icd9Charlson}{data.frame}(x, visitId = NULL,
  scoringSystem = c("original", "charlson", "quan"), return.df = FALSE,
  stringsAsFactors = getOption("stringsAsFactors"), ...)

icd9CharlsonComorbid(x, visitId = NULL, applyHierarchy = FALSE,
  scoringSystem = c("original", "charlson", "quan"))
}
\arguments{
\item{x}{vector of ICD-9 codes}

\item{visitId}{Deprecated. Use \code{visit_name}}

\item{scoringSystem}{Deprecated. Use \code{icd_charlson} with
\code{scoring_system}. One of \code{original}, \code{charlson}, or
\code{quan}. The first two will give the original Charlson weights for each
comorbidity, whereas \code{quan} uses the updated weights from Quan 2001.}

\item{return.df}{Deprecated. Use \code{icd_charlson} with \code{return_df}.
single logical value, if true, a two column data frame will be returned,
with the first column named as in input data frame (i.e.
\code{visit_name}), containing all the visits, and the second column
containing the Charlson Comorbidity Index.}

\item{stringsAsFactors}{Single logical value, describing whether the resulting data frame should have strings, e.g.
\code{visit_id} converted to factor. Default is to follow the current session option. This is identical to the
argument used in, among other base functions \code{as.data.frame}.}

\item{...}{arguments passed on to other functions}

\item{applyHierarchy}{single logical value, default is FALSE. If TRUE, will drop
\code{DM} if \code{DMcx} is present, etc.}
}
\description{
Use \code{icd_charlson} instead.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: \code{data.frame} method
}}
\keyword{internal}

