% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9ChildrenShort_alt_11}
\alias{icd9ChildrenShort_alt_11}
\title{Find child codes from vector of ICD-9 codes.}
\usage{
icd9ChildrenShort_alt_11(icd9Short, onlyReal)
}
\description{
Pure C++11 implementation using \code{unordered set} to find children of
given codes
}
\examples{
\dontrun{
if (requireNamespace("microbenchmark")) {
  microbenchmark::microbenchmark(
    icd:::icd9ChildrenShort(c("001", 100:500), onlyReal = TRUE),
    icd:::icd9ChildrenShort_alt11(c("001", 100:500), onlyReal = TRUE),
    times = 5)
    # C++11 about 15\% faster for this data
}
}
}
\keyword{internal}
