% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9ChildrenShort_alt_Std}
\alias{icd9ChildrenShort_alt_Std}
\title{C++ implementation of finding children of short codes}
\usage{
icd9ChildrenShort_alt_Std(icd9Short, onlyReal)
}
\description{
C++ implementation of finding children of short codes
}
\keyword{internal}
