% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9ChildrenShortCppStd}
\alias{icd9ChildrenShortCppStd}
\title{C++ implementation of finding children of short codes}
\usage{
icd9ChildrenShortCppStd(icd9Short, onlyReal)
}
\description{
C++ implementation of finding children of short codes
}
\examples{
\dontrun{
library(microbenchmark)
microbenchmark(icd9ChildrenShortCpp("001", T), icd9ChildrenShortCppStd("001", T), times = 100)
microbenchmark(icd9ChildrenShortCpp(c("001", 100:400), T),
               icd9ChildrenShortCppUnordered(c("001", 100:400), T),
               icd9ChildrenShortCppStd(c("001", 100:400), T),
               times = 10)
}
# unordered set much faster, but may still need to sort result
}
\keyword{internal}

