% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9ChildrenShortStd}
\alias{icd9ChildrenShortStd}
\title{C++ implementation of finding children of short codes}
\usage{
icd9ChildrenShortStd(icd9Short, onlyReal)
}
\description{
C++ implementation of finding children of short codes
}
\examples{
\dontrun{
library(microbenchmark)
microbenchmark(icd9ChildrenShort("001", T), icd9ChildrenShortStd("001", T), times = 100)
microbenchmark(icd9ChildrenShort(c("001", 100:400), T),
               icd9ChildrenShortUnordered(c("001", 100:400), T),
               icd9ChildrenShortStd(c("001", 100:400), T),
               times = 10)
}
# unordered set much faster, but may still need to sort result
}
\keyword{internal}

