% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{icd9cm2014_leaf}
\alias{icd9cm2014_leaf}
\title{The final ICD-9-CM list of leaf (\sQuote{billable}) codes}
\description{
Other years are available from \code{\link{get_icd9cm2013_leaf}}, etc..
}
\keyword{datasets}
