% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.R
\name{icd9cm_get_billable}
\alias{icd9cm_get_billable}
\alias{icd9cm_get_billable.decimal_code}
\alias{icd9cm_get_billable.short_code}
\title{Get billable ICD-9-CM codes}
\usage{
icd9cm_get_billable(x, short_code = icd_guess_short(x), invert = FALSE,
  version = getLatestBillableVersion())

\method{icd9cm_get_billable}{short_code}(x, invert = FALSE,
  version = getLatestBillableVersion())

\method{icd9cm_get_billable}{decimal_code}(x, invert = FALSE,
  version = getLatestBillableVersion())
}
\description{
icd9_is_billable Return only those codes which are leaf (billable)
  codes in the hierarchy.
}
\section{Methods (by class)}{
\itemize{
\item \code{short_code}: Get the billable ICD-9-CM codes from vector of short codes

\item \code{decimal_code}: Get the billable ICD-9-CM codes from vector of decimal codes
}}

