% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{icd9cm_hierarchy}
\alias{icd9cm_hierarchy}
\alias{get_icd9cm2005}
\alias{get_icd9cm2006}
\alias{get_icd9cm2007}
\alias{get_icd9cm2008}
\alias{get_icd9cm2009}
\alias{get_icd9cm2010}
\alias{get_icd9cm2011}
\alias{get_icd9cm2012}
\alias{get_icd9cm2013}
\alias{get_icd9cm2014}
\title{ICD-9-CM diagnosis codes including leaf nodes and branch names up to the
three-digit codes.}
\source{
\url{http://wonder.cdc.gov/wonder/sci_data/codes/icd9/type_txt/icd9cm.asp}

Rich text descriptions here:
  \url{http://www.cdc.gov/nchs/icd/icd9cm.htm}
  \url{http://www.cms.gov/Medicare/Coding/ICD9ProviderDiagnosticCodes/codes.html}
   This page has versions 23 to 32 (2005 to 2014). At present, only the 2014
  data is included in this package.

\url{http://wonder.cdc.gov/wonder/sci_data/codes/icd9/type_txt/icd9abb.asp}

\url{http://wonder.cdc.gov/wonder/sci_data/codes/icd9/type_txt/icd9cm.asp}

\url{http://wonder.cdc.gov/wonder/sci_data/codes/icd9/type_txt/icdcm.asp}

\url{http://wonder.cdc.gov/wonder/sci_data/codes/icd9/type_txt/icd9abb.asp}
}
\description{
Unlike \code{\link{get_icd9cm2014_leaf}} and friends, these data frames
contain the full structure of the ICD-9-CM scheme up to the three-digit
codes. Unlike the equivalent ICD-10-CM data frames (e.g.,
\code{\link{icd10cm2026}}), they do not have columns indicating billable/leaf
status, or chapter designations.
}
\details{
\code{icd9cm_hierarchy} is the deprecated name, which currently points to the
final ICD-9-CM 2014 release. 2011 -- 2014 are in fact identical. These are
generated dynamically by parsing an awkward RTF file, and only the 2011 --
2014 data has been tested and used thoroughly.

Format:  data frame
}
\keyword{datasets}
