% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{icd9cm_hierarchy_hotfix}
\alias{icd9cm_hierarchy_hotfix}
\title{fix some \code{RTF} parsing errors}
\usage{
icd9cm_hierarchy_hotfix(x)
}
\description{
These are relevant to the most recent ICD-9-CM code sets, not the older
historic versions. There will be no further updates, so this is reasonable.
}
\keyword{internal}
\keyword{manip}
