% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{icd9cm_latest_edition}
\alias{icd9cm_latest_edition}
\title{Latest ICD-9-CM edition}
\usage{
icd9cm_latest_edition()
}
\description{
Returns a single character value with the number of the latest edition,
currently \strong{32}.
}
\details{
Implemented as a function to give flexibility to calculate this, or use an
option override. Duplicated in \code{icd.data} package.
}
\keyword{internal}
