% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{icd9cm_make_chapters_hierarchy}
\alias{icd9cm_make_chapters_hierarchy}
\title{generate ICD-9-CM hierarchy}
\usage{
icd9cm_make_chapters_hierarchy(save_data = FALSE, verbose = FALSE,
  offline = TRUE, perl = TRUE, useBytes = TRUE)
}
\arguments{
\item{save_data}{single logical value, if \code{TRUE} will save the generated
data}

\item{verbose}{single logical value, defaults to \code{FALSE} in most functions.}

\item{offline}{single logical, if \code{TRUE} then don't pull the file from
internet, only return path and file name if the file already exists in
data-raw. This is helpful for testing without using the internet.}
}
\description{
For each row of billing code, give the chapter, sub-chapter, major code and
description, and short and long descriptions. Currently this is specifically
for the 2011 ICD-9-CM after which there have been minimal changes.
Thankfully, ICD-10-CM has machine readable data available.
}
\details{
SOMEDAY add 'billable' column, and make consistent icd9 and icd10 lookup
tables
}
\keyword{internal}

