% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{icd9Comorbid}
\alias{icd9Comorbid}
\alias{icd9ComorbidAhrq}
\alias{icd9ComorbidElix}
\alias{icd9ComorbidQuanDeyo}
\alias{icd9ComorbidQuanElix}
\alias{icd9ComorbidQuanElixhauser}
\alias{icd9ComorbidShort}
\alias{icd9Comorbidities}
\alias{icd9ComorbiditiesAhrq}
\alias{icd9ComorbiditiesElixHauser}
\alias{icd9ComorbiditiesQuanDeyo}
\alias{icd9ComorbiditiesQuanElixhauser}
\title{Deprecated: get comorbidities from patient data}
\usage{
icd9Comorbid(icd9df, icd9Mapping, visitId = NULL, icd9Field = NULL,
  isShort = icd_guess_short(icd9df),
  isShortMapping = icd_guess_short(icd9Mapping), return.df = FALSE, ...)

icd9ComorbidShort(...)

icd9Comorbidities(...)

icd9ComorbiditiesAhrq(...)

icd9ComorbiditiesElixHauser(...)

icd9ComorbiditiesQuanDeyo(...)

icd9ComorbiditiesQuanElixhauser(...)

icd9ComorbidQuanElixhauser(...)

icd9ComorbidAhrq(icd9df, visitId = NULL, icd9Field = NULL,
  isShort = icd9GuessIsShort(icd9df[1:100, icd9Field]), return.df = FALSE,
  abbrevNames = TRUE, applyHierarchy = TRUE, ...)

icd9ComorbidElix(icd9df, visitId = NULL, icd9Field = NULL,
  isShort = icd9GuessIsShort(icd9df[1:100, icd9Field]), return.df = FALSE,
  abbrevNames = TRUE, applyHierarchy = TRUE, ...)

icd9ComorbidQuanElix(icd9df, visitId = NULL, icd9Field = NULL,
  isShort = icd9GuessIsShort(icd9df[1:100, icd9Field]), return.df = FALSE,
  abbrevNames = TRUE, applyHierarchy = TRUE, ...)

icd9ComorbidQuanDeyo(icd9df, visitId = NULL, icd9Field = NULL,
  isShort = icd9GuessIsShort(icd9df[1:100, icd9Field]), return.df = FALSE,
  abbrevNames = TRUE, applyHierarchy = TRUE, ...)
}
\arguments{
\item{icd9df}{Deprecated. Use \code{x} to provide a data frame input}

\item{icd9Mapping}{named list containing vectors of ICD-9 codes. E.g. the
AHRQ comorbidities, contains \code{list(OBESE = c("2780", "27800", "27801",
"27803", "V8554", "79391", "64910", "64911", "64912", "64913", "64914",
"V8530", "V8531", "V8532", "V8533", "V8534", "V8535", "V8536", "V8537",
"V8538", "V8539", "V8541", "V8542", "V8543", "V8544", "V8545" ), DEPRESS =
c("3004", "30112", "3090", "3091", "311"))} amongst other longer groups.}

\item{visitId}{Deprecated. Use \code{visit_name}}

\item{icd9Field}{The column in the data frame which contains the ICD codes.
This is a character vector of length one. If it is \code{NULL}, \code{icd9}
will attempt to guess the column name, looking for progressively less
likely possibilities until it matches a single column. Failing this, it
will take the first column in the data frame. Specifying the column using
this argument avoids the guesswork.}

\item{isShort}{Deprecated. Single logical value which determines whether the
ICD-9 code provided is in short (\code{TRUE}) or decimal (\code{FALSE})
form. Where reasonable, this is guessed from the input data.}

\item{isShortMapping}{logical}

\item{return.df}{Deprecated. Use \code{return_df} instead. Single logical
value, if \code{TRUE}, return the result as a data frame with the first
column being the \code{visitId}, and the second being the count. If
\code{visitId} was a factor or named differently in the input, this is
preserved.}

\item{...}{arguments passed to the corresponding function from the alias.
E.g. all the arguments passed to \code{icd9ComorbiditiesAhrq} are passed on
to \code{icd9ComorbidAhrq}}

\item{abbrevNames}{logical}

\item{applyHierarchy}{logical}
}
\description{
Use \code{icd_comorbid} instead
}

