% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9ComorbidMatMul}
\alias{icd9ComorbidMatMul}
\title{prototype to do entire comorbidity calculation as a matrix multiplication}
\usage{
icd9ComorbidMatMul(icd9df, icd9Mapping, visitId, icd9Field, threads = 8L,
  chunk_size = 256L, omp_chunk_size = 1L, aggregate = TRUE)
}
\description{
The problem is that the matrices could be huge: the patient-icd matrix would
be millions of patient rows, and ~15000 columns for all AHRQ comorbidities.
}
\details{
Several ways of reducing the problem: firstly, as with existing code, we can
drop any ICD codes from the map which are not in the patient data. With many
patients, this will be less effective as the long tail becomes apparent.
However, with the (small) Vermont data, we see ~15,000 codes being reduced to
339.
}
\examples{
# show how many discrete ICD codes there are in the AHRQ map, before reducing
# to the number which actually appear in a group of patient visits
sapply(icd::icd9_map_ahrq, length) \%>\% sum
icd_comorbid_ahrq(vermont_dx \%>\% icd_wide_to_long, comorbid_fun = icd:::icd9ComorbidShortCpp)
\dontrun{
# remove _alt line in .Rbuildignore, then these will be available
icd_comorbid_ahrq(vermont_dx \%>\% icd_wide_to_long, comorbid_fun = icd:::icd9ComorbidMatMul)
icd_comorbid_ahrq(vermont_dx \%>\% icd_wide_to_long, comorbid_fun = icd:::icd9ComorbidSparseOmp)
}
}
\keyword{internal}
