% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9ComorbidTaskloop}
\alias{icd9ComorbidTaskloop}
\title{Work towards simpler comorbidity assignment, possibly using OpenMP taskloops
up <- icd::uranium_pathology
up[["case"]] <- as.character(up[["case"]])
icd9ComorbidTaskloop(icd9df = up, icd9Mapping = icd10_map_ahrq, visitId = "case", icd9Field = "icd10", TRUE)}
\usage{
icd9ComorbidTaskloop(icd9df, icd9Mapping, visitId, icd9Field,
  aggregate = TRUE)
}
\description{
Work towards simpler comorbidity assignment, possibly using OpenMP taskloops
up <- icd::uranium_pathology
up[["case"]] <- as.character(up[["case"]])
icd9ComorbidTaskloop(icd9df = up, icd9Mapping = icd10_map_ahrq, visitId = "case", icd9Field = "icd10", TRUE)
}
\keyword{internal}
