% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{icd9Count}
\alias{icd9Count}
\alias{icd9CountComorbidBin}
\alias{icd9CountWide}
\title{Deprecated: count number of ICD codes per patient}
\usage{
icd9Count(x, visitId = get_visit_name(x), return.df = FALSE)

icd9CountWide(x, visitId = get_visit_name(x), return.df = FALSE,
  aggregate = FALSE)

icd9CountComorbidBin(x, visitId = get_visit_name(x), return.df = FALSE)
}
\arguments{
\item{x}{\code{data.frame}}

\item{visitId}{Deprecated. Use \code{visit_name}}

\item{return.df}{Deprecated. Use \code{return_df} instead. Single logical
value, if \code{TRUE}, return the result as a data frame with the first
column being the \code{visitId}, and the second being the count. If
\code{visitId} was a factor or named differently in the input, this is
preserved.}

\item{aggregate, }{single logical, default is FALSE. If TRUE, the length (or
rows) of the output will no longer match the input, but duplicate
\code{visitId}s will be counted together.}
}
\description{
Use \code{icd_count_codes} instead.
}

