% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{icd9DecimalToParts}
\alias{icd9DecimalToParts}
\alias{icd9DecimalToShort}
\alias{icd9ShortToDecimal}
\alias{icd9ShortToParts}
\title{Deprecated: convert between short, decimal ICD-9 codes}
\usage{
icd9DecimalToParts(icd9Decimal, minorEmpty = "")

icd9DecimalToShort(icd9Short)

icd9ShortToDecimal(icd9Short)

icd9ShortToParts(icd9Short)
}
\arguments{
\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five characters is given in a
code, then the digits are greedily assigned to hundreds, then tens, then
units, before the decimal parts. E.g. "10" becomes "010", not "0010"}

\item{minorEmpty}{vector of length one, to be used in place of minor part of
zero. Defaults to ""}

\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than five characters is
given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010"}
}
\description{
Use \code{icd_decimal_to_short} etc. instead.
}

