% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convert.R
\name{icd9DecimalToParts_R}
\alias{icd9DecimalToParts_R}
\title{convert decimal-form ICD-9 code to major and minor parts}
\usage{
icd9DecimalToParts_R(icd9Decimal, minorEmpty = "")
}
\arguments{
\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five
characters is given in a code, then the digits are greedily assigned to
hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{minorEmpty}{vector of length one, to be used in place of
minor part of zero. Defaults to ""}
}
\description{
convert decimal-form ICD-9 code to major and minor parts
}
\keyword{manip}

