% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convert.R
\name{icd9DecimalToShort_R}
\alias{icd9DecimalToShort_R}
\title{convert between icd9 decimal and short formats}
\usage{
icd9DecimalToShort_R(icd9Decimal)
}
\arguments{
\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five
characters is given in a code, then the digits are greedily assigned to
hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}
}
\value{
character vector of converted ICD-9 codes
}
\description{
converted decimal ICD9 code, e.g. 123.45 to 'short' e.g. 12345
  non-decimal format
}
\seealso{
Other ICD-9 convert: \code{\link{icd9DropLeadingZeroes}},
  \code{\link{icd9DropLeadingZeroesDecimal}},
  \code{\link{icd9DropLeadingZeroesMajor}},
  \code{\link{icd9DropLeadingZeroesShort}};
  \code{\link{icd9MajMinToDecimal_R}},
  \code{\link{icd9MajMinToParts_R}},
  \code{\link{icd9MajMinToShort_R}},
  \code{\link{icd9PartsRecompose}},
  \code{\link{icd9PartsToDecimal}},
  \code{\link{icd9PartsToDecimal_R}},
  \code{\link{icd9PartsToShort}},
  \code{\link{icd9PartsToShort_R}};
  \code{\link{icd9ShortToDecimal_R}}
}
\keyword{manip}

