% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/comorbid.R
\name{icd9DiffComorbid}
\alias{icd9DiffComorbid}
\title{show the difference between two comorbidity mappings}
\usage{
icd9DiffComorbid(x, y, x.names = NULL, y.names = NULL, show = TRUE,
  explain = TRUE)
}
\description{
show the difference between two comorbidity mappings
}

