% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R, R/ranges.R
\name{icd9ExpandMinor}
\alias{icd9ExpandMinor}
\alias{icd9ExpandMinorE}
\alias{icd9ExpandMinorNV_R}
\title{Expand a root 'minor' part of code into all possible children.}
\usage{
icd9ExpandMinor(x, isE)

icd9ExpandMinorNV_R(minor = "")

icd9ExpandMinorE(minor = "")
}
\description{
Expand a root 'minor' part of code into all possible children.
}
\examples{
microbenchmark(icd9:::icd9ExpandMinor("7", FALSE), icd9:::icd9ExpandMinor_R("7", FALSE))
microbenchmark(icd9:::icd9ExpandMinor("", FALSE), icd9:::icd9ExpandMinor_R("", FALSE))
microbenchmark(icd9:::icd9ExpandMinor("7", TRUE), icd9:::icd9ExpandMinor_R("7", TRUE))
microbenchmark(icd9:::icd9ExpandMinor("", TRUE), icd9:::icd9ExpandMinor_R("", TRUE))
}

