% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{icd9ExpandRange}
\alias{icd9ExpandRange}
\alias{icd9ExpandRangeDecimal}
\alias{icd9ExpandRangeMajor}
\alias{icd9ExpandRangeShort}
\title{Deprecated: expand range of ICD-9 codes}
\usage{
icd9ExpandRange(start, end, isShort = icd_guess_short.icd9(c(start, end)),
  onlyReal = TRUE, excludeAmbiguousStart = TRUE,
  excludeAmbiguousEnd = TRUE)

icd9ExpandRangeShort(start, end, onlyReal = TRUE,
  excludeAmbiguousStart = TRUE, excludeAmbiguousEnd = TRUE)

icd9ExpandRangeDecimal(start, end, onlyReal = TRUE,
  excludeAmbiguousStart = TRUE, excludeAmbiguousEnd = TRUE)

icd9ExpandRangeMajor(start, end, onlyReal = TRUE)
}
\arguments{
\item{start}{ICD-9 code}

\item{end}{ICD-9 code}

\item{isShort}{Deprecated. Single logical value which determines whether the
ICD-9 code provided is in short (\code{TRUE}) or decimal (\code{FALSE})
form. Where reasonable, this is guessed from the input data.}

\item{onlyReal}{logical}

\item{excludeAmbiguousStart}{logical}

\item{excludeAmbiguousEnd}{logical}
}
\description{
Use \code{icd_expand_range} instead.
}

