% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ranges.R
\name{icd9ExpandRangeMajor}
\alias{\%i9mj\%}
\alias{icd9ExpandRangeMajor}
\title{create range of icd9 major parts}
\usage{
icd9ExpandRangeMajor(start, end)

start \%i9mj\% end
}
\arguments{
\item{start,end}{is a character vector of ICD-9 codes. If fewer than five characters is
given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010"}
}
\value{
character vector with range inclusive of start and end
}
\description{
accepts V, E or numeric codes. Does not validate codes beyond
  ensuring that the start and end of the range are of the same type. Will add
  leading zeroes when appropriate. User can strip them out with
  icd9DropLeadingZeroes if they wish.
}
\seealso{
Other ICD-9 ranges: \code{\link{icd9ChildrenDecimal_R}};
  \code{\link{icd9ChildrenDecimal}};
  \code{\link{icd9Children}},
  \code{\link{icd9ChildrenShort}},
  \code{\link{icd9ChildrenShort_R}},
  \code{\link{icd9Children_R}};
  \code{\link{icd9CondenseToExplain}};
  \code{\link{icd9CondenseToMajor}};
  \code{\link{icd9ExpandMinor_R}};
  \code{\link{icd9ExpandRange}},
  \code{\link{icd9ExpandRangeDecimal}},
  \code{\link{icd9ExpandRangeShort}}
}

