% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/explain.R
\name{icd9ExplainCondense}
\alias{icd9ExplainCondense}
\title{condense list of short ICD-9 code into minimal set of parent descriptions}
\usage{
icd9ExplainCondense(icd9Short)
}
\arguments{
\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{icd9}{is a character vector of ICD-9 codes. If fewer than five characters is
given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010"}

\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five
characters is given in a code, then the digits are greedily assigned to
hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{isShort}{single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form.}
}
\description{
This can be thought of as the inverse operation to expanding a
  range. The list given must already contain the parents, because this
  function will never add a parent ICD-9 which, although may have all
  children present, may itself have an additional clinical meaning. In
  addition, in contrast to \code{icd9CondenseToMajor}, this function only walks
  back up to parents which have descriptions in \code{icd9Hierarchy}, so it is
  useful for generating a minimal textual description of a set of ICD-9
  codes.
}
\seealso{
Other ICD-9 ranges: \code{\link{\%i9d\%}},
  \code{\link{\%i9mj\%}}, \code{\link{\%i9s\%}},
  \code{\link{icd9ExpandRange}},
  \code{\link{icd9ExpandRangeDecimal}},
  \code{\link{icd9ExpandRangeMajor}},
  \code{\link{icd9ExpandRangeShort}};
  \code{\link{icd9Children}},
  \code{\link{icd9ChildrenDecimal}},
  \code{\link{icd9ChildrenShort}};
  \code{\link{icd9CondenseToMajorShort}};
  \code{\link{icd9ExpandMinor}}
}
\keyword{manip}

