% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ranges.R
\name{icd9GenerateSysData}
\alias{icd9GenerateSysData}
\title{Generate sysdata.rda}
\usage{
icd9GenerateSysData(sysdata.path = file.path("R", "sysdata.rda"))
}
\description{
Generate correctly ordered look-up tables of numeric-only, V and E codes. This is
quick, but much to slow when it appears many times in a loop.
}
\keyword{internal}

