% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/explain.R
\name{icd9GetChapters}
\alias{icd9GetChapters}
\title{get ICD-9 Chapters from vector of ICD-9 codes}
\usage{
icd9GetChapters(icd9, isShort = icd9GuessIsShort(icd9))
}
\arguments{
\item{isShort}{}

\item{icd9-any}{}

\item{invalid}{}
}
\description{
This runs quite slowly. Used too rarely to be worth optimizing
  now. This is used to build a master list of ICD-9 codes with their
  respective chapters, sub-chapters, etc.. The querying of the web page to
  get these is already done, and the results saved in the lists
  \code{icd9Chapters} etc which define ranges.
}
\keyword{internal}

