% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{icd9GetMajor}
\alias{icd9GetMajor}
\alias{icd9IsMajor}
\alias{icd9IsValidMajor}
\title{Deprecated: get or identify major part of an ICD-9 code}
\usage{
icd9GetMajor(x, isShort)

icd9IsMajor(x)

icd9IsValidMajor(icd)
}
\arguments{
\item{x}{vector of ICD-9 codes}

\item{isShort}{Deprecated. Single logical value which determines whether the
ICD-9 code provided is in short (\code{TRUE}) or decimal (\code{FALSE})
form. Where reasonable, this is guessed from the input data.}

\item{icd}{vector of ICD-9 codes}
}
\description{
Not a public function anymore.
If you really need it, use \code{icd:::icd_get_major} instead.
}

