% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{icd9GetValid}
\alias{icd9GetInvalid}
\alias{icd9GetInvalidDecimal}
\alias{icd9GetInvalidShort}
\alias{icd9GetValid}
\alias{icd9GetValidDecimal}
\alias{icd9GetValidShort}
\title{Deprecated: get valid ICD-9 codes}
\usage{
icd9GetValid(icd9, isShort = icd_guess_short(icd9))

icd9GetValidDecimal(icd9Decimal)

icd9GetValidShort(icd9Short)

icd9GetInvalid(icd9, isShort = icd_guess_short(icd9))

icd9GetInvalidDecimal(icd9Decimal)

icd9GetInvalidShort(icd9Short)
}
\arguments{
\item{icd9}{is a character vector or factor of ICD-9 codes. If fewer than five characters
is given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010".}

\item{isShort}{Deprecated. Single logical value which determines whether the
ICD-9 code provided is in short (\code{TRUE}) or decimal (\code{FALSE})
form. Where reasonable, this is guessed from the input data.}

\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five characters is given in a
code, then the digits are greedily assigned to hundreds, then tens, then
units, before the decimal parts. E.g. "10" becomes "010", not "0010"}

\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than five characters is
given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010"}
}
\description{
Use \code{icd_get_valid} instead.
}

