% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/explain.R
\name{icd9GuessIsShort}
\alias{icd9GuessIsShort}
\title{guess whether short or long}
\usage{
icd9GuessIsShort(icd9)
}
\value{
single logical value, \code{TRUE} if input data are predominantly
  short type. If there is some uncertainty, then return NA.
}
\description{
partly implemented. Goal is to guess whether codes are short or
  decimal form, then to call icd9Explain with the condense argument.
  Currently condense works, but not with the icd9 lookup table currently in
  use. Not exporting this function until it works as intended. Of note,
  validation is a bit different here, since we don't know the type until
  after we guess. We could look for where both short and long are invalid,
  and otherwise assume valid, even if the bulk are short. However, it may be
  more useful to check validity after the guess.
}
\keyword{internal}

