% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/valid.R
\name{icd9InvalidDecimal}
\alias{icd9InvalidDecimal}
\alias{icd9InvalidShort}
\title{invalid subset of decimal or short ICD-9 codes}
\usage{
icd9InvalidDecimal(icd9Decimal)

icd9InvalidShort(icd9Short)
}
\arguments{
\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five
characters is given in a code, then the digits are greedily assigned to
hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}
}
\description{
given vector of short or decimal ICD-9 codes in
  \code{icd9Decimal} or \code{icd9Short}, return (in the same format) those
  codes which are invalid. Useful for generating error messages with the
  faulty codes if validation fails.
}
\keyword{manip}

