% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9Is}
\alias{icd9Iicd9E}
\alias{icd9Is}
\alias{icd9IsE}
\alias{icd9IsN}
\alias{icd9IsV}
\title{are the given codes numeric, V or E type?}
\usage{
icd9IsV(icd9)

icd9IsE(icd9)

icd9Iicd9E(icd9)

icd9IsN(icd9)
}
\arguments{
\item{icd9}{is a character vector of ICD-9 codes. If fewer than five characters is
given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010"}
}
\description{
Quickly find V or E codes, without any validation.
}

