% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9IsA}
\alias{icd9IsA}
\title{test whether elements of vector begin with V, E (or any other
  character)}
\usage{
icd9IsA(sv, x, invert = FALSE)
}
\arguments{
\item{sv}{std::vector<std::string>&}

\item{x}{const char* of choices of first character to match}
}
\description{
Current returns a std::vector<bool> which is not thread safe,
  or particularly fast, although it is memory efficient in the standard
  borked implementation. As of icd9 version 1.2, this is not called by
  threaded code, but this could change, so beware! ASCII spaces are trimmed
  from the start of the string before testing, but no other whitesapce
}
\keyword{internal}

