% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{icd9IsBillable}
\alias{icd9GetBillable}
\alias{icd9GetBillableDecimal}
\alias{icd9GetBillableShort}
\alias{icd9GetNonBillable}
\alias{icd9GetNonBillableDecimal}
\alias{icd9GetNonBillableShort}
\alias{icd9IsBillable}
\alias{icd9IsBillableDecimal}
\alias{icd9IsBillableShort}
\title{Deprecated: get or test for billable (leaf node) ICD-9 codes}
\usage{
icd9IsBillable(icd9, isShort = icd9GuessIsShort(icd9),
  version = getLatestBillableVersion())

icd9IsBillableShort(icd9Short, version = icd9cm_latest_edition())

icd9IsBillableDecimal(icd9Decimal, version = icd9cm_latest_edition())

icd9GetBillable(icd9, isShort = icd9GuessIsShort(icd9), invert = FALSE,
  version = icd9cm_latest_edition())

icd9GetBillableShort(icd9Short, version = getLatestBillableVersion())

icd9GetBillableDecimal(icd9Decimal, version = getLatestBillableVersion())

icd9GetNonBillableShort(icd9Short, version = getLatestBillableVersion())

icd9GetNonBillableDecimal(icd9Decimal, version = getLatestBillableVersion())

icd9GetNonBillable(icd9, isShort = icd9GuessIsShort(icd9),
  version = getLatestBillableVersion())
}
\arguments{
\item{icd9}{is a character vector or factor of ICD-9 codes. If fewer than five characters
is given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010".}

\item{isShort}{Deprecated. Single logical value which determines whether the
ICD-9 code provided is in short (\code{TRUE}) or decimal (\code{FALSE})
form. Where reasonable, this is guessed from the input data.}

\item{version}{character}

\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than five characters is
given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010"}

\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five characters is given in a
code, then the digits are greedily assigned to hundreds, then tens, then
units, before the decimal parts. E.g. "10" becomes "010", not "0010"}

\item{invert}{logical}
}
\description{
Use \code{icd_is_billable} or \code{icd_get_billable} instead.
}

