% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/valid.R
\name{icd9IsN}
\alias{icd9IsE}
\alias{icd9IsN}
\alias{icd9IsV}
\title{are codes puerly numeric}
\usage{
icd9IsN(icd9)

icd9IsV(icd9)

icd9IsE(icd9)
}
\value{
logical vector
}
\description{
For each code, return \code{TRUE} if numric or \code{FALSE} if a
  V or E code.
}
\section{Functions}{
\itemize{
\item \code{icd9IsV}: 

\item \code{icd9IsE}: 
}}

