% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/valid.R
\name{icd9IsReal}
\alias{icd9GetReal}
\alias{icd9GetRealDecimal}
\alias{icd9GetRealShort}
\alias{icd9IsReal}
\alias{icd9IsRealDecimal}
\alias{icd9IsRealShort}
\title{Does ICD-9 code exist}
\usage{
icd9IsReal(icd9, isShort, majorOk = TRUE)

icd9IsRealShort(icd9Short, majorOk = TRUE)

icd9IsRealDecimal(icd9Decimal, majorOk = TRUE)

icd9GetReal(icd9, isShort, majorOk = TRUE)

icd9GetRealShort(icd9Short, majorOk = TRUE)

icd9GetRealDecimal(icd9Decimal, majorOk = TRUE)
}
\arguments{
\item{icd9}{is a character vector of ICD-9 codes. If fewer than five characters is
given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010"}

\item{isShort}{single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form.}

\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five
characters is given in a code, then the digits are greedily assigned to
hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{icd9}{is a character vector of ICD-9 codes. If fewer than five characters is
given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010"}

\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five
characters is given in a code, then the digits are greedily assigned to
hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}
}
\value{
logical vector
}
\description{
This is different from syntactic validity: it looks it up in the
  list of icd9 codes. This may have been easier all along, but checking
  syntactic validity still very useful, with a changing list of icd-9 codes
  over time, and possibly imperfections in the master list derived from CMS.
}

