% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/valid.R
\name{icd9IsValid}
\alias{icd9IsValid}
\title{check whether any ICD-9 code is syntactically valid}
\usage{
icd9IsValid(icd9, isShort)
}
\arguments{
\item{icd9}{is a character vector of ICD-9 codes. If fewer than five characters is
given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010"}

\item{isShort}{single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form. Where reasonable, this
is guessed from the input data.}
}
\description{
This does not check whether the code coresponds to a real
  ICD-9-CM billing code, or parent grouping. For that, see
  \code{\link{icd9IsReal}}.

  factors are accepted, and since the validation is done with \code{grepl}
  these are handled correctly.
}
\seealso{
\code{\link{icd9IsValidDecimal}} and \code{\link{icd9IsValidShort}}

Other ICD9 validation: \code{\link{icd9GetInvalidMappingDecimal}},
  \code{\link{icd9GetInvalidMappingShort}},
  \code{\link{icd9IsValidMapping}},
  \code{\link{icd9IsValidMappingDecimal}},
  \code{\link{icd9IsValidMappingShort}};
  \code{\link{icd9IsValidDecimal}},
  \code{\link{icd9IsValidDecimalE}},
  \code{\link{icd9IsValidDecimalN}},
  \code{\link{icd9IsValidDecimalV}};
  \code{\link{icd9IsValidMajor}},
  \code{\link{icd9IsValidMajorE}},
  \code{\link{icd9IsValidMajorN}},
  \code{\link{icd9IsValidMajorV}};
  \code{\link{icd9IsValidShort}},
  \code{\link{icd9IsValidShortE}},
  \code{\link{icd9IsValidShortN}},
  \code{\link{icd9IsValidShortV}}
}

