% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/valid.R
\name{icd9IsValidMajor}
\alias{icd9IsValidMajor}
\alias{icd9IsValidMajorE}
\alias{icd9IsValidMajorN}
\alias{icd9IsValidMajorV}
\title{validate a major part}
\usage{
icd9IsValidMajor(major)

icd9IsValidMajorN(major)

icd9IsValidMajorV(major)

icd9IsValidMajorE(major)
}
\arguments{
\item{major}{character vector of 'major' part of ICD-9 codes, i.e. that part
which falls before the decimal point, in decimal notation. (In 5 digit
notation, the 'major' part is be the first three characters (with leading
zeroes), and includes V or E prefix. xyz}
}
\description{
validation for just the 'major' part of an ICD-9 code. This can
  in fact be provided as a numeric, since there is no ambiguity. Numeric-only
  codes should be one to three digitis, V codes are followed by one or two
  digits, and E codes always by three digits between 800 and 999.

  example/test grepl(c("", "1", "22", "333", "4444", "V", "V2", "V34",
  "V567", "E", "E1", "E70", "E"))
}
\seealso{
Other ICD9 validation: \code{\link{icd9GetInvalidMappingDecimal}},
  \code{\link{icd9GetInvalidMappingShort}},
  \code{\link{icd9IsValidMapping}},
  \code{\link{icd9IsValidMappingDecimal}},
  \code{\link{icd9IsValidMappingShort}};
  \code{\link{icd9IsValidDecimal}},
  \code{\link{icd9IsValidDecimalE}},
  \code{\link{icd9IsValidDecimalN}},
  \code{\link{icd9IsValidDecimalV}};
  \code{\link{icd9IsValidShort}},
  \code{\link{icd9IsValidShortE}},
  \code{\link{icd9IsValidShortN}},
  \code{\link{icd9IsValidShortV}};
  \code{\link{icd9IsValid}}
}

