% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{icd9IsValidMapping}
\alias{icd9GetInvalidMappingDecimal}
\alias{icd9GetInvalidMappingShort}
\alias{icd9IsValidMapping}
\alias{icd9IsValidMappingDecimal}
\alias{icd9IsValidMappingShort}
\title{Deprecated: check whether a mapping has only valid ICD-9 codes}
\usage{
icd9IsValidMapping(icd9Mapping, isShort)

icd9IsValidMappingShort(icd9Mapping)

icd9IsValidMappingDecimal(icd9Mapping)

icd9GetInvalidMappingShort(icd9Mapping)

icd9GetInvalidMappingDecimal(icd9Mapping)
}
\arguments{
\item{icd9Mapping}{named list containing vectors of ICD-9 codes. E.g. the
AHRQ comorbidities, contains \code{list(OBESE = c("2780", "27800", "27801",
"27803", "V8554", "79391", "64910", "64911", "64912", "64913", "64914",
"V8530", "V8531", "V8532", "V8533", "V8534", "V8535", "V8536", "V8537",
"V8538", "V8539", "V8541", "V8542", "V8543", "V8544", "V8545" ), DEPRESS =
c("3004", "30112", "3090", "3091", "311"))} amongst other longer groups.}

\item{isShort}{Deprecated. Single logical value which determines whether the
ICD-9 code provided is in short (\code{TRUE}) or decimal (\code{FALSE})
form. Where reasonable, this is guessed from the input data.}
}
\description{
Use \code{icd_is_valid} instead.
}

