% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/valid.R
\name{icd9IsValidShort}
\alias{icd9IsValidShort}
\alias{icd9IsValidShortE}
\alias{icd9IsValidShortN}
\alias{icd9IsValidShortV}
\title{validate ICD-9 short form code}
\usage{
icd9IsValidShort(icd9Short)

icd9IsValidShortV(icd9Short)

icd9IsValidShortE(icd9Short)

icd9IsValidShortN(icd9Short)
}
\arguments{
\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}
}
\value{
logical vector with T or F for each icd9 code provided according to
  validity
}
\description{
As in \code{icd9IsValidDecimal}, character type of the
  input is enforced. An integer 100 could correspond to decimal ICD-9 codes
  1.00 10.0 or 100.
}
\seealso{
\url{http://www.stata.com/users/wgould/icd9/icd9.hlp}
  \url{http://www.sascommunity.org/wiki/Validate_the_format_of_ICD-9_codes}

Other ICD9 validation: \code{\link{icd9GetInvalidMappingDecimal}},
  \code{\link{icd9GetInvalidMappingShort}},
  \code{\link{icd9IsValidMapping}},
  \code{\link{icd9IsValidMappingDecimal}},
  \code{\link{icd9IsValidMappingShort}};
  \code{\link{icd9IsValidDecimal}},
  \code{\link{icd9IsValidDecimalE}},
  \code{\link{icd9IsValidDecimalN}},
  \code{\link{icd9IsValidDecimalV}};
  \code{\link{icd9IsValidMajor}},
  \code{\link{icd9IsValidMajorE}},
  \code{\link{icd9IsValidMajorN}},
  \code{\link{icd9IsValidMajorV}};
  \code{\link{icd9IsValid}}
}

