% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9LongToWideCpp}
\alias{icd9LongToWideCpp}
\title{Convert long to wide form as matrix}
\usage{
icd9LongToWideCpp(icd9df, visitId = "visitId", icd9Field = "icd9",
  aggregate = TRUE)
}
\description{
Take a data frame with visits and ICD codes in two columns, and convert to a matrix with one row per visit.
Since multiple rows are combined when visits are out of sequence, no guarantee is made about the returned order. We sort implicitly.
For guaranteed order, we can't de-duplicate disordered visitIds, just aggregate contiguous blocks: icd9LongOrderedToWide does this quickly.
}

