% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9MajMinToParts}
\alias{icd9MajMinToParts}
\alias{icd9MajMinToParts_list}
\title{icd9MajMinToParts}
\usage{
icd9MajMinToParts_list(mjr, mnr)
}
\description{
Convert vectors of major and minor components into a data frame with a column for major and a column for minor.
This is probably just as quick in R, but I think I may rely on having a data frame output (which is slower to construct in R)
}

