% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9MajMinToParts_df}
\alias{icd9MajMinToParts}
\alias{icd9MajMinToParts_df}
\title{icd9MajMinToParts_df}
\usage{
icd9MajMinToParts(mjr, mnr)
}
\description{
Convert vectors of major and minor components into a data frame with a column for major and a column for minor.
This is 20x faster in this C++ implementation, and important for a very frequently used function.
}

