% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9MajMinToParts_slower}
\alias{icd9MajMinToParts_slower}
\title{icd9MajMinToParts_slower}
\usage{
icd9MajMinToParts_slower(mjr, mnr)
}
\description{
Convert vectors of major and minor components into a data frame with a column for major and a column for minor.
This is slightly slower than the R implementation. using Rcpp11 std::to_string is even slower than this.
}

