% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9MajMinToShort_alt_Std}
\alias{icd9MajMinToShort_alt_Std}
\title{initialize a std::vector of strings with repeated value of the minor}
\usage{
icd9MajMinToShort_alt_Std(mjr, mnr)
}
\description{
initialize a std::vector of strings with repeated value of the minor
}
\keyword{internal}
