% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9MajMinToShortStd}
\alias{icd9MajMinToShortStd}
\title{initialize a std::vector of strings with repeated value of the minor}
\usage{
icd9MajMinToShortStd(mjr, mnr)
}
\description{
initialize a std::vector of strings with repeated value of the minor
}
\keyword{internal}

