% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/manip.R
\name{icd9ParentMinor}
\alias{icd9ParentMinor}
\title{get parent of an ICD-9 code}
\usage{
icd9ParentMinor(minor)
}
\arguments{
\item{minor}{character vector of 'minor' part of ICD-9 codes, i.e. that part
which falls after the decimal point, in decimal notation. (In 5 digit
notation, the 'major' part is be xyz, even when x and y are zero, thus the
last two characters represent the 'minor' part, in this case.)}
}
\value{
character vector of ICD-9 codes, one for each code given.
}
\description{
not likely to be useful for end-user
}
\keyword{internal}

