% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse.R
\name{icd9ParseAndSaveMappings}
\alias{icd9ParseAndSaveMappings}
\title{parse all known mappings and save to development tree}
\usage{
icd9ParseAndSaveMappings()
}
\arguments{
\item{path}{directory to save in, default is \code{data}}
}
\description{
parse all known mappings and save to development tree
}
\keyword{internal}

