% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convert.R
\name{icd9PartsRecompose}
\alias{icd9MajMinToDecimal_R}
\alias{icd9MajMinToParts_R}
\alias{icd9MajMinToShort_R}
\alias{icd9PartsRecompose}
\alias{icd9PartsToDecimal}
\alias{icd9PartsToDecimal_R}
\alias{icd9PartsToShort}
\alias{icd9PartsToShort_R}
\title{recompose major and minor parts into icd9 codes}
\usage{
icd9PartsRecompose(parts, isShort)

icd9PartsToShort_R(parts)

icd9PartsToDecimal_R(parts)

icd9MajMinToParts_R(major, minor)

icd9MajMinToShort_R(major, minor)

icd9MajMinToDecimal_R(major, minor)
}
\arguments{
\item{parts}{data.frame with major and minor fields. This can be given
instead of major and minor vectors}

\item{isShort}{single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form.}
}
\value{
character vector. Deliberately returns zero-padded major, because
  otherwise we are creating ambiguous codes (even if we know what we mean)
}
\description{
internal function which checks vector lengths to avoid
  unintentional recycling of vectors when lengths differ. Length of one is
  fine for major or minor.

icd9MajMinToDf simply composes the data frame needed
  as input to the PartsToXxxx functions

icd9MajMinTo\{Short|Decimal\} simply composes the data frame
  needed as input to the PartsToXxxx functions. Having two inputs breaks the
  ability to 'pipe' commands together using \code{magrittr}, so passing a
  single \code{data.frame} is preferred.
}
\seealso{
Other ICD-9 convert: \code{\link{icd9DecimalToShort}};
  \code{\link{icd9DropLeadingZeroes}},
  \code{\link{icd9DropLeadingZeroesDecimal}},
  \code{\link{icd9DropLeadingZeroesMajor}},
  \code{\link{icd9DropLeadingZeroesShort}};
  \code{\link{icd9ShortToDecimal_R}}
}
\keyword{internal}

