% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9RandomShort}
\alias{icd9RandomShort}
\title{Generate random short-form ICD-9 E codes}
\usage{
icd9RandomShort(n = 5L)
}
\description{
Quick pseudo-random by picking numeric, 'V' or 'E' based on modulo three of
the number
}
\keyword{internal}

