% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9RandomShortV}
\alias{icd9RandomShortV}
\title{generate random short-form icd9 V codes}
\usage{
icd9RandomShortV(n = 5L)
}
\description{
Generate a character vector of random short-form ICD-9 V codes
}
\keyword{internal}

