% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convert.R
\name{icd9ShortToDecimal_R}
\alias{icd9ShortToDecimal_R}
\title{convert short-form ICD-9 code to decimal form}
\usage{
icd9ShortToDecimal_R(icd9Short)
}
\arguments{
\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}
}
\description{
converts ICD-9 'short' form to decimal form
}
\seealso{
Other ICD-9 convert: \code{\link{icd9DecimalToShort_R}};
  \code{\link{icd9DropLeadingZeroes}},
  \code{\link{icd9DropLeadingZeroesDecimal}},
  \code{\link{icd9DropLeadingZeroesMajor}},
  \code{\link{icd9DropLeadingZeroesShort}};
  \code{\link{icd9MajMinToDecimal_R}},
  \code{\link{icd9MajMinToParts_R}},
  \code{\link{icd9MajMinToShort_R}},
  \code{\link{icd9PartsRecompose}},
  \code{\link{icd9PartsToDecimal}},
  \code{\link{icd9PartsToDecimal_R}},
  \code{\link{icd9PartsToShort}},
  \code{\link{icd9PartsToShort_R}}
}
\keyword{manip}

