% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convert.R
\name{icd9ShortToParts_R}
\alias{icd9ShortToParts_R}
\title{extract major and minor parts of a decimal ICD-9 code}
\usage{
icd9ShortToParts_R(icd9Short, minorEmpty = "")
}
\arguments{
\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{minorEmpty}{vector of length one, to be used in place of
minor part of zero. Defaults to ""}
}
\value{
data.frame with two columns. At least the minor part must be
  character, because "03" is different to "3", but "30" is the same as "3"
}
\description{
accepts Vxxxx Exxxx or xxxxx
}
\keyword{manip}

