% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ranges.R
\name{icd9SortShort}
\alias{icd9SortShort}
\title{sort short-form icd9 codes}
\usage{
icd9SortShort(icd9Short)
}
\arguments{
\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}
}
\value{
sorted vector of ICD-9 codes
}
\description{
should work with numeric only, V or E codes. Note that a numeric
  sort does not work for ICD-9 codes, since "162" > "1620" TODO: write tests.
  TODO: reply with the actual items given (not trimmed, etc.) TODO: need to
  be able to compare a pair of codes quickly, then use built-in sort. This
  becomes easier when I move to S3 classes for ICD-9.
}
\keyword{manip}

