% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{icd9Valid}
\alias{icd9IsValid}
\alias{icd9IsValidDecimal}
\alias{icd9IsValidDecimalE}
\alias{icd9IsValidDecimalN}
\alias{icd9IsValidDecimalV}
\alias{icd9IsValidMajorE}
\alias{icd9IsValidMajorN}
\alias{icd9IsValidMajorV}
\alias{icd9IsValidShort}
\alias{icd9IsValidShortE}
\alias{icd9IsValidShortN}
\alias{icd9IsValidShortV}
\alias{icd9Valid}
\alias{icd9ValidDecimal}
\alias{icd9ValidShort}
\title{Deprecated: check validity of ICD-9 codes}
\usage{
icd9Valid(icd9, isShort)

icd9IsValid(icd9, isShort)

icd9IsValidMajorN(major)

icd9IsValidMajorV(major)

icd9IsValidMajorE(major)

icd9ValidShort(icd9)

icd9IsValidShort(icd9)

icd9IsValidShortN(icd9)

icd9IsValidShortV(icd9)

icd9IsValidShortE(icd9)

icd9ValidDecimal(icd9Decimal)

icd9IsValidDecimal(icd9Decimal)

icd9IsValidDecimalN(icd9Decimal)

icd9IsValidDecimalV(icd9Decimal)

icd9IsValidDecimalE(icd9Decimal)
}
\arguments{
\item{icd9}{is a character vector or factor of ICD-9 codes. If fewer than five characters
is given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010".}

\item{isShort}{Deprecated. Single logical value which determines whether the
ICD-9 code provided is in short (\code{TRUE}) or decimal (\code{FALSE})
form. Where reasonable, this is guessed from the input data.}

\item{major}{character vector of 'major' part of ICD-9 codes, i.e. that part
which falls before the decimal point, in decimal notation. (In five digit
notation, the 'major' part is be the first three characters (with leading
zeroes), and includes V or E prefix.}

\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five characters is given in a
code, then the digits are greedily assigned to hundreds, then tens, then
units, before the decimal parts. E.g. "10" becomes "010", not "0010"}
}
\description{
Use \code{icd_is_valid} instead.
}

