% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{icd9VanWalraven}
\alias{icd9VanWalraven}
\alias{icd9VanWalravenComorbid}
\title{Deprecated: calculate Van Walraven score from ICD-9 codes}
\usage{
icd9VanWalraven(x, visitId = NULL, return.df = FALSE,
  stringsAsFactors = getOption("stringsAsFactors"), ...)

icd9VanWalravenComorbid(x, visitId = NULL, applyHierarchy = FALSE)
}
\arguments{
\item{x}{vector of ICD-9 codes}

\item{visitId}{Deprecated. Use \code{visit_name}}

\item{return.df}{Deprecated. Use \code{return_df} instead. Single logical
value, if \code{TRUE}, return the result as a data frame with the first
column being the \code{visitId}, and the second being the count. If
\code{visitId} was a factor or named differently in the input, this is
preserved.}

\item{stringsAsFactors}{Single logical value, describing whether the resulting data frame should have strings, e.g.
\code{visit_id} converted to factor. Default is to follow the current session option. This is identical to the
argument used in, among other base functions \code{as.data.frame}.}

\item{...}{arguments passed on to other functions}

\item{applyHierarchy}{single logical value that defaults to \code{TRUE}, in
which case the hierarchy defined for the mapping is applied. E.g. in
Elixhauser, you can't have uncomplicated and complicated diabetes both
flagged.}
}
\description{
Use \code{icd_van_walraven} instead.
}

