% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/valid.R
\name{icd9WarnStopMessage}
\alias{icd9WarnStopMessage}
\title{warn or stop with message based on switch}
\usage{
icd9WarnStopMessage(..., invalidAction = c("stop", "warn", "silent",
  "ignore"))
}
\arguments{
\item{...}{message components passed on to \code{warning} or \code{stop}}

\item{invalid}{}
}
\description{
Note that the default action is "stop," not "warn" here.
  "ignore" and "silent" are synonymous.
}
\keyword{internal}

