% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{icd_check_class_order}
\alias{icd_check_class_order}
\title{prefer an order of classes}
\usage{
icd_check_class_order(x)
}
\arguments{
\item{x}{any object which may or may not have classes from this package}
}
\description{
The order of classes can matter because, for some functions,
  we'd prefer to decide what to do based on whether the structure is a
  comorbidity map before caring if it is icd-9 or 10. I can't see how it
  matters whether we prioritize long/wide and short/decimal yet, so won't
  test.
}
\keyword{internal}

