% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges-icd10.R
\name{icd_children_real.icd10}
\alias{icd_children_real.icd10}
\title{real icd10 children based on 2016 ICD-10-CM list}
\usage{
\method{icd_children_real}{icd10}(x, short = icd_guess_short(x))
}
\description{
real icd10 children based on 2016 ICD-10-CM list
}
\keyword{internal}

