% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children.R
\name{icd_children_real}
\alias{icd_children_real}
\alias{icd_children_real.icd10cm}
\title{real icd10 children based on 2016 ICD-10-CM list}
\usage{
icd_children_real(x)

\method{icd_children_real}{icd10cm}(x, short_code = icd_guess_short(x))
}
\description{
real icd10 children based on 2016 ICD-10-CM list
}
\section{Methods (by class)}{
\itemize{
\item \code{icd10cm}: get the children of ICD-10 code(s)
}}
\keyword{internal}

