% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-test.R
\name{icd_classes_are_ordered}
\alias{icd_classes_are_ordered}
\title{expectation for ICD classes to be well-ordered}
\usage{
icd_classes_are_ordered()
}
\value{
a function, mimicking how \code{testthat} works
}
\description{
expectation for ICD classes to be well-ordered
}
\examples{
  testthat::expect_that(icd9("V10"), icd:::icd_classes_are_ordered())
}
\keyword{internal}

