% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{icd_classes_conflict}
\alias{icd_classes_conflict}
\title{check whether there are any ICD class conflicts}
\usage{
icd_classes_conflict(x)
}
\arguments{
\item{x}{input object to test for class conflicts}
}
\description{
E.g. both icd10who and icd10cm
}
\keyword{internal}

