% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcc.R
\name{icd_comorbid_hcc_worker}
\alias{icd_comorbid_hcc_worker}
\title{apply HCC rules to either ICD-9 or ICD-10 codes}
\usage{
icd_comorbid_hcc_worker(x, map, date_name, visit_name, icd_name)
}
\description{
apply HCC rules to either ICD-9 or ICD-10 codes
}
\keyword{internal}
\keyword{manip}
