% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{icd_comorbid_matrix_to_df}
\alias{icd_comorbid_matrix_to_df}
\title{convert matrix of comorbidities into data frame, preserving visitId
  information}
\usage{
icd_comorbid_matrix_to_df(x, visit_name = "visitId",
  stringsAsFactors = getOption("stringsAsFactors"))
}
\arguments{
\item{x}{Matrix of comorbidities, with row and columns names defined}

\item{stringsAsFactors}{whether the resulting data frame should have strings,
i.e. visitId converted to factor. Default is to follow the current session
option.}

\item{visitId}{Single character string with name for new column in output
data frame. Everywhere else, \code{visitId} describes the input data, but
here it is for output data.}
}
\examples{
longdf <- icd_long_data(
             data.frame(visitId = c("a", "b", "b", "c"),
                        icd9 = icd9(c("441", "4424", "443", "441"))))
mat <- icd_comorbid_elix(longdf)
class(mat)
typeof(mat)
rownames(mat)
df.out <- icd_comorbid_matrix_to_df(mat)
stopifnot(is.data.frame(df.out))
# output data frame has a factor for the visitId column
stopifnot(identical(rownames(mat), as.character(df.out$visitId)))
df.out[, 1:4]
}

