% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/score.R
\name{icd9CountWide}
\alias{icd9CountWide}
\alias{icd_count_codes_wide}
\title{Count ICD codes given in wide format}
\usage{
icd9CountWide(x, visitId = get_visit_name(x), return.df = FALSE,
  aggregate = FALSE)

icd_count_codes_wide(x, visit_name = get_visit_name(x), return_df = FALSE,
  aggr = FALSE)
}
\arguments{
\item{aggregate, }{single logical, default is FALSE. If TRUE, the length (or
rows) of the output will no longer match the input, but duplicate visit_names
will be counted together.}
}
\description{
For \code{icd_count_codes}, it is assumed that all the columns apart
  from \code{visit_name} represent actual or possible ICD-9 codes. Duplicate
  \code{visit_name}s are repeated as given and aggregated.
}

