% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/score.R
\name{icd9CountComorbidBin}
\alias{icd9CountComorbidBin}
\alias{icd_count_comorbid}
\title{icd_count_comorbid}
\usage{
icd9CountComorbidBin(x, visitId = get_visit_name(x), return.df = FALSE)

icd_count_comorbid(x, visit_name = get_visit_name(x), return_df = FALSE)
}
\description{
\code{icd_count_comorbid} differs from the other counting
  functions in that it counts _comorbidities_, not individual diagnoses. It
  accepts any data frame with either logicals or zero/non-zero contents, with
  a single column for visit_name. No checks are made to see whether visit_name is
  duplicated.
}

