% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{icd_decimal_to_parts}
\alias{icd_decimal_to_parts}
\alias{icd_decimal_to_parts.icd10}
\alias{icd_decimal_to_parts.icd9}
\title{Convert decimal ICD codes to component parts}
\usage{
icd_decimal_to_parts(x, minor_empty = "")

\method{icd_decimal_to_parts}{icd9}(x, minor_empty = "")

\method{icd_decimal_to_parts}{icd10}(x, minor_empty = "")
}
\description{
Convert decimal ICD codes to component parts
}
\section{Methods (by class)}{
\itemize{
\item \code{icd9}: Convert decimal ICD-9 code to parts

\item \code{icd10}: Convert decimal ICD-9 code to parts
}}
\keyword{internal}
\keyword{manip}

