% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{icd_deprecated}
\alias{icd_deprecated}
\title{wrapper for \code{.Deprecated}}
\usage{
icd_deprecated(...)
}
\arguments{
\item{...}{arguments passed to \code{.Deprecated}}
}
\description{
Don't show warnings when testing deprecated code. I don't really want
deprecated warnings in any testing, with testthat directly, or in CRAN
checks.
}
\keyword{internal}

