% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{icd_expand_range_major}
\alias{icd_expand_range_major}
\alias{icd_expand_range_major.icd10cm}
\alias{icd_expand_range_major.icd9}
\title{Expand major codes to range}
\usage{
icd_expand_range_major(start, end)

\method{icd_expand_range_major}{icd10cm}(start, end)

\method{icd_expand_range_major}{icd9}(start, end, defined = TRUE)
}
\description{
Expand a pair of major codes into a range of major codes. This was previously
exported, but is now going to be available only with \code{:::}, and
icd_expand_range will follow. This was never supposed to be a user facing
function.
}
\section{Methods (by class)}{
\itemize{
\item \code{icd10cm}: Expand range of top-level ICD-10 codes

\item \code{icd9}: Expand a range of ICD-9 major codes into
major codes
}}
\keyword{internal}

