% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explaintable.R
\name{icd_explain_table}
\alias{icd_explain_table}
\alias{icd_explain_table.default}
\title{Explain ICD-9 and ICD-10 codes in English from decimal  (123.45 style),
Tabulates the decimal format alongside converted short form (cms format)
with a boolean column to indicate if the code is parent.
A column for source year may be added in the future.}
\usage{
icd_explain_table(...)

\method{icd_explain_table}{default}(x, short_code = icd_guess_short(x),
  condense = FALSE, brief = TRUE, warn = TRUE, ...)
}
\arguments{
\item{...}{arguments passed on to other functions}

\item{x}{vector or other structure of ICD codes to explain in human language}

\item{short_code}{single logical value which determines whether the ICD-9
code provided is in short (\code{TRUE}) or decimal (\code{FALSE}) form.
Where reasonable, this is guessed from the input data.}

\item{condense}{single logical value which indicates whether to condense the
given set of ICD-9 codes by replacing subsets of codes with 'parent' codes
which exactly encompass certain subsets. E.g. If all cholera diagnoses are
provided, only '001 - Cholera' needs to be displayed, not all sub-types.}

\item{brief}{single logical value, default is \code{FALSE}. If \code{TRUE},
the short description from the canonical CMS descriptions (included in
\code{data-raw}) will be used, otherwise the long description is used.}

\item{warn}{single logical value, default is \code{TRUE}, meaning that codes
which do not correspond to diagnoses, or to three-digit codes, will trigger
a warning.}
}
\value{
data frame, or list of data frames, with fields for ICD-9 code, name
  and description. The ordering is in the same order as input.
}
\description{
Convert 'decimal' format (123.45 style) ICD-9 codes into the name and
description for human review there are official ICD9-CM data tables, not with
conversion to decimal notation, but to the textual format.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: explaining ICD codes from a character vector, guessing ICD version
}}
\keyword{internal}

