% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain-table.R
\name{icd_explain_table_worker}
\alias{icd_explain_table_worker}
\title{generate table of ICD code explanations}
\usage{
icd_explain_table_worker(x, hierarchy, short_code, condense, brief, warn, ...)
}
\description{
common code for generating full table of explanations of ICD codes
}
\author{
Ed Lee
}
\keyword{internal}

