% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.R
\name{icd_extract_alpha_numeric}
\alias{icd_extract_alpha_numeric}
\title{extract alphabetic, and numeric part of icd9 code prefix}
\usage{
icd_extract_alpha_numeric(x)
}
\arguments{
\item{icd9}{is
a character vector or factor of ICD-9 codes. If fewer than five characters
is given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010".}
}
\value{
vector or matrix, with first item of each row (or whole vector)
  containing V, E or "". The second part contains the numeric parts of the
  code, which may include a decimal point.
}
\description{
removes whitespace and separates V or E if present.
}
\keyword{internal}
\keyword{manip}

