% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd-gem.R
\name{icd_gem}
\alias{icd_gem}
\alias{icd_gem.default}
\alias{icd_gem.icd9}
\alias{icd_gem.icd9cm_pc}
\alias{icd_gem.icd10}
\alias{icd_gem.icd10cm_pc}
\title{Create An ICD-9 / ICD-10 GEM for the Specified Codes}
\usage{
icd_gem(codes, condense = FALSE)

\method{icd_gem}{default}(codes, condense = FALSE)

\method{icd_gem}{icd9}(codes, condense = FALSE)

\method{icd_gem}{icd9cm_pc}(codes, condense = FALSE)

\method{icd_gem}{icd10}(codes, condense = FALSE)

\method{icd_gem}{icd10cm_pc}(codes, condense = FALSE)
}
\arguments{
\item{codes}{A character vector or subclass of a character vector providing
codes that we wish to obtain the mappings for.  In the event that you have
procedure codes, it is necessary that \code{codes} is a subclass of either
\code{icd9cm_pc}, or \code{icd10cm_pc}.  If \code{codes} represents
diagnosis codes and does not have a subclass known to \code{icd}, then the
function will try to guess whether they are ICD-9 or ICD-10 diagnosis
codes.}

\item{condense}{Either a value of \code{TRUE} or \code{FALSE} specifying
whether entries should be condensed into a single row.}
}
\value{
A data frame that takes different forms depending on the value of
  \code{condense}.
}
\description{
Takes a set of ICD-9 or ICD-10 codes as inputs and returns a representation
of the General Equivalence Mappings (GEMs) for the target code set.  The
input codes may be either diagnosis or procedure codes.  The returned mapping
does not take the same structure as the GEMs, but instead takes an equivalent
form that is intended to be more amenable for querying.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Obtain a representation of the GEM, where the direction
of the mapping is guessed in lieu of knowing the type of the input codes.
**Warning:** that this function is only able to guess ICD-9 or ICD-10
_diagnosis_ codes, and it will never guess procedure codes.  In the event
that you do have procedure codes, then you will have to explicitly mark
your data using one of `icd9cm_pc`, `as.icd9cm_pc`, `icd10cm_pc`, or
`as.icd10cm_pc` prior to calling `icd_gem`.

\item \code{icd9}: Obtain a representation of the ICD-9 to ICD-10 GEM.

\item \code{icd9cm_pc}: Obtain a representation of the ICD-9 PC to ICD-10 PCS
GEM.

\item \code{icd10}: Obtain a representation of the ICD-10 to ICD-9 GEM.

\item \code{icd10cm_pc}: Obtain a representation of the ICD-10 PCS to ICD-9 PC
GEM.
}}

\section{Representing the GEMs (uncondensed representation)}{


The form of the mappings returned by \code{icd_gem} is slightly different
depending on the value of \code{condense} (note that the form of the mappings
is in fact the same whether you are converting from ICD-9 codes to ICD-10
codes or in the opposite direction).  We first describe the form of the
mappings for when \code{condense} has a value of \code{FALSE}, and then later
describe the form of the mappings for when \code{condense} has a value of
\code{TRUE} (the latter of which are a simple transformation of former).

The mappings returned by \code{icd_gem} when \code{condense} has a value of
\code{FALSE} take a form where each row in the data frame corresponds to a
single scenario for a given code provided in the input to \code{codes}.
Thus, there are at least as many rows as the number of elements in
\code{codes} since there is at least one row per element, and there will be
more rows than elements if any of the codes in \code{codes} have more than
one scenario.  The value of \code{source} and \code{scenario} together form a
unique key for a given row.

The columns in the returned data frame are named \code{source},
\code{scenario}, \code{type}, \code{approx}, and \code{codes}, and when the
codes are provided in a decimal form then an additional column
\code{converted} is provided which shows the converted short form
representation of the codes.  For an example of utilizing this representation
of the GEMs see the documentation for the \code{\link{icd_gem_9_to_10}}.

\itemize{

  \item \code{source}: A vector of the same type as the input to
    \code{source} providing of ICD codes corresponding to one of the input
    codes.

  \item \code{scenario}: A character vector providing the scenario for each
    row.  In the case where the input source code either doesn't have a value
    in the GEM or doesn't have a representation in the target code set,  then
    this is given a value of \code{"0"}.

  \item \code{type}: A character vector taking possible values
    \code{"simple"}, \code{"combination"}, \code{"no mapping"}, or
    \code{"not in GEM"}.  A value of \code{"simple"} means that the
    corresponding source code and scenario pair requires only a match from a
    target code to a single collection of candidate codes (i.e. it is not a
    combination code).  A value of \code{"combination"} means that a match
    must be found to each of multiple collections of candidate codes (i.e. it
    is a combination code).  A value of \code{"no mapping"} means that the
    source code doesn't have a representation in the target code set, and a
    value of \code{"not in GEM"} means that source code isn't found in the
    corresponding GEM.

  \item \code{approx}: A logical vector providing whether the mapping is an
    exact or an approximate mapping.  In the case where the input source code
    either doesn't have a value in the GEM or doesn't have a representation
    in the target code set this is given a value of \code{TRUE}.

  \item \code{codes}: A list such that each element is a list with zero or
    more character vectors of target codes, and where the elements are taken
    to have the following semantics.  For a given row (i.e. a source code and
    scenario), a target code in the target code set must be found in each
    vector in the corresponding element of \code{codes}.  For example,
    suppose that we are converting from ICD-9 codes to ICD-10 codes.  Then
    for a given row in the data frame, if the contents of \code{codes} is a
    single character vector then an ICD-10 code must be found in that vector
    for an event to have been observed corresponding to the source ICD-9 code
    for that given scenario (however, since there may be multiple scenarios
    it is possible that the event can occur based on codes in a different
    row).

}
}

\section{Representing the GEMs (condensed representation)}{


When \code{condense} has a value of \code{TRUE}, then scenarios corresponding
to the same code are combined into the same row.  This has the advantage of
providing a dataset where the rows have a one-to-one correspondence with the
inputs corresponding to \code{codes}.  In this setting, the return object is
a data frame with columns named \code{source} and \code{mapping}.

\itemize{

  \item \code{source}: A vector of the same type as the input to
    \code{source} providing of ICD codes corresponding to one of the input
    codes.

  \item \code{mapping}: A data frame providing with the number of rows equal
    to the number of scenarios for the corresponding source code, and the
    columns \code{scenario}, \code{type}, \code{approx}, and \code{codes}.
    These columns have the same meaning as the corresponding columns in the
    uncondensed representation of the GEMs.

}
}

\examples{
# specify short form ICD-9 codes
short_codes <- c("0010", "24951", "36570")
short_codes_explicit <- as.icd9cm(as.short_diag(c("0010", "24951", "36570")))
decimal_codes_explicit <- as.icd9cm(as.decimal_diag(c("001.0", "249.51", "365.70")))

# has to guess the type of codes
icd_gem(short_codes)

# use short form codes
icd_gem(short_codes_explicit)

# use decimal form codes
icd_gem(decimal_codes_explicit)

# return the condensed form version
icd_gem(decimal_codes_explicit, TRUE)
}
\seealso{
\code{\link{icd_gem_9_to_10}}.
}
