% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/real.R
\name{icd9GetBillable}
\alias{icd9GetBillable}
\alias{icd9GetBillableDecimal}
\alias{icd9GetBillableShort}
\alias{icd9GetNonBillable}
\alias{icd9GetNonBillableDecimal}
\alias{icd9GetNonBillableShort}
\alias{icd_get_billable}
\alias{icd_get_billable.default}
\alias{icd_get_billable.icd9}
\alias{icd_get_billable.icd9cm}
\title{Get billable ICD codes}
\usage{
icd9GetBillable(icd9, isShort = icd9GuessIsShort(icd9), invert = FALSE,
  version = icd9cm_latest_edition())

icd9GetBillableShort(icd9Short, version = getLatestBillableVersion())

icd9GetBillableDecimal(icd9Decimal, version = getLatestBillableVersion())

icd9GetNonBillableShort(icd9Short, version = getLatestBillableVersion())

icd9GetNonBillableDecimal(icd9Decimal, version = getLatestBillableVersion())

icd9GetNonBillable(icd9, isShort = icd9GuessIsShort(icd9),
  version = getLatestBillableVersion())

icd_get_billable(...)

\method{icd_get_billable}{default}(x, ...)

\method{icd_get_billable}{icd9cm}(x, short_code = icd_guess_short.icd9(x),
  invert = FALSE, version = icd9cm_latest_edition(), ...)

\method{icd_get_billable}{icd9}(...)
}
\arguments{
\item{isShort}{Deprecated. Single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form. Where reasonable, this
is guessed from the input data.}

\item{invert}{Single logical value. Returns the inverse of the result. E.g.
if seeking valid ICD-9 codes, the invalid ones are returned.}

\item{version}{e.g. "32", not ICD-9 vs ICD-10}

\item{x}{input vector of ICD codes}

\item{short_code}{single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form. Where reasonable, this
is guessed from the input data.}
}
\description{
Get billable ICD codes, implicitly, this refers to an ICD
  implementation which is specialized for a country, typically for billing,
  e.g. ICD-9-CM in the USA.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Get billable ICD codes, guessing ICD version,
and code short vs decimal type.

\item \code{icd9cm}: Get billable ICD-9-CM codes

\item \code{icd9}: Get billable ICD-9 codes, which is currently implemented assuming ICD-9-CM
}}
\keyword{internal}

