% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.R
\name{icd_get_defined}
\alias{icd_get_defined}
\title{Return only those codes which are heading or leaf (billable), specifying
whether codes are all short-form or all decimal-form}
\usage{
icd_get_defined(x, short_code = icd_guess_short(x), billable = FALSE)
}
\description{
Return only those codes which are heading or leaf (billable), specifying
whether codes are all short-form or all decimal-form
}

