% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{icd_get_major.icd10}
\alias{icd_get_major.icd10}
\title{initial implementation via stringr. If speed needed, then can probably reuse
C++ ICD-9 version: i'm just grabbing the first three characters, after all,
and this is much easier in ICD-10 then ICD-9}
\usage{
\method{icd_get_major}{icd10}(x)
}
\description{
initial implementation via stringr. If speed needed, then can probably reuse
C++ ICD-9 version: i'm just grabbing the first three characters, after all,
and this is much easier in ICD-10 then ICD-9
}
\keyword{internal}

