% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{icd_get_major}
\alias{icd_get_major}
\alias{icd_get_major.icd10}
\title{Get major part of an ICD code}
\usage{
icd_get_major(x)

\method{icd_get_major}{icd10}(x)
}
\description{
Returns major component of codes, i.e. the part before the decimal,
or where the decimal would be.
}
\details{
For ICD-10, this is an initial implementation. If speed needed, then
  can re-use C++ ICD-9 version: just grabbing the first three characters,
  after all, and this is much easier in ICD-10 then ICD-9
}
\section{Methods (by class)}{
\itemize{
\item \code{icd10}: Get major part of an ICD-10 code
}}
\keyword{internal}

