% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{icd_get_missing_kids}
\alias{icd_get_missing_kids}
\title{Find children of a code which are not in a given environment}
\usage{
icd_get_missing_kids(code, env, defined)
}
\description{
Find children of a code which are not in a given environment
}
\keyword{internal}

