% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/guess.R
\name{icd9GuessIsShort}
\alias{icd9GuessIsShort}
\alias{icd_guess_short}
\alias{icd_guess_short.data.frame}
\alias{icd_guess_short.icd9}
\title{guess whether codes are \code{short_code} or \code{decimal_code}}
\usage{
icd9GuessIsShort(icd9)

icd_guess_short(x, short_code = NULL, test_n = 1000L)

\method{icd_guess_short}{data.frame}(x, short_code = NULL, test_n = 1000L,
  icd_name = get_icd_name(x))

\method{icd_guess_short}{icd9}(x, short_code = NULL, test_n = 1000L)
}
\value{
single logical value, \code{TRUE} if input data are predominantly
  short_code type. If there is some uncertainty, then return NA.
}
\description{
TODO: Partly implemented. Goal is to guess whether codes are
  short_code or decimal form. Currently condense works, but not with the icd9
  lookup table currently in use. Not exporting this function until it works
  as intended. Of note, validation is a bit different here, since we don't
  know the type until after we guess. We could look for where both short_code
  and long are invalid, and otherwise assume valid, even if the bulk are
  short_code. However, it may be more useful to check validity after the
  guess.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Guess whether a data frame has ICD-9 or ICD-10
codes

\item \code{icd9}: Guess whether an ICD-9 code is in short_code form
}}
\keyword{internal}

