% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/real.R
\name{icd9IsBillable}
\alias{icd9IsBillable}
\alias{icd9IsBillableDecimal}
\alias{icd9IsBillableShort}
\alias{icd_is_billable}
\alias{icd_is_billable.default}
\alias{icd_is_billable.icd10}
\alias{icd_is_billable.icd10cm}
\alias{icd_is_billable.icd9}
\alias{icd_is_billable.icd9cm}
\title{Determine whether codes are billable leaf-nodes}
\usage{
icd9IsBillable(icd9, isShort = icd9GuessIsShort(icd9),
  version = getLatestBillableVersion())

icd9IsBillableShort(icd9Short, version = icd9cm_latest_edition())

icd9IsBillableDecimal(icd9Decimal, version = icd9cm_latest_edition())

icd_is_billable(...)

\method{icd_is_billable}{icd9}(x, short_code = icd_guess_short.icd9(x),
  version = icd9cm_latest_edition(), ...)

\method{icd_is_billable}{icd10cm}(x, short_code = icd_guess_short.icd10(x),
  version = "2016", ...)

\method{icd_is_billable}{icd10}(x, short_code = icd_guess_short.icd10(x),
  version = "2016", ...)

\method{icd_is_billable}{icd9cm}(x, short_code = icd_guess_short(x),
  version = icd9cm_latest_edition(), ...)

\method{icd_is_billable}{default}(x, short_code = icd_guess_short(x),
  version = icd9cm_latest_edition(), ...)
}
\arguments{
\item{icd9}{is
a character vector or factor of ICD-9 codes. If fewer than five characters
is given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010".}

\item{version}{single character string, default is "32" which is the latest
release from CMS. Currently anything from "23" to "32" is accepted. Not
numeric because there are possible cases with non-numeric names, e.g.
revisions within one year, although none currently implemented.}

\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five
characters is given in a code, then the digits are greedily assigned to
hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{short_code}{single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form. Where reasonable, this
is guessed from the input data.}

\item{invert}{Single logical value. Returns the inverse of the result. E.g.
if seeking valid ICD-9 codes, the invalid ones are returned.}
}
\value{
logical vector of same length as input
}
\description{
Codes provided are compared to the most recent version of the
  CMS list of billable codes, or another version if specified.
}
\section{Methods (by class)}{
\itemize{
\item \code{icd9}: Which of the given ICD-9 codes are leaf nodes in
ICD-9-CM. Currently assumes ICD-9 codes are ICD-9-CM

\item \code{icd10cm}: Which of the given ICD-10 codes are leaf nodes in
ICD-10-CM. Currently assumes ICD-10 codes are ICD-10-CM

\item \code{icd10}: Which of the given ICD-10 codes are leaf nodes in
ICD-10-CM. Currently assumes ICD-10 codes are ICD-10-CM

\item \code{icd9cm}: Which of the given ICD-9 codes are leaf nodes in
ICD-9-CM

\item \code{default}: Which of the given ICD codes are leaf nodes in
ICD version guessed from the codes themselves.
}}
\keyword{internal}

