% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/real.R
\name{icd9IsReal}
\alias{icd9GetReal}
\alias{icd9GetRealDecimal}
\alias{icd9GetRealShort}
\alias{icd9IsReal}
\alias{icd9IsRealDecimal}
\alias{icd9IsRealShort}
\alias{icd_is_defined}
\title{Check whether ICD-9 codes exist}
\usage{
icd9IsReal(icd9, isShort = icd_guess_short(icd9), onlyBillable = FALSE)

icd9IsRealShort(icd9Short, onlyBillable = FALSE)

icd9IsRealDecimal(icd9Decimal, onlyBillable = FALSE)

icd9GetReal(icd9, isShort = icd_guess_short.icd9(icd9),
  onlyBillable = FALSE)

icd9GetRealShort(icd9Short, onlyBillable = FALSE)

icd9GetRealDecimal(icd9Decimal, onlyBillable = FALSE)

icd_is_defined(x, short_code = icd_guess_short(x), ...)
}
\arguments{
\item{icd9}{is
a character vector or factor of ICD-9 codes. If fewer than five characters
is given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010".}

\item{onlyBillable}{single logical value, if \code{TRUE}, describes the input
data, stating that it only contains billable codes. Usually, the function
will try to guess this, but if you know in advance what they should be, the
functions can optionally warn if this is incorrect, and save some
computation time. The billable codes are derived from the CMS list. The
most recent version is used by default.}

\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five
characters is given in a code, then the digits are greedily assigned to
hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{short_code}{single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form. Where reasonable, this
is guessed from the input data.}

\item{...}{arguments passed on to other functions}

\item{billable}{single logical value (default \code{FALSE}), if
\code{TRUE} will divert to test whether the codes are in the billable list
instead of seeing if they are any leaf or branch node. TODO: template}
}
\value{
logical vector
}
\description{
This is different from syntactic validity: it looks it up in the
  canonical list of ICD-9 codes published by the CMS, and which are included
  in this package under \code{data-raw}. Checking syntactic validity using
  \code{link{icd9IsValid}} etc. is still useful, with a changing list of
  icd-9 codes over time, and possible imperfections in the master lists
  derived from CMS.
}

