% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{icd_is_major}
\alias{icd_is_major}
\alias{icd_is_major.icd10}
\alias{icd_is_major.icd10cm}
\alias{icd_is_major.icd10who}
\alias{icd_is_major.icd9}
\title{check whether a code is major}
\usage{
icd_is_major(icd)

\method{icd_is_major}{icd10}(icd)

\method{icd_is_major}{icd10cm}(icd)

\method{icd_is_major}{icd10who}(icd)

\method{icd_is_major}{icd9}(icd9)
}
\arguments{
\item{icd}{character vector of ICD codes.}
}
\description{
move to C++ or own R file:
}
\section{Methods (by class)}{
\itemize{
\item \code{icd10}: check whether a code is an ICD-10 major

\item \code{icd10cm}: check whether a code is an ICD-10-CM major

\item \code{icd10who}: check whether a code is an ICD-10 WHO major

\item \code{icd9}: check whether a code is an ICD-9 major
}}
\keyword{internal}

