% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{icd_is_valid.character}
\alias{icd_is_valid.character}
\title{valid, guessing both type and version of the ICD codes}
\usage{
icd_is_valid.character(x, short_code = icd_guess_short(x),
  whitespace_ok = TRUE, ...)
}
\description{
valid, guessing both type and version of the ICD codes
}

