% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid-icd10.R
\name{icd_is_valid}
\alias{icd_is_valid}
\alias{icd_is_valid.icd10}
\title{Test whether ICD-9 or ICD-10 codes are valid}
\usage{
icd_is_valid(x)

\method{icd_is_valid}{icd10}(icd)
}
\arguments{
\item{icd}{An ICD-9 or 10 code. If the class is set to \code{'icd9'},
\code{'icd10'}, \code{'icd10cm'} etc then perform appropriate validation.}
}
\description{
TODO: add default (when there is no class) which detected icd9 vs 10 if
possible. TODO: use "short" or "long" attribute if available to tighten
validation, or guess if missing.
}
\section{Methods (by class)}{
\itemize{
\item \code{icd10}: Test whether generic ICD-10 code is valid
}}

