% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/valid.R
\name{icd9IsValidMajor}
\alias{icd9IsValidMajor}
\alias{icd9_is_valid_major_e}
\alias{icd9_is_valid_major_n}
\alias{icd9_is_valid_major_v}
\alias{icd_is_valid_major}
\alias{icd_is_valid_major.icd9}
\title{test whether codes are of major type}
\usage{
icd9IsValidMajor(icd)

icd_is_valid_major(x)

\method{icd_is_valid_major}{icd9}(major)

icd9_is_valid_major_n(major)

icd9_is_valid_major_v(major)

icd9_is_valid_major_e(major)
}
\arguments{
\item{x}{ICD codes}
}
\description{
codes without real or implied decimal place return TRUE
}
\section{Methods (by class)}{
\itemize{
\item \code{icd9}: Test whether an ICD-9 code is of major type.
}}
\keyword{internal}

