% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{icd_long_to_wide}
\alias{icd9LongToWide}
\alias{icd_long_to_wide}
\title{convert ICD data from long to wide format}
\usage{
icd_long_to_wide(x, visit_name = get_visit_name(x),
  icd_name = get_icd_name(x), prefix = "icd_", min.width = 0,
  aggregate = TRUE, return_df = FALSE)

icd9LongToWide(icd9df, visitId = NULL, icd9Field = NULL, prefix = "icd_",
  min.width = 0, aggregate = TRUE, return.df = FALSE)
}
\arguments{
\item{x}{data.frame of long-form data, one column for visit_name and one
for ICD code}

\item{visit_name}{The name of the column in the data frame which contains the
patient or visit identifier. Typically this is the visit identifier, since
patients come leave and enter hospital with different ICD-9 codes. It is a
character vector of length one. If left empty, or \code{NULL}, then an
attempt is made to guess which field has the ID for the patient encounter
(not a patient ID, although this can of course be specified directly). The
guesses proceed until a single match is made. Data frames may be wide with
many matching fields, so to avoid false positives, anything but a single
match is rejected. If there are no successful guesses, and \code{visitId}
was not specified, then the first column of the data frame is used.}

\item{icd_name}{The column in the data frame which contains the ICD codes.
This is a character vector of length one. If it is \code{NULL}, \code{icd9}
will attempt to guess the column name, looking for progressively less
likely possibilities until it matche a single column. Failing this, it will
take the first column in the data frame. Specifying the column using this
argument avoids the guesswork.}

\item{prefix}{character, default "icd_" to prefix new columns}

\item{min.width, }{single integer, if specified, writes out this many columns
even if no patients have that many codes. Must be greater than or equal to
the maximum number of codes per patient.}

\item{aggregate}{single logical value, if TRUE (the default) will take more
time to find out-of-order visit_names, and combine all the codes for each
unique visit_name. If \code{FALSE}, then out-of-order visit_names will result in
a row in the output data per contiguous block of identical visit_names.}

\item{return_df}{single logical value, if \code{TRUE}, return a data frame
with a field for the visit_name. This may be more convenient, but the default
of \code{FALSE} gives the more natural return data of a matrix with
rownames being the visit_names.}

\item{visitId}{Deprecated. Use \code{visit_name}}

\item{icd9Field}{Deprecated. Use \code{icd_name}}
}
\description{
This is more complicated than reshape or reshape2::dcast allows.
  This is a reasonably simple solution using built-in functions.
}
\examples{
  longdf <- data.frame(visit_name = c("a", "b", "b", "c"),
    icd9 = c("441", "4424", "443", "441"))
  icd9LongToWide(longdf)
  icd9LongToWide(longdf, prefix = "ICD10_")
}
\seealso{
Other ICD-9 convert: \code{\link{icd9ChaptersToMap}},
  \code{\link{icd9DropLeadingZeroes}},
  \code{\link{icd9PartsToShort}},
  \code{\link{icd_wide_to_long}}
}
\keyword{manip}

