% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{icd_map_cc_hcc}
\alias{icd_map_cc_hcc}
\title{Map of hierarchical condition codes}
\format{An object of class \code{data.frame} with 540 rows and 8 columns.}
\usage{
icd_map_cc_hcc
}
\description{
See \code{icd9_map_cc}
}
\keyword{datasets}

