% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{icd_names_cc}
\alias{icd_names_cc}
\title{Names of hierarchical condition codes}
\format{An object of class \code{list} of length 119.}
\usage{
icd_names_cc
}
\description{
See \code{icd9_map_cc}
}
\keyword{datasets}

