% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-hcc.R
\name{icd_parse_cc_labels}
\alias{icd_parse_cc_labels}
\title{Import the CMS CC Labels}
\usage{
icd_parse_cc_labels(save_data = FALSE)
}
\arguments{
\item{save_data}{logical whether to save the result in the source tree.
Defaults to \code{FALSE}.}

\item{condense}{Deprecated. The map is not condensed by default. If required,
it can reduced to a smaller set of codes using \code{icd_condense}.}
}
\value{
invisibly returns the list of vectors, where a co-morbidity name is
  associated with a character vector of ICD-9 codes.
}
\description{
The CMS Condition Category labels were obtained from CMS \url{https://www.cms.gov/Medicare/Health-Plans/MedicareAdvtgSpecRateStats/Risk-Adjustors.html}.
Due to the complex file structure of the original data (many nested zip files), they have
been organized in the folder data/icd_hcc_rawdata/labels
This function creates an .RData file containing labels for CC and HCCs
}
\keyword{internal}

