% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{icd_set_decimal_code}
\alias{icd_set_decimal_code}
\title{set class to describe decimal form ICD codes, i.e. with a decimal place}
\usage{
icd_set_decimal_code(x)
}
\description{
set class to describe decimal form ICD codes, i.e. with a decimal place
}

