% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{icd_set_map}
\alias{icd_set_map}
\title{set class to describe a comorbidity map}
\usage{
icd_set_map(x)
}
\description{
This, I think, should take priority over ICD-9 vs ICD-10 when doing S3 dispatching
Maybe should call this class icd_comorbidity_map instead
}

