% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{icd_set_short_code}
\alias{icd_set_short_code}
\title{set class to describe short form ICD codes, i.e. with no decimal place}
\usage{
icd_set_short_code(x)
}
\description{
set class to describe short form ICD codes, i.e. with no decimal place
}

