% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd_short_to_decimal.icd9}
\alias{icd_short_to_decimal.icd9}
\title{Convert ICD-9 codes between short and decimal forms}
\usage{
icd_short_to_decimal.icd9(x)
}
\arguments{
\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five
characters is given in a code, then the digits are greedily assigned to
hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}
}

