% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{icd_short_to_decimal}
\alias{icd_short_to_decimal}
\title{Convert ICD from short to decimal forms}
\usage{
icd_short_to_decimal(x)
}
\arguments{
\item{x}{ICD codes}
}
\description{
Convert codes between short and decimal forms
}

