% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{icd_short_to_parts}
\alias{icd_decimal_to_parts.character}
\alias{icd_short_to_parts}
\alias{icd_short_to_parts.character}
\alias{icd_short_to_parts.icd9}
\title{Convert short format ICD codes to component parts}
\usage{
icd_short_to_parts(x, minor_empty = "")

\method{icd_short_to_parts}{icd9}(x, minor_empty = "")

\method{icd_short_to_parts}{character}(x, minor_empty = "")

\method{icd_decimal_to_parts}{character}(x, minor_empty = "")
}
\description{
Convert short format ICD codes to component parts
}
\section{Methods (by class)}{
\itemize{
\item \code{icd9}: Convert short format ICD-9 code to parts

\item \code{character}: Convert short format ICD code to parts,
guessing whether ICD-9 or ICD-10

\item \code{character}: Convert decimal ICD code to parts, guessing
ICD version
}}
\keyword{internal}
\keyword{manip}

