% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{is.icd9}
\alias{is.icd9}
\alias{is.icd9cm}
\alias{is.icd9cm_pc}
\alias{is.icd9who}
\alias{is.icd10}
\alias{is.icd10cm}
\alias{is.icd10cm_pc}
\alias{is.icd10who}
\alias{is.icd10fr}
\alias{is.comorbidity_map}
\title{Test presence of ICD classes}
\usage{
is.icd9(x)

is.icd9cm(x)

is.icd9cm_pc(x)

is.icd9who(x)

is.icd10(x)

is.icd10cm(x)

is.icd10cm_pc(x)

is.icd10who(x)

is.icd10fr(x)

is.comorbidity_map(x)
}
\arguments{
\item{x}{Any object which may have ICD-related classes set}
}
\description{
This merely checks whether the given object is a certain type of ICD code, it
does no validation of any kind. For validation, see \code{\link{is_valid}}.
}
\examples{
# A character string is not itself an ICD code
is.icd9("100.1")
is_valid("100.1")
is.icd9(as.icd9cm("100.1"))
}
