% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{is.icd_short_diag}
\alias{is.icd_short_diag}
\title{Get or set whether ICD codes have have an attribute indicating 'short' or
'decimal' format}
\usage{
is.icd_short_diag(x, must_work = FALSE)
}
\arguments{
\item{x}{ICD data}

\item{must_work}{single logical value, if \code{FALSE} (the default) this may
return \code{NULL} if the attribute is not present. If \code{TRUE}, then
either \code{TRUE} or \code{FALSE} is returned.}

\item{value}{\code{TRUE} or \code{FALSE}, default is \code{TRUE} which sets
the attribute to whatever is indicated in the function name. See examples.}
}
\description{
Get or set whether ICD codes have have an attribute indicating 'short' or
'decimal' format
}
\section{Getting the attribute}{
 \code{is.short_diag} tests for presence of an attribute, not whether the code
is a valid ICD code. If \code{must_work} is \code{TRUE} then \code{NULL}
(i.e. no attribute set) returns \code{FALSE}, otherwise \code{NULL} is
returned. To test validity, see \code{\link{is_valid}}.
}

\section{Setting the attribute}{
 Similarly, \code{as.icd_short_diag} and \code{as.icd_decimal_diag} set the
attribute, but do not convert the codes themselves. For conversion between
'short' and 'decimal' forms, use \code{\link{decimal_to_short}} and
\code{\link{short_to_decimal}}.

The attribute \code{icd_short_code} should be either \code{TRUE} or
\code{FALSE}. There is no attribute named \code{icd_decimal_code}. These
functions set and get the attribute safely.
}

\examples{
as.icd_short_diag("6670")

as.icd_short_diag("667.0") # no warning or error!

is.icd_short_diag(decimal_to_short("667.0"))

decimal_type_code <- as.icd_short_diag("667.0", FALSE)
stopifnot(is.icd_decimal_diag(decimal_type_code))

codes <- as.icd9(c("100.1", "441.3"))
codes <- as.decimal_diag(codes)
codes
}
