% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{is.short_diag}
\alias{is.short_diag}
\alias{is.decimal_diag}
\alias{is.icd_short_diag}
\alias{is.icd_decimal_diag}
\alias{as.short_diag}
\alias{as.decimal_diag}
\alias{as.icd_short_diag}
\alias{as.icd_decimal_diag}
\title{Get or set whether ICD codes have have an attribute indicating 'short' or 'decimal' format}
\usage{
is.short_diag(x, must_work = FALSE)

is.decimal_diag(x, must_work = FALSE)

is.icd_short_diag(x, must_work = FALSE)

is.icd_decimal_diag(x, must_work = FALSE)

as.short_diag(x, value = TRUE)

as.decimal_diag(x, value = TRUE)

as.icd_short_diag(x, value = TRUE)

as.icd_decimal_diag(x, value = TRUE)
}
\arguments{
\item{x}{ICD data}

\item{must_work}{single logical value, if \code{FALSE} (the default) this may
return \code{NULL} if the attribute is not present. If \code{TRUE}, then
either \code{TRUE} or \code{FALSE} is returned.}

\item{value}{\code{TRUE} or \code{FALSE}, default is \code{TRUE} which sets the attribute to whatever is indicated in the function name. See examples.}
}
\description{
\code{is.short_diag} tests for presence of an attribute, not whether the code
is a valid ICD code. If \code{must_work} is \code{TRUE} then \code{NULL}
(i.e. no attribute set) returns \code{FALSE}, otherwise \code{NULL} is
returned.
}
\details{
The attribute \code{icd_short_code} should be either \code{TRUE} or
\code{FALSE}. There is no attribute named \code{icd_decimal_code}. These
functions set and get the attribute safely.
}
\examples{
as.icd_short_diag("6670")
decimal_type_code <- as.icd_short_diag("667.0", FALSE)
stopifnot(is.icd_decimal_diag(decimal_type_code))

codes <- as.icd9(c("100.1", "441.3"))
codes <- as.decimal_diag(codes)
codes
}
