% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{listTrimFlat}
\alias{listTrimFlat}
\title{trim null or empty values from a list}
\usage{
listTrimFlat(x)
}
\arguments{
\item{x}{list}
}
\value{
trimmed list
}
\description{
For a given unnested list, x, return the list without any NA or NULL elements.
}
\keyword{internal}

