% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lookupComorbid_alt_ByChunkForTaskloop}
\alias{lookupComorbid_alt_ByChunkForTaskloop}
\title{alternate comorbidity search}
\usage{
lookupComorbid_alt_ByChunkForTaskloop(vcdb, map, out)
}
\description{
alternate version using much simplified with Openmp taskloop, only in OMP4.5
}
\keyword{internal}
