% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lookupComorbidByChunkFor}
\alias{lookupComorbidByChunkFor}
\title{core search for ICD code in a map}
\usage{
lookupComorbidByChunkFor(vcdb, map, chunkSize, ompChunkSize, out)
}
\description{
core search for ICD code in a map
}
\keyword{internal}
