% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lookupComorbidByChunkForTaskloop}
\alias{lookupComorbidByChunkForTaskloop}
\title{alternate comorbidity search}
\usage{
lookupComorbidByChunkForTaskloop(vcdb, map, out)
}
\description{
alternate version using much simplified with Openmp taskloop, only in OMP4.5
}
\keyword{internal}
