% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lookupComorbidSparse}
\alias{lookupComorbidSparse}
\title{comorbidity search with sparse matrix result}
\usage{
lookupComorbidSparse(vcdb, map)
}
\description{
Much less memory competition in writing output. As an example the Vermont
data has 29,000 comorbidity flags (29 for each patient) Whereas only 2367
AHRQ comorbidity are positive, so under 10%.
}
\keyword{internal}
