% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/util.R
\name{match_rcpp}
\alias{match_rcpp}
\alias{fin}
\alias{\%fin\%}
\alias{\%fnin\%}
\title{Faster match}
\usage{
match_rcpp(x, table)

fin(x, table)

x \%fin\% table

x \%fnin\% table
}
\description{
Faster match
}
\section{Functions}{
\itemize{
\item \code{fin}: Use faster matching for %in% equivalent.

\item \code{\%fin\%}: Use faster matching for %in%

\item \code{\%fnin\%}: Use faster matching for %nin%
}}

\keyword{internal}
