% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{named_list}
\alias{named_list}
\title{make a list using input argument names as names}
\usage{
named_list(...)
}
\arguments{
\item{...}{arguments whose names become list item names, and values become
the values in the list}
}
\description{
make a list using input argument names as names
}
\keyword{internal}

