% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdocs.R
\docType{data}
\name{names_elix}
\alias{names_elix}
\alias{names_elix_abbrev}
\alias{names_elix_htn}
\alias{names_elix_htn_abbrev}
\alias{names_quan_elix}
\alias{names_quan_elix_abbrev}
\alias{names_quan_elix_htn}
\alias{names_quan_elix_htn_abbrev}
\alias{names_ahrq}
\alias{names_ahrq_abbrev}
\alias{names_ahrq_htn}
\alias{names_ahrq_htn_abbrev}
\alias{names_charlson}
\alias{names_charlson_abbrev}
\alias{names_cc}
\alias{icd_names_elix}
\alias{icd_names_elix_abbrev}
\alias{icd_names_elix_htn}
\alias{icd_names_elix_htn_abbrev}
\alias{icd_names_quan_elix}
\alias{icd_names_quan_elix_abbrev}
\alias{icd_names_quan_elix_htn}
\alias{icd_names_quan_elix_htn_abbrev}
\alias{icd_names_ahrq}
\alias{icd_names_ahrq_abbrev}
\alias{icd_names_ahrq_htn}
\alias{icd_names_ahrq_htn_abbrev}
\alias{icd_names_charlson}
\alias{icd_names_charlson_abbrev}
\alias{icd_names_cc}
\alias{icd9_names_multi_ccs}
\alias{icd9_names_single_ccs}
\alias{icd10_names_ccs}
\alias{names_pccc}
\alias{names_pccc_abbrev}
\title{Comorbidity names}
\format{
list, with character/numeric code. 'Hypertension,
  uncomplicated' and 'Hypertension, complicated' are labelled '6a'
  and '6b'. Diabetes, cancer, and metastasis are counted
  independently, as in the original paper, giving the original 30
  groups. "01" to "30"
}
\description{
These lists provide correctly sorted names of the comorbidities and
their particular permutations in both full and abbreviated forms.
}
\details{
In the Elixhauser derived mappings, uncomplicated and complicated
hypertension are listed separately, but are always combined in the
final analyses. Uncomplicated and complicated hypertension are list
separately and as "Hypertension, combined." _abbrev suffix indicates
a very short space-free description. Quan's version of Elixhauser is
identical. AHRQ updates drops the arrythmia field. The naming
convention is a root, e.g. \code{icd9_map_elix} or
\code{icd10_map_elix}, with neither/either/both suffixes \code{_htn}
and \code{_abbrev}. The Charlson derived mappings do not include
hypertension. _abbreviated comorbidity names are helpful for
interactive work, whereas the full names might be preferred for
plotting.
}
\keyword{datasets}
