% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_leaf_desc_icd9cm_v27}
\alias{parse_leaf_desc_icd9cm_v27}
\title{Parse billable codes for ICD-9-CM version 27}
\usage{
parse_leaf_desc_icd9cm_v27(offline = FALSE)
}
\arguments{
\item{offline}{logical value}
}
\description{
These have a quirk which needs a different approach
}
\keyword{internal}

