% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse.R
\name{parseIcd9Chapters}
\alias{parseIcd9Chapters}
\title{Read higher-level ICD-9 structure from a reliable web site}
\usage{
parseIcd9Chapters(year = NULL, save = FALSE)
}
\description{
Previous iteration attempted to use the canonical data from
  annual RTF files from the CDC, however, this was ridiculously tricky and
  error prone, so now using a reliable website and scraping. Will still
  confirm the results with tests.
}
\keyword{internal}

