% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse.R
\name{parseIcd9LeafDescriptionsAll}
\alias{parseIcd9LeafDescriptionsAll}
\title{get billable codes from all available years}
\source{
http://www.cms.gov/Medicare/Coding/ICD9ProviderDiagnosticCodes/codes.html
}
\usage{
parseIcd9LeafDescriptionsAll(save = FALSE)
}
\arguments{
\item{save}{single logical value, if \code{TRUE} the source text or CSV file
will be saved in \code{inst/extdata}, otherwise (the default) the data is
simply returned invisibly.}
}
\value{
data frame with icd9, descShort and descLong columns. NA is placed in
  descLong when not available.
}
\description{
for versions 23 to 32, those which are on the CMS web site, get
  any codes with long or short descriptions. Earlier years only have
  abbreviated descriptions.
}
\keyword{internal}

