% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parseLeafDescriptionsAll}
\alias{parseLeafDescriptionsAll}
\title{get billable codes from all available years}
\source{
http://www.cms.gov/Medicare/Coding/ICD9ProviderDiagnosticCodes/codes.html
}
\usage{
parseLeafDescriptionsAll(save_data = FALSE, fromWeb = FALSE, save = NULL)
}
\arguments{
\item{save_data}{single logical value, if \code{TRUE} the source text or CSV file
will be saved in \code{data-raw}, otherwise (the default) the data is
simply returned invisibly.}
}
\value{
data frame with icd9, descShort and descLong columns. NA is placed in
  descLong when not available.
}
\description{
for versions 23 to 32, those which are on the CMS web site, get
  any codes with long or short descriptions. Earlier years only have
  abbreviated descriptions.
}
\examples{
  # To populate the data-raw directory with the ICD-9 source:
  # not included in installed package, run using the full source from github,
  # e.g. using devtools::load_all()
  \dontrun{
  parseLeafDescriptionsAll(save_data = TRUE, fromWeb = TRUE)
  }
}
\keyword{internal}

