% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-rtf.R
\name{parseRtfFifthDigitRanges}
\alias{parseRtfFifthDigitRanges}
\title{parse a row of RTF source data for ranges to apply fifth digit
  sub-classifications}
\usage{
parseRtfFifthDigitRanges(row_str, verbose = FALSE)
}
\arguments{
\item{verbose}{single logical value, defaults to \code{FALSE} in most functions.}
}
\description{
returns all the possible 5 digit codes encompassed by the given
  definition. This needs to be whittled down to just those matching fifth
  digits, but we haven't parsed them yet.
}
\keyword{internal}

