% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-rtf.R
\name{parseRtfYear}
\alias{parseRtfYear}
\title{parse RTF description of entire ICD-9-CM for a specific year}
\source{
http://ftp.cdc.gov/pub/Health_Statistics/NCHS/Publications/ICD9-CM/2011/Dtab12.zip
  and similar files run from 1996 to 2011.
}
\usage{
parseRtfYear(year = "2011", save = FALSE, fromWeb = FALSE,
  verbose = FALSE)
}
\arguments{
\item{year}{from 1996 to 2012 (this is what CDC has published). Only 2012
implemented thus far}

\item{verbose}{single logical value, defaults to \code{FALSE} in most functions.}
}
\description{
Currently only the most recent update is implemented. Note that
  CMS have published additional ICD-9-CM billable code lists since the last
  one from the CDC: I think these have been the same every year since 2011,
  though. THe last CDC release is Dtab12.rtf from 2011.
}
\keyword{internal}

