% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbid.R
\docType{data}
\name{poa_choices}
\alias{poa_choices}
\title{Present-on-admission flags}
\format{
An object of class \code{character} of length 4.
}
\usage{
poa_choices
}
\description{
See \link{filter_poa} for more details.
}
\examples{
poa_choices
}
\keyword{character}
