% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.icd_comorbidity_map.R
\name{print.icd_comorbidity_map}
\alias{print.icd_comorbidity_map}
\title{Print a comorbidity map}
\usage{
\method{print}{icd_comorbidity_map}(x, n_comorbidities = 7, n_codes = 7)
}
\description{
The default is to summarize by printing the first seven comorbidities, and
the first seven codes for each. To print the whole thing, just convert it to
a list.
}
\examples{
print(icd9_map_ahrq)
print(icd9_map_ahrq, n_comorbidities = 3, n_codes = 3)
\dontrun{
print.list(icd9_map_ahrq)
print(list(icd9_map_ahrq))
}
}

