% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{quanElixComorbid}
\alias{quanElixComorbid}
\title{Quan adaptation of Elixhauser comorbidities}
\format{list of character vectors, each named by co-morbidity}
\description{
These were transcribed directly from the Quan paper referenced.
}
\references{
Quan, Hude, Vijaya Sundararajan, Patricia Halfon, Andrew Fong,
  Bernard Burnand, Jean-Christophe Luthi, L. Duncan Saunders, Cynthia A.
  Beck, Thomas E. Feasby, and William A. Ghali. "Coding Algorithms for
  Defining Comorbidities in ICD-9-CM and ICD-10 Administrative Data." Medical
  Care 43, no. 11 (November 1, 2005): 1130-39.
  \url{http://www.ncbi.nlm.nih.gov/pubmed/16224307}
  \url{http://web.archive.org/web/20110225042437/http://www.chaps.ucalgary.ca/sas}
}
\keyword{datasets}

