% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-test.R
\name{random_string}
\alias{random_string}
\title{generate random strings}
\usage{
random_string(n, max_chars = 4)
}
\description{
Mixed upper and lower case, with replacement
}
\keyword{datagen}
\keyword{debugging}
\keyword{internal}

