% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rbind_with_empty}
\alias{rbind_with_empty}
\title{fast expand of logical matrix to add rows filled with false}
\usage{
rbind_with_empty(a, b_rows)
}
\description{
fast expand of logical matrix to add rows filled with false
}
\examples{
nrow = 1e5
a <- matrix(rnorm(nrow * 30) < 0.5, nrow = nrow, ncol = 30)
identical(rbind(a, matrix(FALSE, nrow = 1e5, ncol = 30)), icd:::rbind_with_empty(a, nrow))
microbenchmark::microbenchmark(
  rbind(a, matrix(FALSE, nrow = 1e5, ncol = 30)),
  icd:::rbind_with_empty(a, nrow)
)
}
\keyword{internal}
