% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{re_just}
\alias{re_just}
\alias{re_just_ws}
\title{limit a regex to just what is given}
\usage{
re_just(x, whitespace_ok = FALSE)

re_just_ws(x)
}
\arguments{
\item{x}{single string containing a regex}

\item{whitespace_ok}{Single logical, if \code{TRUE}, the default, matches for
ICD codes will accept leading and trailing whitespace.}
}
\description{
update regex to limit by start and end, with or without whitespace
}
\section{Functions}{
\itemize{
\item \code{re_just_ws}: allow whitespace
}}
\keyword{internal}

