% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.R
\name{read.zip.url.lines}
\alias{read.zip.url.lines}
\title{read file from zip at URL}
\usage{
read.zip.url.lines(url, filename, file_enc = "latin1", ...)
}
\arguments{
\item{url}{character vector of length one containing URL of zip file.}

\item{filename}{character vector of length one containing name of file to
extract from zip. If not specified, and the zip contains a single file,
then this single file will be used.}

\item{encoding}{passed to file when the contents of the zip is read, default
is "", i.e. R uses current locale to guess}
}
\description{
downloads zip file, and opens named file \code{filename}, or the
  single file in zip if \code{filename} is not specified. FUN is a function,
  with additional arguments to FUN given by \dots.
}
\keyword{internal}

