% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_nosort.R
\name{refactor}
\alias{refactor}
\title{Refactor by integer matching levels in C++}
\usage{
refactor(x, levels, na.rm = FALSE, exclude_na = TRUE)
}
\description{
Slightly slower for small factors, three times faster for one hundred million
elemeents with two million new levels. Three times faster for any \code{n > 1e6}.
With NAs, margin is smaller, but still beats base \code{factor}.
}
\examples{
\donttest{

}
}
\keyword{internal}
\keyword{manip}
