% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{regexec32}
\alias{regexec32}
\title{regexec which accepts \code{perl} argument even in older R}
\usage{
regexec32(pattern, text, ...)
}
\description{
regexec which accepts \code{perl} argument even in older R
}
\keyword{internal}

