% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-rtf.R
\name{rtf_fetch_year}
\alias{rtf_fetch_year}
\title{Fetch RTF for a given year}
\usage{
rtf_fetch_year(year, offline = TRUE)
}
\arguments{
\item{year}{character vector of length one, e.g. "2011"}

\item{offline}{single logical value}
}
\description{
Will return NULL if offline and not available
}
\keyword{internal}
