% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-rtf.R
\name{rtf_fifth_digit}
\alias{rtf_fifth_digit}
\title{add descriptions and codes with fifth digit}
\usage{
rtf_fifth_digit(out, lookup_fifth)
}
\description{
Use data table to has lookup in existing codes
}
\keyword{internal}

