% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-rtf.R
\name{rtf_filter_excludes}
\alias{rtf_filter_excludes}
\title{exclude some unwanted rows from filtered RTF}
\usage{
rtf_filter_excludes(filtered, ...)
}
\description{
Ignore excluded codes, and fix some odd-balls so they don't get dropped
"707.02Upper back", "066.40West Nile fever, unspecified", etc
}
\keyword{datagen}
\keyword{internal}
