% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-rtf.R
\name{rtf_fix_duplicates}
\alias{rtf_fix_duplicates}
\title{fix duplicates detected in RTF parsing}
\usage{
rtf_fix_duplicates(out, verbose)
}
\description{
clean up duplicates (about 350 in 2015 data), mostly one very brief
description and a correct longer one; or, identical descriptions
}
\keyword{internal}

