% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-rtf.R
\name{rtf_fix_quirks_2015}
\alias{rtf_fix_quirks_2015}
\title{fix quirks for 2015 RTF parsing}
\usage{
rtf_fix_quirks_2015(out)
}
\description{
2015 quirks (many more are baked into the parsing: try to splinter out the
most specific) some may well apply to other years 650-659 ( and probably many
others don't use whole subset of fourth or fifth digit qualifiers) going to
have to parse these, e.g. [0,1,3], as there are so many...
}
\keyword{internal}
\keyword{manip}

