% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-rtf.R
\name{rtf_fix_unicode}
\alias{rtf_fix_unicode}
\title{Fix Unicode characters in RTF}
\usage{
rtf_fix_unicode(filtered, ...)
}
\description{
fix ASCII, Code Page 1252 and Unicode horror: some character definitions are
split over lines... This needs care in Windows, or course. Maybe Mac, too?
}
\details{
First: c cedilla, e grave, e acute Then:  n tilde, o umlaut
}
\keyword{datagen}
\keyword{internal}
