% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-rtf.R
\name{rtf_generate_fourth_lookup}
\alias{rtf_generate_fourth_lookup}
\title{generate look-up for four digit codes}
\usage{
rtf_generate_fourth_lookup(filtered, fourth_rows, verbose = FALSE)
}
\value{
named character vector, names are the ICD codes, values are the
  descriptions
}
\description{
\code{lookup_fourth} will contain vector of suffices, with names being the
codes they augment
}
\keyword{datagen}
\keyword{internal}
