% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-rtf.R
\name{rtf_lookup_fourth}
\alias{rtf_lookup_fourth}
\title{apply fourth digit qualifiers}
\usage{
rtf_lookup_fourth(out, lookup_fourth, verbose = FALSE)
}
\description{
use the lookup table of fourth digit
}
\keyword{datagen}
\keyword{internal}
