% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-rtf.R
\name{rtf_lookup_fourth_alt_base}
\alias{rtf_lookup_fourth_alt_base}
\title{apply fourth digit qualifiers}
\usage{
rtf_lookup_fourth_alt_base(out, lookup_fourth, verbose = FALSE)
}
\description{
use the lookup table of fourth digit
}
\keyword{internal}

