% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-rtf.R
\name{rtf_main_filter}
\alias{rtf_main_filter}
\title{filter RTF for actual ICD-9 codes}
\usage{
rtf_main_filter(filtered, ...)
}
\description{
Keep some more information, but we'll just take the primary description for
each item, i.e. where a code begins a line. Some codes have ten or so
alternative descriptions, e.g. 410.0
}
\keyword{internal}
