% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{sas_extract_let_strings}
\alias{sas_extract_let_strings}
\title{Extract quoted or unquoted SAS string definitions}
\usage{
sas_extract_let_strings(x)
}
\arguments{
\item{x}{is a vector of character strings, typically taken from something
like \code{readLines(some_sas_file_path)}}
}
\description{
Finds all the LET statements in some SAS code and writes them to an R list.
The list item name is the SAS variable name, and each list item is a
character vector of the contents. This is specifically for string
assignments, but probably easy to adapter to numbers if ever needed.
}
\keyword{internal}
\keyword{list}
\keyword{programming}
