% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{sas_format_extract}
\alias{sas_format_extract}
\title{extract assignments from a SAS FORMAT definition.}
\usage{
sas_format_extract(sas_lines)
}
\arguments{
\item{sas_lines}{is a character vector, with one item per line, e.g. from
\code{readLines}}
}
\value{
list (of lists)
}
\description{
this is modelled entirely on a single chunk of SAS code, but
  hopefully will have some generalizability. It relies heavily on lists and
  regex, but, as you will see from the code, R is not a great language with
  which to write a SAS parser.
}
\examples{
  \dontrun{
  sas_format_extract(readLines('data-raw/comformat2012-2013.txt'))
  }
}
\references{
\url{http://support.sas.com/documentation/cdl/en/proc/61895/HTML/default/viewer.htm#a002473474.htm}
\url{https://communities.sas.com/thread/47571?start=0&tstart=0}
\url{https://communities.sas.com/message/165945}
}
\keyword{internal}
\keyword{list}
\keyword{programming}

