% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sas.R
\name{sasDropOtherAssignment}
\alias{sasDropOtherAssignment}
\title{drop superfluous assignment name when the name is already defined.}
\usage{
sasDropOtherAssignment(x)
}
\arguments{
\item{x}{in this case is '079,305 = "YES"' (quotes may be present in the
string itself)}
}
\description{
this is happening in the DRG definitions for AHRQ comorbidities.
  We have data like: "HTNDRG" 079,305 = "YES" I would like this to be
  list(HTNDRG="079,305")
}
\keyword{internal}
\keyword{manip}
\keyword{util}

