% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{sasParseAssignments}
\alias{sasParseAssignments}
\title{get assignments from a character string strings.}
\usage{
sasParseAssignments(x, strip_whitespace = TRUE, strip_quotes = TRUE)
}
\arguments{
\item{x}{is a character string containing space delimited assignments, in SAS
declaration format.}

\item{strip_whitespace}{will strip all whitespace from the returned values}

\item{strip_quotes}{will strip all double quotation marks from the returned
values}
}
\value{
list with each list item containing a matrix of "char ranges",
  "assigned value" pairs
}
\description{
The format of assignments is best seen in the SAS source files.
}
\keyword{internal}
\keyword{list}
\keyword{programming}

