% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_class_icd10cm}
\alias{set_class_icd10cm}
\title{Set class to icd10cm and icd10 in C__}
\usage{
set_class_icd10cm(x)
}
\arguments{
\item{x}{R object}
}
\description{
Assume either of those classes not already set
}
\examples{
j <- "100"
icd:::set_class_icd10cm(j)
stopifnot(is.icd10(j))
stopifnot(is.icd10cm(j))
}
\keyword{internal}

