% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-test.R
\name{set_full_test_options}
\alias{set_full_test_options}
\title{Set test options to do everything}
\usage{
set_full_test_options()
}
\description{
Default without setting options is for slow and online tests to be skipped,
and warnings to be generated for deprecated functions. This function
explicitly sets options to do slow and online tests, and not to warn for
deprecated functions. This is intended for local testing.
}
\keyword{debugging}
\keyword{internal}

