% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd-cm-ver.R
\name{set_icd10cm_active_ver}
\alias{set_icd10cm_active_ver}
\alias{get_icd10cm_active_ver}
\title{Get or set the annual version of ICD-10-CM to use}
\usage{
set_icd10cm_active_ver(ver, check_exists = TRUE)

get_icd10cm_active_ver()
}
\arguments{
\item{ver}{Character version of ICD-10-CM to use, corresponding to the suffix
of the \code{data.frame} name, e.g. for 2019 ICD-10-CM, use
\code{"icd10cm2019"} for Dutch 2014 ICD-CM translations, use
'icd10cm2014_nl"}

\item{check_exists}{\code{TRUE} by default, which forces a check that the
requested version is actually available in this R session.}
}
\description{
Get or set the annual version of ICD-10-CM to use
}
