% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{set_re_globals}
\alias{set_re_globals}
\title{Put ICD validation regexes in the icd::: namespace only}
\usage{
set_re_globals(env = parent.frame())
}
\arguments{
\item{env}{target environment to save the data}
}
\description{
Put ICD validation regexes in the icd::: namespace only
}
\keyword{data}
\keyword{internal}
\keyword{sysdata}
