% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R, R/resource.R
\name{setup_icd_data}
\alias{setup_icd_data}
\alias{icd_data_dir}
\title{Set up the data download cache, give permission to download data}
\usage{
setup_icd_data(path = NULL)

icd_data_dir(path)
}
\arguments{
\item{path}{Path to a directory where cached online raw and parsed data will
be cached. It will be created if it doesn't exist.}
}
\value{
The path to the resource directory, or \code{NULL} if it could not be
  found.

Invisibly returns the data path which was set, or NULL if not done.
}
\description{
This must be called by the user, as prompted on package attach with
\code{library(icd)}.
}
\section{Functions}{
\itemize{
\item \code{icd_data_dir}: Return the currently active data directory. If
missing, it will return \code{NULL} and, depending on
\code{getOption("icd.data.absent_action")}, will stop, give a message, or
do nothing.
}}

\examples{
\dontrun{
setup_icd_data()
setup_icd_data("/var/cache/icd.data")
setup_icd_data(path = ".local/icd.data")
icd_data_dir()
}
}
\seealso{
\code{\link{download_icd_data}}
}
