% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-test.R
\name{setup_test_check}
\alias{setup_test_check}
\title{Set-up test options}
\usage{
setup_test_check()
}
\description{
Checks shell environment for whether to do slow or online tests.
Checks whether we should warn for deprecated functions (usually not
when testing). If \code{covr} is detected (an option is set), then
we may be in a subprocess and not see any shell environment or options
from the calling process, so try to set slow tests on and warnings off.
}

