% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{short_to_parts}
\alias{short_to_parts}
\alias{short_to_parts.icd9}
\alias{short_to_parts.character}
\title{Convert short format ICD codes to component parts}
\usage{
short_to_parts(x, mnr_empty = "")

\method{short_to_parts}{icd9}(x, mnr_empty = "")

\method{short_to_parts}{character}(x, mnr_empty = "")
}
\description{
Convert short format ICD codes to component parts
}
\section{Methods (by class)}{
\itemize{
\item \code{icd9}: Convert short format ICD-9 code to parts

\item \code{character}: Convert short format ICD code to parts,
guessing whether ICD-9 or ICD-10
}}

\keyword{internal}
\keyword{manip}
