% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain-table.R
\name{shortcode_icd9}
\alias{shortcode_icd9}
\alias{shortcode_icd10}
\title{set \code{short_to_decimal} attribute}
\usage{
shortcode_icd9(x, short_code = guess_short(x))

shortcode_icd10(x, short_code = guess_short(x))
}
\description{
Does not convert between decimal and short codes. Calling
\code{short_to_decimal} should convert and set the attribute.
}
\keyword{internal}
