% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-test.R
\name{show_test_options}
\alias{show_test_options}
\title{Show options which control testing}
\usage{
show_test_options()
}
\description{
Get the options for all currently used \code{icd} testing options.
}
\keyword{internal}

