% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sort_std}
\alias{sort_std}
\title{Sort using STL, possibly parallelized}
\usage{
sort_std(x)
}
\description{
Sort using STL, possibly parallelized
}
\examples{
\dontrun{
pts <- icd9:::randomPatients(1e7)
microbenchmark::microbenchmark(sort_std(pts$icd9), sort(pts$icd9), times = 5)
# four times faster on 4 real core (8 with HT) machine.
}
}

