% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/comorbid.R
\name{spawnReferenceChildren}
\alias{spawnReferenceChildren}
\title{spawn reference codes into all possible lower-level codes (and memoise)}
\usage{
spawnReferenceChildren(icd9Reference, isShortReference)
}
\description{
Take a regular string of an ICD9 code of format (ABC.zxyz) with or without
leading and trailing zeroes. top level ICD9 code and return T/F if the icd9
fall within subgroups. This takes several seconds on an unimpressive desktop
PC, so would benefit from memoization.
}
\keyword{internal}

