% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{str_extract}
\alias{str_extract}
\title{\code{stringr} does this, but here we have a small amount of base R code}
\usage{
str_extract(string, pattern, ...)
}
\description{
\code{stringr} does this, but here we have a small amount of base R code
}
\keyword{internal}
