% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{strim}
\alias{strim}
\title{Trim leading and trailing whitespace from a single string}
\usage{
strim(x)
}
\arguments{
\item{x}{character vector of length one}
}
\value{
character vector of length one
}
\description{
\code{NA} is accepted (and returned as \code{NA_character_})
}
\keyword{internal}

