% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-rtf.R
\name{strip_rtf}
\alias{strip_rtf}
\title{Strip RTF}
\usage{
strip_rtf(x)
}
\arguments{
\item{x}{vector of character strings containing RTF}
}
\description{
Take a vector of character strings containing RTF, replace each \\tab with a
space and eradicate all other RTF symbols
}
\keyword{internal}

