% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-rtf.R
\name{stripRtf}
\alias{stripRtf}
\title{Strip RTF}
\usage{
stripRtf(x)
}
\arguments{
\item{x}{vector of character strings containing RTF}
}
\description{
Take a vector of character strings containing RTF, replace each \tab with a
space and eradicate all other RTF symbols
}
\keyword{internal}

