% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{strMultiMatch}
\alias{strMultiMatch}
\alias{strPairMatch}
\title{return the actual matches from a bracketed regex}
\usage{
strMultiMatch(pattern, text, dropEmpty = FALSE, ...)

strPairMatch(pattern, text, swap = FALSE, dropEmpty = FALSE, pos = c(1,
  2), ...)
}
\arguments{
\item{pattern}{regular expression: if it has bracketed sections, these
submatches are returned}

\item{text}{is the string to match against. This vector should be the same
length as the pattern vector, or the patern vector should be length one.}

\item{dropEmpty}{logical whether to drop rows with no matches}

\item{...}{are additional parameters passed to regexec and regmatches. I
haven't tried this: it may need two separate variables containing lists of
params, since this will send everything to both functions.}

\item{swap}{logical scalar, whether to swap the names and values. Default is
not to swap, so the first match becomes the name.}
}
\value{
list of character vectors, list length being the length of the inptu
  text vector.
}
\description{
Be careful: this may throw funny results for exotic regex, but
  so far, it seems okay. it also drops the first result which always seems to
  be a duplicate or whole-string match.

\code{strPaitMatch} differs in that there should only be two
  pairs of parenthesis, then the first (by default) becomes the name, and the
  second the value.
}
\keyword{internal}

