% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{[.icd_map}
\alias{[.icd_map}
\title{extract elements of an ICD comorbidity map}
\usage{
\method{[}{icd_map}(x, index, ...)
}
\arguments{
\item{x}{\code{icd_map} object}

\item{index}{index element to index}

\item{...}{arguments to pass to \code{list.[}}
}
\description{
Equivalent to a list, but preserves class of extracted elements
}

