% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{[[.comorbidity_map}
\alias{[[.comorbidity_map}
\title{Extract vector of codes from an ICD comorbidity map}
\usage{
\method{[[}{comorbidity_map}(x, index, ...)
}
\arguments{
\item{x}{comorbidity map, which is a named list}

\item{index}{integer}

\item{...}{arguments passed on to other functions}
}
\description{
Equivalent to a list, but preserves class of extracted vector.
}
\examples{
# show that attributes are preserved when subsetting
stopifnot(is.short_diag(icd10_map_ahrq[[1]]))
}
\keyword{internal}
