% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{subset_icd}
\alias{[.icd10}
\alias{[.icd9}
\alias{[[.icd10}
\alias{[[.icd9}
\alias{subset_icd}
\title{extract subset from ICD data}
\usage{
\method{[}{icd9}(x, ...)

\method{[[}{icd9}(x, ...)

\method{[}{icd10}(x, ..., drop = TRUE)

\method{[[}{icd10}(x, ..., exact = TRUE)
}
\arguments{
\item{x}{object to extract from}

\item{...}{other arguments to pass to next S3 methods}
}
\description{
exactly the same as using x[n] or x[[n]] but preserves the ICD
  classes in result
}
\details{
TODO:Potential here to use attributes, since we can (as base R does
  in \code{datediff}, \code{POSIXct}, etc.) recreate the attributes after a
  subsetting operation. This would simplify the class system.
}

