% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{swap_names_vals}
\alias{swap_names_vals}
\title{swap names and values of a vector}
\usage{
swap_names_vals(x)
}
\arguments{
\item{x}{named vector}
}
\value{
vector, with values being the names of the input vector, and names
being the previous values.
}
\description{
Swap names and values of a vector. Non-character values are implicitly
converted to names.
}
\keyword{internal}

