% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.R
\name{swapNamesWithVals}
\alias{swapNamesWithVals}
\title{swap names and values of a vector}
\usage{
swapNamesWithVals(x)
}
\arguments{
\item{x}{named vector}
}
\value{
vector
}
\description{
swap names and values of a vector
}
\keyword{internal}

