% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{trim}
\alias{trim}
\title{Trim leading and trailing white space}
\usage{
trim(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector
}
\description{
\code{NA} is accepted and returned, probably as \code{NA_character_}
}
\examples{
\dontrun{
s <- random_string(250);
microbenchmark::microbenchmark(trim(s), trimws(s)) # trimws from R version 3.2
}
}
\keyword{internal}
