% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-file.R
\name{unzip_single}
\alias{unzip_single}
\title{unzip a single file from URL}
\usage{
unzip_single(url, file_name, save_path)
}
\description{
take a single file from zip located at a given URL, unzip into temporary
directory, and copy to the given \code{save_path}
}
\keyword{internal}

