% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{update_class_from_columns}
\alias{update_class_from_columns}
\title{If only one ICD code type is defined in columns in a matrix or data.frame,
then update the class of the data frame to relfect that type.}
\usage{
update_class_from_columns(x)
}
\description{
If only one ICD code type is defined in columns in a matrix or data.frame,
then update the class of the data frame to relfect that type.
}
\keyword{internal}

