% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{update_data_frame_class}
\alias{update_data_frame_class}
\title{Update the ICD version class of data.frame if there is an unambiguous version
column therein}
\usage{
update_data_frame_class(x, icd_name = get_icd_name(x), must_work = TRUE)
}
\arguments{
\item{x}{\code{data.frame} of patient data}

\item{icd_name}{The column in the data frame which contains the ICD codes.
This is a character vector of length one. If it is \code{NULL}, \code{icd9}
will attempt to guess the column name, looking for progressively less
likely possibilities until it matche a single column. Failing this, it will
take the first column in the data frame. Specifying the column using this
argument avoids the guesswork.}

\item{must_work}{single logical, if \code{TRUE} (the default) will stop if a
single ICD version class cannot be applied}
}
\description{
If only one ICD code type is defined in columns in a matrix or data.frame,
then update the class of the data frame to relfect that type.
}
\keyword{internal}

