% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{update_everything}
\alias{update_everything}
\title{generate all package data}
\usage{
update_everything()
}
\description{
Parses (and downloads if necessary) CDC annual revisions of ICD-9-CM to get
the 'billable' codes. Also parses the AHRQ, Quan/Deyo, and CMS HCC comorbidity mappings
from the source SAS data. Elixhauser and Quan/Elixhauser mappings are
generated from transcribed codes.
}
\keyword{internal}

