% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{vec_to_env}
\alias{vec_to_env}
\title{create environment from vector}
\usage{
vec_to_env(x, val = TRUE, env = new.env(hash = TRUE, parent = baseenv()))
}
\description{
create an environment by inserting the value \code{val} with names taken from
\code{x}
}
\keyword{internal}

