% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{ahrqComorbid}
\alias{ahrqComorbid}
\title{AHRQ comorbidities}
\format{list of character vectors}
\source{
\url{http://www.hcup-us.ahrq.gov/toolssoftware/comorbidity/comorbidity.jsp}
}
\description{
This mapping of comorbidities to ICD-9 codes is dervied directly
  from SAS code provided by AHRQ, and translated into this R data structure.
  This is a revision of the Elixhauser system, notably exluding cradiac
  arrythmia.
}
\keyword{datasets}

