% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{as.comorbidity_map}
\alias{as.comorbidity_map}
\title{Set the class of a named list to show it is a comorbidity map.}
\usage{
as.comorbidity_map(x)
}
\description{
Set the class of a named list to show it is a comorbidity map.
}
\seealso{
Other ICD data types: \code{\link{set_icd_class}},
  \code{\link{wide_vs_long}}
}
\concept{ICD data types}
