% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/attr.R
\name{.attr_short_diag}
\alias{.attr_short_diag}
\alias{as.icd_short_diag}
\alias{as.icd_decimal_diag}
\title{Change whether ICD code has short or long attribute}
\usage{
.attr_short_diag(x, value = TRUE)

as.icd_short_diag(x, value = TRUE)

as.icd_decimal_diag(x, value = TRUE)
}
\arguments{
\item{x}{ICD data}

\item{value}{True or false, default is \code{TRUE}}
}
\description{
The attribute \code{icd_short_code} should be either \code{TRUE} or
\code{FALSE}.
}
\examples{
codes <- as.icd9(c("100.1", "441.3"))
codes <- as.icd_decimal_diag(codes)
codes
}
\keyword{attribute}
\keyword{internal}
