% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{as_char_no_warn}
\alias{as_char_no_warn}
\title{convert to character vector without warning}
\usage{
as_char_no_warn(x)
}
\arguments{
\item{x}{vector, typically numeric or a factor}
}
\description{
convert to character vector without warning
}
\keyword{internal}

