% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{chapter_to_desc_range}
\alias{.chapter_to_desc_range}
\alias{chapter_to_desc_range}
\alias{chapter_to_desc_range.icd10}
\alias{chapter_to_desc_range.icd9}
\title{Parse a (sub)chapter text description with parenthesised range}
\usage{
.chapter_to_desc_range(x, re_major)

chapter_to_desc_range.icd9(x)

chapter_to_desc_range.icd10(x)
}
\arguments{
\item{x}{vector of descriptions followed by ICD code ranges}
}
\value{
list of two-element character vectors, the elements being named
  'start' and 'end'.
}
\description{
Parse a (sub)chapter text description with parenthesised range
}
\keyword{internal}
\keyword{manip}

