% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{combine}
\alias{c.icd10}
\alias{c.icd10cm}
\alias{c.icd10who}
\alias{c.icd9}
\alias{c.icd9cm}
\alias{combine}
\title{combine ICD codes}
\usage{
\method{c}{icd9}(...)

\method{c}{icd9cm}(...)

\method{c}{icd10}(...)

\method{c}{icd10cm}(...)

\method{c}{icd10who}(...)
}
\description{
These function implement combination of lists or vectors of
  codes, while preserving ICD classes.
}

