% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R, R/RcppExportsMjrMnr.R, R/convert.R
\name{icd9_MajMinToCode}
\alias{convert}
\alias{icd9DecimalToParts}
\alias{icd9DecimalToShort}
\alias{icd9GetMajor}
\alias{icd9MajMinToCode}
\alias{icd9MajMinToDecimal}
\alias{icd9MajMinToParts}
\alias{icd9MajMinToShort}
\alias{icd9PartsToDecimal}
\alias{icd9PartsToShort}
\alias{icd9ShortToDecimal}
\alias{icd9ShortToParts}
\alias{icd9_MajMinToCode}
\alias{icd9_MajMinToDecimal}
\alias{icd9_MajMinToParts}
\alias{icd9_MajMinToShort}
\title{Convert ICD9 codes between formats and structures.}
\usage{
icd9_MajMinToCode(mjr, mnr, isShort)

icd9_MajMinToShort(mjr, mnr)

icd9_MajMinToDecimal(mjr, mnr)

icd9PartsToShort(parts)

icd9PartsToDecimal(parts)

icd9_MajMinToParts(mjr, mnr)

icd9ShortToParts(icd9Short, minorEmpty = "")

icd9DecimalToParts(icd9Decimal, minorEmpty = "")

icd9ShortToDecimal(x)

icd9DecimalToShort(x)

icd9GetMajor(icd9, isShort)

icd9MajMinToCode(major, minor, isShort)

icd9MajMinToShort(major, minor)

icd9MajMinToDecimal(major, minor)

icd9MajMinToParts(major, minor)
}
\arguments{
\item{isShort}{single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form.}

\item{parts}{data.frame with major and minor fields. This can be given
instead of major and minor vectors}

\item{minorEmpty}{vector of length one, to be used in place of
minor part of zero. Defaults to ""}

\item{major}{}

\item{minor}{}

\item{icd9-any}{}

\item{icd9-short}{}

\item{icd9-long}{}
}
\value{
Character vector or list. Deliberately returns zero-padded major, because
  otherwise we are creating ambiguous codes (even if we know what we mean)
}
\description{
ICD-9 codes are represented in \emph{short} and \emph{decimal} forms.
The short form has up to 5 digits, or V or E followed by up to four digits. The decimal form
has a dedcimal point to delimit the top-level (henceforth \emph{major}) category, and the
\emph{minor} part containing the subsidiary classifications.
}
\section{Structures used in this package are}{

\itemize{
  \item vector (usually character vectors) of \emph{short} or \emph{long} codes
  \item list, containing two elements with equal length vectors called \code{major}
  \code{minor}. Yes, this is very like a data frame, but we do not need the overhead
  of structuring it that way.
  \item two vectors of separated major and minor parts
  }
}
\seealso{
Other ICD-9 convert: \code{\link{icd9ChaptersToMap}};
  \code{\link{icd9DropLeadingZeroes}},
  \code{\link{icd9DropLeadingZeroesDecimal}},
  \code{\link{icd9DropLeadingZeroesMajor}},
  \code{\link{icd9DropLeadingZeroesShort}};
  \code{\link{icd9LongToWide}};
  \code{\link{icd9WideToLong}}
}

