% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert ICD9 codes between formats and structures.}
\description{
ICD-9 codes are represented in \emph{short} and \emph{decimal} forms. The
short form has up to 5 digits, or V or E followed by up to four digits. The
decimal form has a decimal point to delimit the top-level (henceforth
\emph{major}) category, and the \emph{minor} part containing the subsidiary
classifications.
}
\details{
For conversions of ICD-9 or ICD-10 codes between the \emph{short} and
\emph{decimal} forms, use \code{\link{short_to_decimal}} and
\code{\link{decimal_to_short}}.

\pkg{icd} does not covert ICD-9 to ICD-10 codes yet.
}
\seealso{
Other ICD code conversion: \code{\link{as.icd_long_data}},
  \code{\link{long_to_wide}}, \code{\link{wide_to_long}}
}
\concept{ICD code conversion}
