% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbid.R
\name{cr}
\alias{cr}
\title{sequence columns of comorbidities}
\usage{
cr(x)
}
\arguments{
\item{x}{matrix or data.frame of comorbidities}
}
\description{
Get sequence of column indices of comorbidity data frame, which differ
between tidy data frames and clean, row-named matrices
}
\keyword{internal}
