% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbid.R
\name{cr}
\alias{cr}
\title{sequence columns of comorbidities}
\usage{
cr(x)
}
\arguments{
\item{cbd}{matrix or data.frame of comorbidities}
}
\description{
Get sequence of column indices of comorbidity data frame
}
\keyword{internal}
