% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{expand_range_worker}
\alias{expand_range_worker}
\title{expand range worker function}
\usage{
expand_range_worker(start, end, lookup, defined, ex_ambig_start, ex_ambig_end)
}
\description{
expand range worker function
}
\keyword{internal}

