% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-test.R
\name{expect_equal_no_icd}
\alias{expect_equal_no_icd}
\title{expect equal, ignoring any ICD classes}
\usage{
expect_equal_no_icd(object, expected, ..., info = NULL, label = NULL,
  expected.label = NULL)
}
\description{
expect equal, ignoring any ICD classes
}
\keyword{internal}

