% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{factor_}
\alias{factor_}
\title{Fast Factor Generation}
\usage{
factor_(x, levels = NULL, labels = levels, na.last = NA)
}
\arguments{
\item{x}{An object of atomic type \code{integer}, \code{numeric}, \code{character} or \code{logical}.}

\item{levels}{An optional character vector of levels. Is coerced to the same type as \code{x}. By default, we
compute the levels as \code{sort(unique.default(x))}.}

\item{labels}{A set of labels used to rename the levels, if desired.}

\item{na.last}{If \code{TRUE} and there are missing values, the last level is set as \code{NA}; otherwise; they are
removed.}
}
\description{
This function generates factors more quickly, by leveraging \code{fastmatch::\link{fmatch}}. The speed increase for
ICD-9 codes is about 33% reduction for 10 million codes.
}
\details{
\code{NaN}s are converted to \code{NA} when used on numerics. Extracted from https://github.com/kevinushey/Kmisc.git
}
\examples{
\dontrun{
pts <- icd9:::randomUnorderedPatients(1e7)
u <- unique.default(pts$icd9)
# this shows that stringr (which uses stringi) sort takes 50\% longer than built-in R sort.
microbenchmark::microbenchmark(sort(u), stringr::str_sort(u))

# this shows that \\code{factor_} is about 50\% faster than \\code{factor} for big vectors of strings
microbenchmark::microbenchmark(factor(pts$icd9), factor_(pts$icd9), times = 10)
}
}
\author{
Kevin Ushey, adapted by Jack Wasey
}
\keyword{internal}

