% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastIntToString}
\alias{fastIntToString}
\alias{fastIntToStringRcpp}
\title{fast convert integer vector to character vector}
\usage{
fastIntToString(x, bufferSize = 16L)

fastIntToStringRcpp(x, bufferSize = 16L)
}
\arguments{
\item{x}{td::vector<int>}

\item{bufferSize}{int if any input strings are longer than this number (default 16) there will be memory errors.
No checks done for speed.}
}
\description{
fast convert integer vector to character vector
}
\examples{
\dontrun{
# this is actually slower than as.character
microbenchmark::microbenchmark(fastIntToString(pts$visitId, buffer = 8),
                               fastIntToString(pts$visitId, buffer = 16),
                               fastIntToString(pts$visitId, buffer = 64),
                               fastIntToStringRcpp(pts$visitId, buffer = 8),
                               fastIntToStringRcpp(pts$visitId, buffer = 16),
                               fastIntToStringRcpp(pts$visitId, buffer = 64),
                               as.character(pts$visitId),
                               asCharacterNoWarn(pts$visitId), times = 5)
}
}
\keyword{internal}

