% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastIntToString}
\alias{fastIntToString}
\alias{fastIntToString_alt_Std}
\alias{fastIntToStringRcpp}
\title{Fast convert integer vector to character vector}
\usage{
fastIntToString_alt_Std(x)

fastIntToStringRcpp(x)
}
\arguments{
\item{x}{vector of integers}

\item{bufferSize}{int if any input strings are longer than this number
(default 16) there will be memory errors. No checks done for speed.}
}
\description{
Fast conversion from integer vector to character vector using C++
}
\section{Functions}{
\itemize{
\item \code{fastIntToStringRcpp}: Same using \code{Rcpp}
}}

\keyword{internal}
