% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-test.R
\name{generate_random_short_icd10cm_bill}
\alias{generate_random_short_icd10cm_bill}
\title{generate random ICD-9 codes}
\usage{
generate_random_short_icd10cm_bill(n = 10)
}
\arguments{
\item{n}{number to select, passed to \code{sample}}
}
\description{
Uses billable ICD-10-CM codes from current master list
}
\keyword{datagen}
\keyword{debugging}
\keyword{internal}

