% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-test.R
\name{generate_random_short_icd9}
\alias{generate_random_short_icd9}
\alias{generate_random_decimal_icd9}
\alias{generate_random_pts}
\alias{generate_random_ordered_pts}
\alias{generate_random_unordered_pts}
\alias{generate_random_short_ahrq_icd9}
\title{generate random ICD-9 codes}
\usage{
generate_random_short_icd9(n = 50000)

generate_random_decimal_icd9(n = 50000)

generate_random_pts(...)

generate_random_ordered_pts(...)

generate_random_unordered_pts(num_patients = 50000,
  dz_per_patient = 20, n = num_patients, np = dz_per_patient,
  fun = generate_random_short_icd9)

generate_random_short_ahrq_icd9(n = 50000)
}
\description{
generate random ICD-9 codes
}
\keyword{datagen}
\keyword{debugging}
\keyword{internal}
