% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sysdata.R
\name{generate_sysdata}
\alias{generate_sysdata}
\title{Generate \code{sysdata.rda}}
\usage{
generate_sysdata(save_data = TRUE)
}
\arguments{
\item{save_data}{single logical value, if \code{TRUE} will save the generated
data}
}
\value{
invisibly returns the data as a list
}
\description{
Generate correctly ordered look-up tables of numeric-only, V and E codes.
This is quick, but much too slow when it appears many times in a loop.
}
\keyword{internal}

