% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mappings.R
\name{generate_uranium_pathology}
\alias{generate_uranium_pathology}
\title{generate uranium pathology data}
\usage{
generate_uranium_pathology(save_data = FALSE)
}
\description{
generate uranium pathology data
}
\keyword{internal}

