% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.R, R/z-deprecated.R
\name{get_defined}
\alias{get_defined}
\alias{icd_get_defined.icd9}
\alias{icd_get_defined}
\title{Select only defined ICD codes}
\usage{
get_defined(x, short_code = guess_short(x), billable = FALSE)

icd_get_defined.icd9(...)

icd_get_defined(...)
}
\arguments{
\item{x}{input vector or factor, possibly with an ICD class}

\item{short_code}{logical value, whether short-form ICD code}

\item{billable}{single logical value, whether to limit return codes also by
whether they are billable, i.e. leaf nodes. This is really only designed
for use with ICD-9-CM, ICD-10-CM etc, since the WHO versions are not
designed for billing, but for public health and death reporting.}

\item{...}{arguments passed on to other functions}
}
\description{
Return only those codes which are heading or leaf (billable), specifying
whether codes are all short-form or all decimal-form
}
\section{Deprecated function names}{
 Future versions of \pkg{icd} will drop
  the \code{icd_} prefix. For example, \code{charlson} should be used in
  favor of \code{icd_charlson}. To distinguish \pkg{icd} function calls,
  consider using the prefix \code{icd::} instead, e.g., \code{icd::charlson}.
  Functions which specifically operate on either ICD-9 or ICD-10 codes or
  their sub-types will retain the prefix. E.g.
  \code{\link{icd9_comorbid_ahrq}}. \pkg{icd} specific classes also retain
  the prefix, e.g., \code{\link{icd_wide_data}}.
}

