% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd-cm-ver.R
\name{get_icd10cm_version}
\alias{get_icd10cm_version}
\alias{get_icd10cm_active}
\title{Get the data for a given version (four-digit year) of ICD-10-CM}
\usage{
get_icd10cm_version(ver, verbose = .verbose())

get_icd10cm_active(verbose = FALSE)
}
\arguments{
\item{ver}{Character version of ICD-10-CM to use, corresponding to the suffix
of the \code{data.frame} name, e.g. for 2019 ICD-10-CM, use
\code{"icd10cm2019"} for Dutch 2014 ICD-CM translations, use
'icd10cm2014_nl"}

\item{verbose}{single logical value, defaults to \code{FALSE} in most functions.}
}
\description{
When called without an argument, it returns the currently active version as
set by \code{set_icd10cm_active_ver()}
}
\section{Functions}{
\itemize{
\item \code{get_icd10cm_active}: Get the currently active version of
ICD-10-CM.
}}

\examples{
\dontrun{
get_icd10cm_version("2018")
}
}
