% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_non_ASCII}
\alias{get_non_ASCII}
\alias{is_non_ASCII}
\alias{get_encodings}
\title{mimic the \code{R CMD check} test}
\usage{
get_non_ASCII(x)

is_non_ASCII(x)

get_encodings(x)
}
\description{
\code{R CMD check} is quick to tell you where \code{UTF-8} characters are not
encoded, but gives no way of finding out which or where
}
\examples{
\dontrun{
sapply(icd9cm_hierarchy, icd:::get_non_ASCII)
icd:::get_encodings(icd9cm_hierarchy)
sapply(icd9cm_billable, icd:::get_non_ASCII)
sapply(icd9cm_billable, icd:::get_encodings)
}
}
\keyword{internal}
