% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{\%fin\%}
\alias{\%fin\%}
\alias{\%fnin\%}
\title{Fast find which \code{x} are \emph{not} in \code{table}}
\usage{
x \%fin\% table

x \%fnin\% table
}
\arguments{
\item{x}{vector}

\item{table}{vector}
}
\description{
Uses \code{fmatch} taken from the \code{fastmatch} package which creates hash
table, and re-uses if \code{table} is re-used.
}
\keyword{internal}

