% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{guess_icd_col_by_name}
\alias{guess_icd_col_by_name}
\alias{guess_icd_col_by_class}
\title{get candidate column(s) from wide or long data frame frame, using hints}
\usage{
guess_icd_col_by_name(x, valid_codes = TRUE, defined_codes = FALSE,
  guesses = c("icd.?(9|10)", "icd.?(9|10).?Code", "icd", "diagnos",
  "diag.?code", "diag", "dx", "i(9|10)", "code"),
  class_pattern = icd_dx_not_generic)

guess_icd_col_by_class(x, pattern)
}
\value{
Zero, one or many names of columns likely to contain ICD codes based
  on the column names.
}
\description{
get candidate column(s) from wide or long data frame frame, using hints
}
\section{Functions}{
\itemize{
\item \code{guess_icd_col_by_class}: Just use the class of columns
}}

\examples{
wide_df <- data.frame(a = letters,
                      dx0 = icd9_map_elix$CHF[1:26],
                      dx1 = icd9_map_elix$PVD[1:26],
                      dx2 = icd9_map_elix$HTN[1:26])
icd:::guess_icd_col_by_name(wide_df)
wide_dc <- data.frame(a = letters,
                      dx0 = as.icd9cm(icd9_map_elix$CHF[1:26]),
                      dx1 = as.icd9cm(icd9_map_elix$PVD[1:26]),
                      dx2 = as.icd9cm(icd9_map_elix$HTN[1:26]),
                      stringsAsFactors = FALSE)
icd:::guess_icd_col_by_name(wide_dc)
}
\keyword{internal}
