% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{icd10_chapters}
\alias{icd10_chapters}
\title{ICD-10 chapters}
\description{
The WHO ICD-10 scheme chapters. The chapter level is the highest in the
hierarchy, each chapter containing sets of codes which span multiple
three-digit 'major' codes, and in some cases also span codes across two
alphabetic initial characters. E.g. Chapter I spans A00 to B99.
}
\details{
2017 ICD-10-CM does not have any U codes (codes for special purposes).
U00-U49 - Provisional assignment of new diseases of uncertain etiology or
emergency use U82-U85 - Resistance to antimicrobial and anti-neoplastic drugs

Format: list with chapter names stored in list names, each with two
  element named character vector with start and end codes.
}
\seealso{
\code{\link[icd.data]{ls_icd_data}}
}
\keyword{datasets}
