% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbid_alt.R
\name{icd10_comorbid_parent_search}
\alias{icd10_comorbid_parent_search}
\title{find ICD-10 comorbidities by checking parents}
\usage{
icd10_comorbid_parent_search(x, map, visit_name = NULL,
  icd_name = get_icd_name(x), short_code = guess_short(x, icd_name =
  icd_name), short_map = guess_short(map), return_df = FALSE, ...)
}
\description{
find ICD-10 comorbidities by checking parents
}
\examples{
up <- uranium_pathology[1:50, ]
if (exists("icd10_comorbid_parent_search_orig"))
  stopifnot(identical(
  icd:::icd10_comorbid_parent_search_orig(up, icd10_map_ahrq,
    visit_name = "case", icd_name ="icd10",
    short_code = FALSE, short_map = TRUE, return_df = FALSE),
  icd:::icd10_comorbid_parent_search_use_cpp(up, icd10_map_ahrq,
    visit_name = "case", icd_name ="icd10",
    short_code = FALSE, short_map = TRUE, return_df = FALSE),
  icd:::icd10_comorbid_reduce(up, icd10_map_ahrq,
    visit_name = "case", icd_name ="icd10",
    short_code = FALSE, short_map = TRUE, return_df = FALSE)
  ))
}
\keyword{internal}
