% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd10_comorbid_parent_search_cpp}
\alias{icd10_comorbid_parent_search_cpp}
\title{Internal function to find ICD-10 parents}
\usage{
icd10_comorbid_parent_search_cpp(x, map, visit_name, icd_name)
}
\arguments{
\item{x}{Character vector (not factor)}
}
\description{
Written in C++ for speed. Takes short code, short map, and pre-defined
visit_name and icd_name.
}
\seealso{
\url{https://github.com/s-u/fastmatch/blob/master/src/fastmatch.c}
}
\keyword{internal}

