% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-ahrq-ccs.R
\name{icd10_fetch_ahrq_ccs}
\alias{icd10_fetch_ahrq_ccs}
\title{Download AHRQ CCS ICD-10 definitions}
\usage{
icd10_fetch_ahrq_ccs(version = "2018.1", offline)
}
\arguments{
\item{offline}{single logical, if \code{TRUE} then don't pull the file from
internet, only return path and file name if the file already exists in the
raw data directory. This is helpful for testing without using the internet.}
}
\description{
Download AHRQ CCS ICD-10 definitions
}
\keyword{internal}
