% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mappings.R
\name{icd10_generate_map_quan_elix}
\alias{icd10_generate_map_quan_elix}
\title{generate ICD-10 Quan/Elixhauser maping}
\usage{
icd10_generate_map_quan_elix(condense = NULL, save = TRUE)
}
\description{
generate ICD-10 Quan/Elixhauser maping
}
\details{
Started with Quan's SAS code (in \code{data-raw}):
}

