% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{icd10_pcs}
\alias{icd10_pcs}
\title{ICD-10-CM Procedure Codes}
\description{
ICD-10-PCS is the annually-updated set of procedure codes designed by 3M for
the US CMS. There is no directory of WHO ICD procedure codes.
}
\details{
Format:  A named list of data frames. The elements of the list are named by
  the year, e.g., \code{"2018"}. Each data frame contains two character
  columns, the first, named \code{code} is the procedure code; the second,
  named \code{desc}, has the description.
}
\seealso{
\code{\link[icd.data]{ls_icd_data}}
}
\keyword{datasets}
