% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-icd10cm-xml.R
\name{icd10cm_extract_sub_chapters}
\alias{icd10cm_extract_sub_chapters}
\title{Get sub-chapters from the 2016 XML for ICD-10-CM}
\usage{
icd10cm_extract_sub_chapters(save_data = FALSE)
}
\arguments{
\item{save_data}{single logical value, if \code{TRUE} will save the generated
data}
}
\description{
This is not quite a super-set of ICD-10 sub-chapters: many more codes than
ICD-10, but some are abbreviated, notably HIV.
}
\details{
This is complicated by the XML document specifying more hierarchical
levels,e.g. \code{C00-C96}, \code{C00-75} are both specified within the
chapter Neoplasms (\code{C00-D49}). A way of determining whether there are
extra levels would be to check the XML tree depth for a member of each
putative sub-chapter.
}
\keyword{internal}
