% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-ahrq-ccs.R
\name{icd9_fetch_ahrq_ccs}
\alias{icd9_fetch_ahrq_ccs}
\title{Download AHRQ CCS ICD-9 definitions}
\usage{
icd9_fetch_ahrq_ccs(single = TRUE, ...)
}
\arguments{
\item{single}{a logical value, if \code{TRUE} then use single level CCS,
otherwise use multi level}

\item{...}{arguments passed on to other functions}
}
\description{
Download AHRQ CCS ICD-9 definitions
}
\keyword{internal}
