% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/valid.R
\name{icd9IsN}
\alias{icd9IsE}
\alias{icd9IsN}
\alias{icd9IsV}
\alias{icd9_is_e}
\alias{icd9_is_n}
\alias{icd9_is_v}
\title{do ICD-9 codes belong to numeric, V or E classes?}
\usage{
icd9IsN(icd9)

icd9IsV(icd9)

icd9IsE(icd9)

icd9_is_n(x)

icd9_is_v(x)

icd9_is_e(x)
}
\arguments{
\item{icd9}{is
a character vector or factor of ICD-9 codes. If fewer than five characters
is given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010".}
}
\value{
logical vector
}
\description{
For each code, return \code{TRUE} if numric or \code{FALSE} if a
  V or E code.
}
\section{Functions}{
\itemize{
\item \code{icd9_is_v}: are the given codes V type?

\item \code{icd9_is_e}: are the given codes E type?
}}

