% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{icd9_map_ahrq}
\alias{@details}
\alias{Beyond}
\alias{CHF}
\alias{HTN,}
\alias{This}
\alias{\code{icd9_ahrq_all}}
\alias{\code{icd9_ahrq}}
\alias{\code{icd9_ahrq},}
\alias{ahrq}
\alias{ahrqComorbid}
\alias{ahrqComorbidAll}
\alias{all}
\alias{and}
\alias{categories.}
\alias{icd9_map_ahrq}
\alias{icd9_map_ahrq_all}
\alias{in}
\alias{includes}
\alias{into}
\alias{is}
\alias{needed}
\alias{not}
\alias{parent}
\alias{probably}
\alias{renal}
\alias{resolution}
\alias{rolled}
\alias{subgroups,}
\alias{the}
\alias{their}
\alias{typical}
\alias{usage:}
\alias{want.}
\alias{what}
\alias{you}
\title{AHRQ comorbidities}
\format{list of character vectors}
\source{
\url{http://www.hcup-us.ahrq.gov/toolssoftware/comorbidity/comorbidity.jsp}
}
\description{
This mapping of comorbidities to ICD-9 codes is dervied directly
  from SAS code provided by AHRQ, and translated into this R data structure.
  This is a revision of the Elixhauser system, notably exluding cardiac
  arrythmia.
}
\keyword{datasets}

