% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{icd9_map_ahrq}
\alias{ahrq}
\alias{icd10_map_ahrq}
\alias{icd9_map_ahrq}
\title{AHRQ comorbidities}
\format{list of character vectors}
\source{
\url{http://www.hcup-us.ahrq.gov/toolssoftware/comorbidity/comorbidity.jsp}
\url{http://www.hcup-us.ahrq.gov/toolssoftware/comorbidityicd10/comorbidity_icd10.jsp}
}
\description{
This mapping of comorbidities to ICD-9 codes is derived directly from SAS
code provided by AHRQ, and translated into this R data structure. This is a
revision of the Elixhauser system, notably excluding cardiac arrythmia.
}
\keyword{datasets}

