% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{icd9_map_pccc}
\alias{icd9_map_pccc}
\alias{icd9_map_pccc_dx}
\alias{icd9_map_pccc_pc}
\alias{icd10_map_pccc_dx}
\alias{icd10_map_pccc_pc}
\title{Pediatric Complex Chronic Conditions}
\description{
There are seven comorbidity maps which represent the combinations of ICD-9,
ICD-10, diagnosis and procedure codes, and three also with a set of 'fixed'
codes. (See reference).
}
\references{
Feudtner C, et al. Pediatric complex chronic conditions
  classification system version 2: updated for ICD-10 and complex medical
  technology dependence and transplantation, BMC Pediatrics, 2014, 14:199,
  DOI: 10.1186/1471-2431-14-199

Feudtner C, Feinstein JA, Zhong W, Hall M, Dai D. Pediatric
  complex chronicconditions classification system version 2: updated for
  ICD-10 and complexmedical technology dependence and transplantation. BMC
  Pediatr. 2014 Aug8;14:199. doi: 10.1186/1471-2431-14-199.
  \url{https://www.ncbi.nlm.nih.gov/pubmed/25102958}
}
\seealso{
\url{https://feudtnerlab.research.chop.edu/ccc_version_2.php}
}
\keyword{datasets}
