% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R, R/manip.R
\name{icd9AddLeadingZeroesMajor}
\alias{icd9AddLeadingZeroes}
\alias{icd9AddLeadingZeroesDecimal_R}
\alias{icd9AddLeadingZeroesMajor}
\alias{icd9AddLeadingZeroesMajor_R}
\alias{icd9AddLeadingZeroesShort}
\alias{icd9AddLeadingZeroesShort_R}
\title{pad decimal ICD-9 codes with leading zeroes}
\usage{
icd9AddLeadingZeroesMajor(mjr)

icd9AddLeadingZeroesShort(icd9Short)

icd9AddLeadingZeroes(icd9, isShort)

icd9AddLeadingZeroesDecimal_R(icd9Decimal)

icd9AddLeadingZeroesShort_R(icd9Short)

icd9AddLeadingZeroesMajor_R(major)
}
\arguments{
\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{icd9}{is a character vector of ICD-9 codes. If fewer than five characters is
given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010"}

\item{isShort}{single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form.}

\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}
}
\value{
character vector of icd9 codes
}
\description{
three digit codes are returned unchanged, one and two digit
  codes are preceded by 00 or 0.

Non-decimal ICD-9 codes with length<5 are often ambiguous. E.g.
  100 could be 1.00 10.0 or 100 if coded incorrectly. We must assume 100 is
  really 100

pad decimal ICD-9 codes with leading zeroes

Non-decimal ICD-9 codes with length<5 are often ambiguous. E.g.
  100 could be 1.00 10.0 or 100 if coded incorrectly. We must assume 100 is
  really 100

three digit codes are returned unchanged, one and two digit
  codes are preceded by 00 or 0.
}
\keyword{internal}
\keyword{manip}

