% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9_add_leading_zeroes.icd_short_code}
\alias{icd9AddLeadingZeroes}
\alias{icd9_add_leading_zeroes.icd_decimal_code}
\alias{icd9_add_leading_zeroes.icd_short_code}
\title{Add leading zeroes to incomplete ICD codes}
\usage{
\method{icd9_add_leading_zeroes}{icd_short_code}(x)

\method{icd9_add_leading_zeroes}{icd_decimal_code}(x)

icd9AddLeadingZeroes(icd9, isShort)
}
\arguments{
\item{icd9}{is a character vector or factor of ICD-9 codes. If fewer than five characters
is given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010".}

\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than five characters is
given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010"}

\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five characters is given in a
code, then the digits are greedily assigned to hundreds, then tens, then
units, before the decimal parts. E.g. "10" becomes "010", not "0010"}

\item{major}{character vector of 'major' part of ICD-9 codes, i.e. that part
which falls before the decimal point, in decimal notation. (In five digit
notation, the 'major' part is be the first three characters (with leading
zeroes), and includes V or E prefix.}
}
\value{
character vector of ICD codes (or major part thereof)
}
\description{
Non-decimal ICD-9 codes with length<5 are often ambiguous. E.g.
  100 could be 1.00 10.0 or 100 if coded incorrectly. We must assume 100 is
  really 100
}
\keyword{internal}
\keyword{manip}

