% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convert.R
\name{icd9ChaptersToMap}
\alias{icd9ChaptersToMap}
\title{convert the chapter headings to lists of codes}
\usage{
icd9ChaptersToMap(x)
}
\arguments{
\item{x}{Either a chapter list itself, or the name of one, e.g.
icd9ChaptersSub}
}
\description{
the chapter headings can be converted into the full set of their
  children, and then used to look-up which chapter, sub-chapter, or 'major' a
  given code belongs. Always returns a map with short-form icd-9 codes. These
  can be converted en masse with \code{lapply} and \code{icd9ShortToDecimal}.
}
\seealso{
Other ICD-9 convert: \code{\link{convert}},
  \code{\link{icd9DecimalToParts}},
  \code{\link{icd9DecimalToShort}},
  \code{\link{icd9GetMajor}},
  \code{\link{icd9MajMinToCode}},
  \code{\link{icd9MajMinToDecimal}},
  \code{\link{icd9MajMinToParts}},
  \code{\link{icd9MajMinToShort}},
  \code{\link{icd9PartsToDecimal}},
  \code{\link{icd9PartsToShort}},
  \code{\link{icd9ShortToDecimal}},
  \code{\link{icd9ShortToParts}};
  \code{\link{icd9DropLeadingZeroes}},
  \code{\link{icd9DropLeadingZeroesDecimal}},
  \code{\link{icd9DropLeadingZeroesMajor}},
  \code{\link{icd9DropLeadingZeroesShort}};
  \code{\link{icd9LongToWide}};
  \code{\link{icd9WideToLong}}
}
\keyword{internal}
\keyword{manip}

