% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9Children}
\alias{icd9Children}
\alias{icd9ChildrenDecimal}
\alias{icd9ChildrenShort}
\title{Expand 5 character form 'short' ICD9 to all possible sub codes}
\usage{
icd9Children(icd9, isShort, onlyReal = FALSE)

icd9ChildrenShort(icd9Short, onlyReal = FALSE)

icd9ChildrenDecimal(icd9Decimal, onlyReal = FALSE)
}
\arguments{
\item{icd9}{is a character vector of ICD-9 codes. If fewer than five characters is
given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010"}

\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five
characters is given in a code, then the digits are greedily assigned to
hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}
}
\description{
Much faster in C++, but gains diminished with increasing numbers of input shorts, probably because I am copying memory unnecessarily.
}
\examples{
library(magrittr)
icd9ChildrenShort("10201", FALSE) # no children other than self
icd9ChildrenShort("10201", TRUE) # empty list because 102.01 is not a meaningful code
icd9ChildrenShort("003", TRUE) \%>\% icd9ExplainShort(condense = FALSE)
icd9ChildrenDecimal("100.0")
icd9ChildrenDecimal("100.00")
icd9ChildrenDecimal("2.34")
}
\seealso{
Other ICD-9 ranges: \code{\link{\%i9d\%}},
  \code{\link{\%i9mj\%}}, \code{\link{\%i9s\%}},
  \code{\link{icd9ExpandRange}},
  \code{\link{icd9ExpandRangeDecimal}},
  \code{\link{icd9ExpandRangeMajor}},
  \code{\link{icd9ExpandRangeShort}};
  \code{\link{icd9CondenseToExplain}};
  \code{\link{icd9CondenseToMajor}};
  \code{\link{icd9ExpandMinor}}
}
\keyword{manip}

