% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R, R/ranges.R
\name{icd9ChildrenShort}
\alias{icd9Children}
\alias{icd9ChildrenShort}
\alias{icd9ChildrenShort_R}
\alias{icd9Children_R}
\title{Expand 5 character form 'short' ICD9 to all possible sub codes}
\usage{
icd9ChildrenShort(icd9Short, onlyReal = FALSE)

icd9Children(icd9, isShort, onlyReal = FALSE)

icd9ChildrenShort_R(icd9Short, onlyReal = FALSE)

icd9Children_R(icd9, onlyReal = FALSE, isShort)
}
\arguments{
\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{onlyReal}{single logical value, if TRUE, will limit the search to those
codes which appear in the master list, not just syntactically valid codes.
Since nearly valid, out-dated or new codes may be missed, not limiting to
'real' values will be useful. Ultimately, there will need to be annual (and
all-time) master lists of codes and the ability to test against a given
master list given the year of the ICD-9 coding.}

\item{icd9}{is a character vector of ICD-9 codes. If fewer than five characters is
given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010"}

\item{isShort}{single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form.}

\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{icd9}{is a character vector of ICD-9 codes. If fewer than five characters is
given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010"}

\item{isShort}{single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form.}
}
\description{
Much faster in C++, but gains diminished with increasing numbers of input shorts, probably because I am copying memory unnecessarily.

expand 5 character form 'short' ICD9 to all possible sub codes
}
\examples{
library(microbenchmark)
"400" \%i9s\% "410" -> shorts
microbenchmark(icd9:::icd9ChildrenShort(shorts), icd9:::icd9ChildrenShort_R(shorts), times = 5)
}
\seealso{
Other ICD-9 ranges: \code{\link{\%i9d\%}},
  \code{\link{\%i9mj\%}}, \code{\link{\%i9s\%}},
  \code{\link{icd9ExpandRange}},
  \code{\link{icd9ExpandRangeDecimal}},
  \code{\link{icd9ExpandRangeMajor}},
  \code{\link{icd9ExpandRangeShort}};
  \code{\link{icd9ChildrenDecimal}};
  \code{\link{icd9CondenseToExplain}};
  \code{\link{icd9CondenseToMajor}};
  \code{\link{icd9ExpandMinor_R}}

Other ICD-9 ranges: \code{\link{\%i9d\%}},
  \code{\link{\%i9mj\%}}, \code{\link{\%i9s\%}},
  \code{\link{icd9ExpandRange}},
  \code{\link{icd9ExpandRangeDecimal}},
  \code{\link{icd9ExpandRangeMajor}},
  \code{\link{icd9ExpandRangeShort}};
  \code{\link{icd9ChildrenDecimal}};
  \code{\link{icd9CondenseToExplain}};
  \code{\link{icd9CondenseToMajor}};
  \code{\link{icd9ExpandMinor_R}}

Other ICD-9 ranges: \code{\link{\%i9d\%}},
  \code{\link{\%i9mj\%}}, \code{\link{\%i9s\%}},
  \code{\link{icd9ExpandRange}},
  \code{\link{icd9ExpandRangeDecimal}},
  \code{\link{icd9ExpandRangeMajor}},
  \code{\link{icd9ExpandRangeShort}};
  \code{\link{icd9ChildrenDecimal}};
  \code{\link{icd9CondenseToExplain}};
  \code{\link{icd9CondenseToMajor}};
  \code{\link{icd9ExpandMinor_R}}

Other ICD-9 ranges: \code{\link{\%i9d\%}},
  \code{\link{\%i9mj\%}}, \code{\link{\%i9s\%}},
  \code{\link{icd9ExpandRange}},
  \code{\link{icd9ExpandRangeDecimal}},
  \code{\link{icd9ExpandRangeMajor}},
  \code{\link{icd9ExpandRangeShort}};
  \code{\link{icd9ChildrenDecimal}};
  \code{\link{icd9CondenseToExplain}};
  \code{\link{icd9CondenseToMajor}};
  \code{\link{icd9ExpandMinor_R}}
}
\keyword{internal}
\keyword{manip}

