% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9ChildrenShort_alt_11}
\alias{icd9ChildrenShort_alt_11}
\title{Find child codes from vector of ICD-9 codes.}
\usage{
icd9ChildrenShort_alt_11(icd9Short, onlyReal)
}
\description{
Pure C++11 implementation using \code{unordered set} to find children of
given codes
}
\keyword{internal}
