% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{icd9cm_billable}
\alias{icd9cm_billable}
\title{Deprecated name for the ICD-9-CM billable/leaf nodes}
\format{list of data frames with columns \code{code}, \code{short_desc}, and
  \code{long_desc}, containing just the leaf (in USA, "billable") codes, and
  their descriptions. See \code{\link{icd9cm_hierarchy}} for a data frame
  with more hierarchical information.}
\source{
\url{http://www.cms.gov/Medicare/Coding/ICD9ProviderDiagnosticCodes/codes.html}
}
\description{
This now just contains the final version (2014, version number 32). This is
actually the same from 2011 to 2014. Use getter functions such as
\code{\link{get_icd9cm2014_leaf}} instead.
}
\keyword{datasets}
