% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{icd9cm_hierarchy}
\alias{icd9cm_hierarchy}
\title{Latest ICD-9-CM diagnosis codes, in flat \code{data.frame} format}
\description{
Short-form ICD-9 codes with short and long descriptions, and description of
each hierarchy level containing each code.
}
\details{
Format:  data frame
}
\seealso{
\code{\link[icd.data]{ls_icd_data}}
}
\keyword{datasets}
