% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/util.R
\name{getLatestBillableVersion}
\alias{getLatestBillableVersion}
\alias{icd9cm_latest_edition}
\title{Latest ICD-9-CM edition}
\usage{
getLatestBillableVersion()

icd9cm_latest_edition()
}
\description{
Returns a single character value with the number of the latests edition, currently \strong{32}.
}
\details{
Implemented as a function to give flexibility to calculate this, or use an option override.
}
\keyword{internal}

