% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9Comorbid_alt_Taskloop}
\alias{icd9Comorbid_alt_Taskloop}
\alias{icd9Comorbid_alt_Taskloop2}
\title{Simpler comorbidity assignment}
\usage{
icd9Comorbid_alt_Taskloop(icd9df, icd9Mapping, visitId, icd9Field,
  threads = 8L, chunk_size = 256L, omp_chunk_size = 1L,
  aggregate = TRUE)

icd9Comorbid_alt_Taskloop2(icd9df, icd9Mapping, visitId, icd9Field,
  threads = 8L, chunk_size = 256L, omp_chunk_size = 1L,
  aggregate = TRUE)
}
\description{
Re-written without OpenMP initially, but structured more simply, with the motivation of
using modern compiler features and OpenMP 4.5 with 'taskloop' construct.
\url{https://developers.redhat.com/blog/2016/03/22/what-is-new-in-openmp-4-5-3/}
}
\details{
# basic test
# use tests/testthat/helper-base.R for two_pts and two_map
icd_comorbid(two_pts, two_map, comorbid_fun = icd:::icd9Comorbid_alt_Taskloop)

vermont_dx %>% icd_wide_to_long() -> vt
microbenchmark::microbenchmark(
  res1 <- icd_comorbid(vt, icd9_map_ahrq, comorbid_fun = icd:::icd9ComorbidShortCpp),
  res2 <- icd_comorbid(vt, icd9_map_ahrq, comorbid_fun = icd:::icd9Comorbid_alt_Taskloop),
  times = 50)
identical(res1, res2)
}
\section{Functions}{
\itemize{
\item \code{icd9Comorbid_alt_Taskloop2}: Taskloop but finish with R transpose
}}

\keyword{internal}
