% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convert.R
\name{icd9ComorbidMatToDf}
\alias{icd9ComorbidMatToDf}
\title{convert matrix of comorbidities into data frame, preserving visitId information}
\usage{
icd9ComorbidMatToDf(x, visitId = NULL,
  stringsAsFactors = getOption("stringsAsFactors"))

icd9ComorbidMatToDf(x, visitId = NULL,
  stringsAsFactors = getOption("stringsAsFactors"))
}
\description{
convert matrix of comorbidities into data frame, preserving visitId information

convert matrix of comorbidities into data frame, preserving visitId information
}

