% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{icd9ComorbidMatToDf}
\alias{icd9ComorbidDfToMat}
\alias{icd9ComorbidMatToDf}
\title{Deprecated: convert matrix of comorbidities to data frame}
\usage{
icd9ComorbidMatToDf(x, visitId = "visit_name",
  stringsAsFactors = getOption("stringsAsFactors"))

icd9ComorbidDfToMat(x, visitId = get_visit_name(x),
  stringsAsFactors = getOption("stringsAsFactors"))
}
\arguments{
\item{x}{matrix or data frame of comorbidities}

\item{visitId}{Deprecated. Use \code{visit_name}}

\item{stringsAsFactors}{Single logical value, describing whether the resulting data frame should have strings, e.g.
\code{visit_id} converted to factor. Default is to follow the current session option. This is identical to the
argument used in, among other base functions \code{as.data.frame}.}
}
\description{
Use \code{icd_comorbid_mat_to_df} or \code{icd_comorbid_df_to_mat} instead.
}

