% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/explain.R
\name{icd9Condense}
\alias{icd9Condense}
\alias{icd9CondenseDecimal}
\alias{icd9CondenseShort}
\title{Condense ICD-9 code by replacing complete families with parent codes}
\usage{
icd9Condense(icd9, isShort = icd9GuessIsShort(icd9), onlyReal = NULL)

icd9Condense(icd9, isShort = icd9GuessIsShort(icd9), onlyReal = NULL)

icd9CondenseDecimal(icd9Decimal, onlyReal = NULL)

icd9CondenseShort(icd9Short, onlyReal = NULL, onlyBillable = NULL,
  warn = FALSE)
}
\arguments{
\item{icd9}{is
a character vector or factor of ICD-9 codes. If fewer than five characters
is given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010".}

\item{isShort}{single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form. Where reasonable, this
is guessed from the input data.}

\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five
characters is given in a code, then the digits are greedily assigned to
hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{onlyBillable}{single logical value, if \code{TRUE}, describes the input
data, stating that it only contains billable codes. Usually, the function
will try to guess this, but if you know in advance what they should be, the
functions can optionally warn if this is incorrect, and save some
computation time. The billable codes are derived from the CMS list. The
most recent version is used by default.}

\item{warnReal}{single logical value, if \code{TRUE}, give warnings when
there is discrepancy between \code{onlyReal} being \code{TRUE} yet data
containing undefined codes.}
}
\description{
This can be thought of as the inverse operation to
  \code{icd9Children}.
}
\seealso{
Other ICD-9 ranges: \code{\link{\%i9d\%}},
  \code{\link{\%i9da\%}}, \code{\link{\%i9mj\%}},
  \code{\link{\%i9s\%}}, \code{\link{\%i9sa\%}},
  \code{\link{icd9ExpandRange}},
  \code{\link{icd9ExpandRangeDecimal}},
  \code{\link{icd9ExpandRangeMajor}},
  \code{\link{icd9ExpandRangeShort}};
  \code{\link{icd9Children}}, \code{\link{icd9Children}},
  \code{\link{icd9ChildrenDecimal}},
  \code{\link{icd9ChildrenShort}};
  \code{\link{icd9ExpandMinor}}
}

