% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{icd9DiffComorbid}
\alias{icd9DiffComorbid}
\title{Deprecated: find difference between comorbidity maps}
\usage{
icd9DiffComorbid(x, y, names = NULL, x.names = NULL, y.names = NULL,
  show = TRUE, explain = TRUE)
}
\arguments{
\item{x}{comorbidity map}

\item{y}{comorbidity map}

\item{names}{character vector of the comorbidity names}

\item{x.names}{character vector of the comorbidity names from \code{x} to
compare}

\item{y.names}{character vector of the comorbidity names from \code{y} to}

\item{show}{logical}

\item{explain}{logical
compare}
}
\description{
Use \code{icd_diff_comorbid} instead.
}

