% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{icd9FilterValid}
\alias{icd9FilterInvalid}
\alias{icd9FilterPoa}
\alias{icd9FilterPoaNo}
\alias{icd9FilterPoaNotNo}
\alias{icd9FilterPoaNotYes}
\alias{icd9FilterPoaYes}
\alias{icd9FilterValid}
\title{Deprecated: filter valid ICD-9 codes in or out}
\usage{
icd9FilterValid(icd9df, icd9Field = NULL, isShort = NULL, invert = FALSE)

icd9FilterInvalid(icd9df, icd9Field = NULL, isShort = NULL,
  invert = FALSE)

icd9FilterPoa(icd9df, poaField = "poa", poa = icd9PoaChoices)

icd9FilterPoaYes(icd9df, poaField = "poa")

icd9FilterPoaNo(icd9df, poaField = "poa")

icd9FilterPoaNotNo(icd9df, poaField = "poa")

icd9FilterPoaNotYes(icd9df, poaField = "poa")
}
\arguments{
\item{icd9df}{Deprecated. Use \code{x} to provide a data frame input}

\item{icd9Field}{The column in the data frame which contains the ICD codes.
This is a character vector of length one. If it is \code{NULL}, \code{icd9}
will attempt to guess the column name, looking for progressively less
likely possibilities until it matches a single column. Failing this, it
will take the first column in the data frame. Specifying the column using
this argument avoids the guesswork.}

\item{isShort}{Deprecated. Single logical value which determines whether the
ICD-9 code provided is in short (\code{TRUE}) or decimal (\code{FALSE})
form. Where reasonable, this is guessed from the input data.}

\item{invert}{logical}

\item{poaField}{Deprecated. Use \code{poa_name} and the \code{icd_filter_poa}
family of functions.}

\item{poa}{single character value, being one of \code{Yes}, \code{No},
\code{NotYes}, and \code{NotNo}, indicating whether to account for
comorbidities flagged as present-on-arrival. This is not a simple flag,
because many codes are exempt, unspecified, or unknown. The intermediate
codes, such as "exempt", "unknown" and \code{NA} mean that "yes" is not the
same as "not no."}
}
\description{
Use \code{icd_filter_valid} instead
}

