% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{icd9GetChapters}
\alias{icd9GetChapters}
\title{Deprecated: get chapters for ICD-9 codes}
\usage{
icd9GetChapters(icd9, isShort = icd_guess_short(icd9), verbose = FALSE)
}
\arguments{
\item{icd9}{is a character vector or factor of ICD-9 codes. If fewer than five characters
is given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010".}

\item{isShort}{Deprecated. Single logical value which determines whether the
ICD-9 code provided is in short (\code{TRUE}) or decimal (\code{FALSE})
form. Where reasonable, this is guessed from the input data.}

\item{verbose}{single logical value, defaults to \code{FALSE} in most functions.}
}
\description{
Use \code{icd9_get_chapters} instead.
}

