% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{icd9IsN}
\alias{icd9IsE}
\alias{icd9IsN}
\alias{icd9IsV}
\title{Deprecated: is a code a particular ICD-9 sub-type?}
\usage{
icd9IsN(icd9)

icd9IsV(icd9)

icd9IsE(icd9)
}
\arguments{
\item{icd9}{is a character vector or factor of ICD-9 codes. If fewer than five characters
is given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010".}
}
\description{
Not exported anymore. If you really need it, use:
\code{icd9_is_n} etc.
}

