% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convert.R
\name{icd9LongToWide}
\alias{icd9LongToWide}
\title{convert ICD data from long to wide format}
\usage{
icd9LongToWide(icd9df, visitId = "visitId", icd9Field = "icd9",
  prefix = "icd_", min.width = 0, aggregate = TRUE, return.df = TRUE)
}
\arguments{
\item{visitId}{single character, if NULL will use "visitId" if exists,
otherwise the first column}

\item{icd9Field}{single character string with name of column containing the ICD
codes. If left as NULL, the field is guessed to be the first matching ICD
or icd, and a warning is given if multiple columns match.}

\item{prefix}{character, default "icd_" to prefix new columns}

\item{min.width,}{single integer, if specified, writes out this many columns even
if no patients have that many codes. Must be greater than or equal to the
maximum number of codes per patient.}

\item{x}{data.frame of long-form data, one column for visitId and one for ICD
code}
}
\description{
This is more complicated than reshape or reshape2::dcast allows.
  This is a reasonably simple solution using built-in functions.
}
\examples{
longdf <- data.frame(visitId = c("a", "b", "b", "c"),
    icd9 = c("441", "4424", "443", "441"))
  icd9LongToWide(longdf)
  icd9LongToWide(longdf, prefix = "ICD10_", empty = "")
}
\seealso{
Other ICD-9 convert: \code{\link{convert}},
  \code{\link{convert}}, \code{\link{icd9DecimalToParts}},
  \code{\link{icd9DecimalToShort}},
  \code{\link{icd9GetMajor}},
  \code{\link{icd9MajMinToCode}},
  \code{\link{icd9MajMinToDecimal}},
  \code{\link{icd9MajMinToParts}},
  \code{\link{icd9MajMinToShort}},
  \code{\link{icd9PartsToDecimal}},
  \code{\link{icd9PartsToShort}},
  \code{\link{icd9ShortToDecimal}},
  \code{\link{icd9ShortToParts}};
  \code{\link{icd9ChaptersToMap}};
  \code{\link{icd9DropLeadingZeroes}},
  \code{\link{icd9DropLeadingZeroesDecimal}},
  \code{\link{icd9DropLeadingZeroesMajor}},
  \code{\link{icd9DropLeadingZeroesShort}};
  \code{\link{icd9WideToLong}}
}
\keyword{internal}

