% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9MajMinToCode}
\alias{icd9MajMinToCode}
\title{Convert major and minor vectors to single code}
\usage{
icd9MajMinToCode(major, minor, isShort)
}
\description{
In debug mode, will check that major and minor are same length.
}
\examples{
\dontrun{
n <- 5
majors <- as.character(sample(1:999, n, replace = TRUE))
minors <- as.character(sample(0:99, n, replace = TRUE))
microbenchmark::microbenchmark(
  icd9MajMinToCode(majors, minors, TRUE),
  icd9MajMinToCodeStd(majors, minors, TRUE),
  times = 3
)
}
# std method vastly quicker, e.g. x100 when n=5000
}

