% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9MajMinToCode}
\alias{icd9MajMinToCode}
\alias{icd9MajMinToCodePrePadded}
\title{Convert mjr and mnr vectors to single code}
\usage{
icd9MajMinToCode(mjr, mnr, isShort)

icd9MajMinToCodePrePadded(mjr, mnr, isShort)
}
\arguments{
\item{isShort}{Deprecated. Single logical value which determines whether the
ICD-9 code provided is in short (\code{TRUE}) or decimal (\code{FALSE})
form. Where reasonable, this is guessed from the input data.}

\item{major}{character vector of 'major' part of ICD-9 codes, i.e. that part
which falls before the decimal point, in decimal notation. (In five digit
notation, the 'major' part is be the first three characters (with leading
zeroes), and includes V or E prefix.}

\item{minor}{character vector of 'minor' part of ICD-9 codes, i.e. that part
which falls after the decimal point, in decimal notation. (In 'short' five
digit notation, the 'major' part is the first three characters including
leading zeroes which may be inferred; the last two characters represent the
'minor' part.)}
}
\description{
In debug mode, will check that mjr and mnr are same length.
}
\section{Functions}{
\itemize{
\item \code{icd9MajMinToCodePrePadded}: Same as \code{icd9MajMinToCode} but assume
codes are already trimmed and correctly padded with zeros, e.g. E001, V09,
001. This version does handle NA values correctly. ' @keywords internal
}}
\examples{
\dontrun{
n <- 5e6
mjrs <- as.character(sample(1:999, n, replace = TRUE))
mnrs <- as.character(sample(0:99, n, replace = TRUE))
microbenchmark::microbenchmark(
  icd9MajMinToCode(mjrs, mnrs, TRUE),
  icd9MajMinToCodeStd(mjrs, mnrs, TRUE),
  icd9MajMinToCodePrePadded(mjrs, mnrs, TRUE),
  times = 10
)
}
# std method about the same with O3 (4\% faster, but no NA handling), but 50\% quicker with O0
# std method without doing padding is 5 times quicker than previous...
}
\keyword{internal}
\keyword{manip}

