% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9RandomShort}
\alias{icd9RandomShort}
\title{generate random short-form icd9 E codes}
\usage{
icd9RandomShort(n = 5L)
}
\description{
Very dirty pseudo-random by picking numeric, V or E based on
  modulo 3 of the number
}
\keyword{internal}

