% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9RandomShortN}
\alias{icd9RandomShortN}
\alias{icd9RandomShortV}
\alias{icd9RandomShortE}
\alias{icd9RandomShort}
\title{Generate random short-form ICD-9 codes}
\usage{
icd9RandomShortN(n = 5L)

icd9RandomShortV(n = 5L)

icd9RandomShortE(n = 5L)

icd9RandomShort(n = 5L)
}
\description{
Quick pseudo-random by picking numeric, 'V' or 'E' based on modulo three of
the number
}
\keyword{internal}
