% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9RandomShortE}
\alias{icd9RandomShortE}
\title{Generate random short-form ICD-9 E codes}
\usage{
icd9RandomShortE(n = 5L)
}
\description{
Generate a character vector of random short-form ICD-9 E codes
}
\keyword{internal}

