% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9RandomShortN}
\alias{icd9RandomShortN}
\title{generate random short-form numeric icd9 codes}
\usage{
icd9RandomShortN(n = 5L)
}
\description{
Generate a character vector of random short-form ICD-9 numeric codes
}
\keyword{internal}

