% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icd9ShortToDecimal}
\alias{icd9DecimalToShortOld}
\alias{icd9ShortToDecimal}
\title{Convert ICD-9 codes between short and decimal forms}
\usage{
icd9ShortToDecimal(icd9Short)

icd9DecimalToShortOld(icd9Decimal)
}
\arguments{
\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five
characters is given in a code, then the digits are greedily assigned to
hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}
}
\description{
Convert ICD-9 codes between short and decimal forms
}

