% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children.R
\name{icd_children_defined}
\alias{icd_children_defined}
\alias{icd_children_defined.icd10cm}
\title{defined children of ICD codes}
\usage{
icd_children_defined(x)

\method{icd_children_defined}{icd10cm}(x,
  short_code = icd_guess_short.icd10(x))
}
\description{
defined ICD-10 children based on 2016 ICD-10-CM list. "defined" may be a
three digit code, or a leaf node. This is distinct from 'billable'.
}
\section{Methods (by class)}{
\itemize{
\item \code{icd10cm}: get the children of ICD-10 code(s)
}}
\keyword{internal}

