% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children.R
\name{icd_children_real}
\alias{icd_children_real}
\alias{icd_children_real.icd10}
\title{real icd10 children based on 2016 ICD-10-CM list}
\usage{
icd_children_real(x)

\method{icd_children_real}{icd10}(icd, short = icd_guess_short(icd))
}
\description{
real icd10 children based on 2016 ICD-10-CM list
}
\section{Methods (by class)}{
\itemize{
\item \code{icd10}: get the children of ICD-10 code(s)
}}
\keyword{internal}

