% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R, R/util-test.R
\name{icd_classes_ordered}
\alias{icd_classes_ordered}
\alias{expect_icd_classes_ordered}
\title{prefer an order of classes}
\usage{
icd_classes_ordered(x)

expect_icd_classes_ordered(x)
}
\arguments{
\item{x}{any object which may or may not have classes from this package}
}
\description{
The order of classes can matter because, for some functions, we'd prefer to
decide what to do based on a higher level structure, e.g. whether the
structure is a comorbidity map before caring if it is ICD-9 or ICD-10. I
can't see how it matters whether we prioritize long/wide and short/decimal
yet, so won't test.
}
\section{Functions}{
\itemize{
\item \code{expect_icd_classes_ordered}: \code{testthat} \code{expect} function
for ICD classes to be in correct order.
}}

\keyword{debugging}
\keyword{internal}
