% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/comorbid.R, R/deprecated.R
\name{icd9ComorbidShortCpp}
\alias{icd9Comorbid}
\alias{icd9ComorbidAhrq}
\alias{icd9ComorbidElix}
\alias{icd9ComorbidQuanDeyo}
\alias{icd9ComorbidQuanElix}
\alias{icd9ComorbidQuanElixhauser}
\alias{icd9ComorbidShort}
\alias{icd9ComorbidShortCpp}
\alias{icd9Comorbidities}
\alias{icd9ComorbiditiesAhrq}
\alias{icd9ComorbiditiesElixHauser}
\alias{icd9ComorbiditiesQuanDeyo}
\alias{icd9ComorbiditiesQuanElixhauser}
\alias{icd_comorbid}
\alias{icd_comorbid.default}
\alias{icd_comorbid.icd10}
\alias{icd_comorbid.icd9}
\alias{icd_comorbid_ahrq}
\alias{icd_comorbid_ahrq.data.frame}
\alias{icd_comorbid_ahrq.icd10}
\alias{icd_comorbid_ahrq.icd9}
\alias{icd_comorbid_common}
\alias{icd_comorbid_elix}
\alias{icd_comorbid_elix.data.frame}
\alias{icd_comorbid_elix.icd9}
\alias{icd_comorbid_quan_deyo}
\alias{icd_comorbid_quan_deyo.data.frame}
\alias{icd_comorbid_quan_deyo.icd9}
\alias{icd_comorbid_quan_elix}
\alias{icd_comorbid_quan_elix.data.frame}
\alias{icd_comorbid_quan_elix.icd9}
\title{find comorbidities from ICD-9 codes.}
\usage{
icd9ComorbidShortCpp(icd9df, icd9Mapping, visitId, icd9Field, threads = 8L,
  chunk_size = 256L, omp_chunk_size = 1L, aggregate = TRUE)

icd_comorbid(...)

\method{icd_comorbid}{default}(x, ...)

\method{icd_comorbid}{icd10}(x, map, visit_name = NULL, icd_name = NULL,
  short_code = NULL, short_map = icd_guess_short.list(map),
  return_df = FALSE, ...)

\method{icd_comorbid}{icd9}(x, map, visit_name = NULL, icd_name = NULL,
  short_code = icd_guess_short.data.frame(x, icd_name = icd_name),
  short_map = icd_guess_short.list(map), return_df = FALSE, ...)

icd_comorbid_common(x, map, visit_name = NULL, icd_name, short_code,
  short_map, return_df = FALSE, ...)

icd_comorbid_ahrq(...)

icd_comorbid_quan_elix(...)

icd_comorbid_quan_deyo(...)

icd_comorbid_elix(...)

\method{icd_comorbid_ahrq}{data.frame}(x, ..., icd_name = get_icd_name(x))

\method{icd_comorbid_elix}{data.frame}(x, ..., icd_name = get_icd_name(x))

\method{icd_comorbid_quan_elix}{data.frame}(x, ...,
  icd_name = get_icd_name(x))

\method{icd_comorbid_quan_deyo}{data.frame}(x, ...,
  icd_name = get_icd_name(x))

\method{icd_comorbid_ahrq}{icd9}(..., abbrev_names = TRUE, hierarchy = TRUE)

\method{icd_comorbid_ahrq}{icd10}(..., abbrev_names = TRUE,
  hierarchy = TRUE)

\method{icd_comorbid_quan_deyo}{icd9}(..., abbrev_names = TRUE,
  hierarchy = TRUE)

\method{icd_comorbid_quan_elix}{icd9}(..., abbrev_names = TRUE,
  hierarchy = TRUE)

\method{icd_comorbid_elix}{icd9}(..., abbrev_names = TRUE, hierarchy = TRUE)

icd9ComorbidShort(...)

icd9Comorbidities(...)

icd9ComorbiditiesAhrq(...)

icd9ComorbiditiesElixHauser(...)

icd9ComorbiditiesQuanDeyo(...)

icd9ComorbiditiesQuanElixhauser(...)

icd9ComorbidQuanElixhauser(...)

icd9Comorbid(icd9df, icd9Mapping, visitId = NULL, icd9Field = NULL,
  isShort = icd_guess_short.icd9(icd9df),
  isShortMapping = icd_guess_short.icd9(icd9Mapping), return.df = FALSE,
  ...)

icd9ComorbidAhrq(icd9df, visitId = NULL, icd9Field = NULL,
  isShort = icd9GuessIsShort(icd9df[1:100, icd9Field]), return.df = FALSE,
  abbrevNames = TRUE, applyHierarchy = TRUE, ...)

icd9ComorbidElix(icd9df, visitId = NULL, icd9Field = NULL,
  isShort = icd9GuessIsShort(icd9df[1:100, icd9Field]), return.df = FALSE,
  abbrevNames = TRUE, applyHierarchy = TRUE, ...)

icd9ComorbidQuanElix(icd9df, visitId = NULL, icd9Field = NULL,
  isShort = icd9GuessIsShort(icd9df[1:100, icd9Field]), return.df = FALSE,
  abbrevNames = TRUE, applyHierarchy = TRUE, ...)

icd9ComorbidQuanDeyo(icd9df, visitId = NULL, icd9Field = NULL,
  isShort = icd9GuessIsShort(icd9df[1:100, icd9Field]), return.df = FALSE,
  abbrevNames = TRUE, applyHierarchy = TRUE, ...)
}
\arguments{
\item{aggregate}{single logical value, if /code{TRUE}, then take (possible
much) more time to aggregate out-of-sequence visit IDs in the icd9df
data.frame. If this is \code{FALSE}, then each contiguous group of visit
IDs will result in a row of comorbidities in the output data. If you know
your visitIds are possible disordered, then use \code{TRUE}.}

\item{...}{arguments passed to the corresponding function from the alias.
E.g. all the arguments passed to \code{icd9ComorbiditiesAhrq} are passed on
to \code{icd9ComorbidAhrq}}

\item{map}{list (or name of a list if character vector of length one is given
as argument) of the comorbidities with each top-level list item containing
a vector of decimal ICD-9 codes. This is in the form of a list, with the
names of the items corresponding to the comorbidities (e.g. 'HTN', or
'diabetes') and the contents of each list item being a character vector of
short-form (no decimal place but ideally zero left-padded) ICD-9 codes. No
default: user should prefer to use the derivative functions, e.g.
\code{icd_comorbid_ahrq}, since these also provide appropriate naming for
the fields, and squashing the hierarchy (see \code{hierarchy} below)}

\item{visit_name}{The name of the column in the data frame which contains the
patient or visit identifier. Typically this is the visit identifier, since
patients come leave and enter hospital with different ICD-9 codes. It is a
character vector of length one. If left empty, or \code{NULL}, then an
attempt is made to guess which field has the ID for the patient encounter
(not a patient ID, although this can of course be specified directly). The
guesses proceed until a single match is made. Data frames may be wide with
many matching fields, so to avoid false positives, anything but a single
match is rejected. If there are no successful guesses, and \code{visit_id}
was not specified, then the first column of the data frame is used.}

\item{icd_name}{The column in the \code{data.frame} which contains the ICD
codes. This is a character vector of length one. If it is \code{NULL},
\code{icd9} will attempt to guess the column name, looking for
progressively less likely possibilities until it matches a single column.
Failing this, it will take the first column in the data frame. Specifying
the column using this argument avoids the guesswork.}

\item{short_code}{single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form. Where reasonable, this
is guessed from the input data.}

\item{short_map}{Same as short, but applied to \code{map} instead of
\code{icd_df}. All the codes in a mapping should be of the same type, i.e.
short or decimal.}

\item{abbrev_names}{single logical value that defaults to \code{TRUE}, in
which case the shorter human-readable names stored in e.g.
\code{ahrqComorbidNamesAbbrev} are applied to the data frame column names.}

\item{hierarchy}{single logical value that defaults to \code{TRUE}, in
which case the hierarchy defined for the mapping is applied. E.g. in
Elixhauser, you can't have uncomplicated and complicated diabetes both
flagged.}

\item{...}{arguments passed to the corresponding function from the alias.
E.g. all the arguments passed to \code{icd9ComorbiditiesAhrq} are passed on
to \code{icd9ComorbidAhrq}}
}
\description{
\link{Rcpp} approach to comorbidity assignment with OpenMP and
  vector of integers strategy. It is very fast, and most time is now spent
  setting up the data to be passed in.

This is the main function which extracts co-morbidities from a set of ICD-9
codes. This is when some trivial post-processing of the comorbidity data is
done, e.g. renaming to human-friendly field names, and updating fields
according to rules. The exact fields from the original mappings can be
obtained using \code{hierarchy = FALSE}, but for comorbidity counting,
Charlson Score, etc., the rules should be applied.
}
\details{
There is a change in behavior from previous versions. The visit_name
  column is (implicitly) sorted by using std::set container. Previously, the
  visit_name output order was whatever R's \code{aggregate} produced.

  The threading of the C++ can be controlled using e.g.
  \code{option(icd.threads = 4)}. If it is not set, the number of cores in
  the machine is used.

For ICD-10, there are two lookup methods. The classic lookup, as
  used for ICD-9 codes, assumes any possible code is available to match in
  the comorbidity map. However, for ICD-10-CM, there are too many possible
  codes, specifying subsequent encounters, laterality, etc., etc., so this is
  too bulky. However for some mappings, there are exact definitions, e.g.
  AHRQ seems to specify everything (for a given ICD-10-CM year)

The common comorbidity calculation code does not depend on ICD type.
  There is some type conversion so the map and input codes are all in 'short'
  format, fast factor generation, then fast comorbidity assignment.

For Deyo's Charlson comorbidities, strictly speaking, there is no
  dropping of more e.g. uncomplicated 'DM' if complicated 'DM' exists,
  however, this is probably useful, in general and is essential when
  calculating the Charlson score.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default method for getting comorbidities, guessing
ICD version. ICD version can be specified by setting the class of the input
data directly, or by calling, for example, \code{icd9(your_data_frame)}.

\item \code{icd10}: Get comorbidities from ICD-10 codes. With this
method, it relies on exact matching, but not every of billions of possible
ICD-10/ICD-10-CM codes are included in the mappings, so it will likely give
incomplete results, without searching for parents of the input codes until
a match is found in the map. TODO: this is incomplete!

\item \code{icd9}: Get comorbidities from \code{data.frame} of ICD-9
codes
}}
\examples{
  # optional but often helpful
  library(magrittr, warn.conflicts = FALSE, quietly = TRUE)

  pts <- data.frame(visit_name = c("2", "1", "2", "3", "3"),
                   icd9 = c("39891", "40110", "09322", "41514", "39891")) \%>\%
                   icd_long_data \%>\% icd9
   icd_comorbid(pts, icd9_map_ahrq, short_code = TRUE) # visit_name is now sorted
}
\keyword{internal}

