% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/convert.R
\name{icd_short_to_parts.icd10}
\alias{icd_short_to_parts.icd10}
\alias{icd_decimal_to_parts.icd10}
\alias{icd_decimal_to_parts}
\alias{icd_decimal_to_parts.icd9}
\title{Convert decimal ICD codes to component parts}
\usage{
\method{icd_short_to_parts}{icd10}(x, mnr_empty = "")

\method{icd_decimal_to_parts}{icd10}(x, mnr_empty = "")

icd_decimal_to_parts(x, mnr_empty = "")

\method{icd_decimal_to_parts}{icd9}(x, mnr_empty = "")
}
\description{
Convert decimal ICD codes to component parts
}
\section{Methods (by class)}{
\itemize{
\item \code{icd10}: Convert short ICD-10 code to parts

\item \code{icd10}: Convert decimal ICD-10 code to parts. This
shares almost 100% code with the ICD-9 version: someday combine the common
code.

\item \code{icd9}: Convert decimal ICD-9 code to parts
}}

\keyword{internal}
\keyword{manip}
