% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{icd_deprecated}
\alias{icd_deprecated}
\title{wrapper for \code{.Deprecated}}
\usage{
icd_deprecated(...)
}
\arguments{
\item{...}{arguments passed to \code{.Deprecated}}
}
\description{
Don't show warnings when using deprecated code. This allows people to
continue using their \code{icd9} code without any modifications, and without
flooding the user with warnings.
}
\examples{
\dontrun{
# turn off all deprecated warnings for this package
options("icd.warn_deprecated" = FALSE)
}
}
\keyword{internal}

