% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{icd_expand_minor}
\alias{icd_expand_minor}
\title{expand decimal part of ICD-9 code to cover all possible sub-codes}
\usage{
icd_expand_minor(mnr, ...)
}
\arguments{
\item{minor}{character vector of 'minor' part of ICD-9 codes, i.e. that part
which falls after the decimal point, in decimal notation. (In 'short' five
digit notation, the 'major' part is the first three characters including
leading zeroes which may be inferred; the last two characters represent the
'minor' part.)}

\item{isE}{single logical, which if TRUE, treats the minor as part of an E
code (which is one character), as opposed to a V or numeric-only code,
which is two character. Default is \code{FALSE}.}
}
\value{
NA for invalid minor, otherwise a vector of all possible (perhaps
  non-existent) sub-divisions.
}
\description{
Accepts a single number or character input starting point for
  generation of all possible decimal parts of ICD9 code. e.g. giving an empty
  input will fill out 111 combinations, e..g .1 .11 .12 .... .2 ....
}
\examples{
\dontrun{
  # return all possible decimal parts of ICD9 codes (111 in total)
  length(icd:::icd_expand_minor(as.icd9(""), is_e = FALSE))
  icd:::icd_expand_minor(as.icd9("1")) # "1"  "10" "11" "12" "13" "14" "15" "16" "17" "18" "19"
}
}
\seealso{
Other ICD-9 ranges: \code{\link{icd_children}},
  \code{\link{icd_condense}},
  \code{\link{icd_expand_range}}
}
\keyword{internal}
\keyword{manip}

