% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{icd_expand_range.icd10cm}
\alias{icd_expand_range.icd10cm}
\title{expand range of ICD-10 codes returning only defined codes in ICD-10-CM}
\usage{
\method{icd_expand_range}{icd10cm}(start, end,
  short_code = icd_guess_short.icd10(c(start, end)), defined = TRUE, ...)
}
\arguments{
\item{start}{character vector of length one containing an ICD code}

\item{end}{character vector of length one containing an ICD code}
}
\description{
This will need generalizing to any list of 'defined' codes, e.g. WHO or other
}
\details{
There are so many permutations of alphaunmerics after the decimal place
thatit would be easy to generate many millions with simple ranges, the vast
majority of which would be undefined.
}
\keyword{internal}

