% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{icd_expand_range}
\alias{\%i9d\%}
\alias{\%i9da\%}
\alias{\%i9mj\%}
\alias{\%i9s\%}
\alias{\%i9sa\%}
\alias{icd9ExpandRange}
\alias{icd9ExpandRangeDecimal}
\alias{icd9ExpandRangeMajor}
\alias{icd9ExpandRangeShort}
\alias{icd9_expand_range_decimal}
\alias{icd9_expand_range_short}
\alias{icd_expand_range}
\alias{icd_expand_range.icd9}
\title{take two ICD-9 codes and expand range to include all child codes}
\usage{
icd_expand_range(start, end, ...)

icd9ExpandRange(start, end, isShort = icd_guess_short_code(c(start, end)),
  onlyReal = TRUE, excludeAmbiguousStart = TRUE,
  excludeAmbiguousEnd = TRUE)

icd9ExpandRangeShort(start, end, onlyReal = TRUE,
  excludeAmbiguousStart = TRUE, excludeAmbiguousEnd = TRUE)

icd9ExpandRangeDecimal(start, end, onlyReal = TRUE,
  excludeAmbiguousStart = TRUE, excludeAmbiguousEnd = TRUE)

icd9ExpandRangeMajor(start, end, onlyReal = TRUE)

\method{icd_expand_range}{icd9}(start, end,
  short_code = icd_guess_short_code(c(start, end)), real = TRUE,
  excludeAmbiguousStart = TRUE, excludeAmbiguousEnd = TRUE)

icd9_expand_range_short(start, end, real = TRUE,
  excludeAmbiguousStart = TRUE, excludeAmbiguousEnd = TRUE)

icd9_expand_range_decimal(start, end, real = TRUE,
  excludeAmbiguousStart = TRUE, excludeAmbiguousEnd = TRUE)

start \%i9da\% end

start \%i9sa\% end

start \%i9d\% end

start \%i9mj\% end

start \%i9s\% end
}
\arguments{
\item{start, end}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{isShort}{single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form. Where reasonable, this
is guessed from the input data.}

\item{onlyReal}{single logical value, if TRUE, will limit the search to those
codes which appear in the master list, not just syntactically valid codes.
Since nearly valid, out-dated or new codes may be missed, not limiting to
'real' values will be useful. Ultimately, there will need to be annual (and
all-time) master lists of codes and the ability to test against a given
master list given the year of the ICD-9 coding.}

\item{excludeAmbiguousStart}{single logical value, if \code{TRUE} the range
returned will not include codes which are explicitly listed in the range,
but would imply a broader range than specified. E.g. \code{V10 %%i9sa%%
V1009} would by default (\code{FALSE}) include \code{V10} even though
\code{V10} itself is parent to everything up to \code{V11}.}

\item{excludeAmbiguousEnd}{single logical, same as
\code{excludeAmbiguousStart} but affects codes at the end of the range.
E.g. 99.99 to 101.01 would by default exclude 101 and 101.0}
}
\description{
this is cumbersome code, covering a whole load of edge cases
  relating to the fact that icd9 codes are \strong{not} in numeric order. An
  alternative strategy would be to list all the ICD9 codes, then a range
  would just pick out start and finish positions, and return subset of the
  list. Not all ICD-9 codes are valid, including some parent codes which have
  valid children. However, I expect at least some of these have been used in
  some billing databases.

  As with \code{link{icd9ExpandRangeShort}} great care is taken not to
  include codes which have children not in the range. E.g. "100.9" to "101.1"
  would _not_ include code "101".

  \code{onlyReal} default is \code{TRUE} (a change from previous versions)
  since this is far more likely to be useful to the end user.
}
\details{
Expand range of short_code-form ICD-9 codes
}
\section{Methods (by class)}{
\itemize{
\item \code{icd9}: Expand a range of ICD-9 codes
}}
\examples{
"4280 " \%i9s\% "4289 "
"4280 " \%i9s\% "42821"
"42799 " \%i9sa\% "42802" # doesn't include 428 or 4280
"427.99 " \%i9da\% "428.02"
"V80 " \%i9s\% " V810 "
}
\seealso{
Other ICD-9 ranges: \code{\link{icd_children}},
  \code{\link{icd_condense}},
  \code{\link{icd_expand_minor}}
}
\keyword{internal}

