% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{icd_explain}
\alias{icd9Explain}
\alias{icd9Explain.numeric}
\alias{icd9ExplainDecimal}
\alias{icd9ExplainShort}
\alias{icd_explain}
\alias{icd_explain.character}
\alias{icd_explain.factor}
\alias{icd_explain.list}
\title{explain ICD9 codes}
\usage{
icd_explain(icd9, short_code = icd_guess_short(icd9), condense = TRUE,
  brief = FALSE, warn = TRUE)

\method{icd_explain}{list}(icd9, short_code = icd_guess_short(icd9),
  condense = TRUE, brief = FALSE, warn = TRUE)

\method{icd_explain}{factor}(icd9, short_code = icd_guess_short(icd9),
  condense = TRUE, brief = FALSE, warn = TRUE)

\method{icd_explain}{character}(icd9, short_code = icd_guess_short(icd9),
  condense = TRUE, brief = FALSE, warn = TRUE)

icd9Explain(icd9, short_code = icd_guess_short(icd9), condense = TRUE,
  brief = FALSE, warn = TRUE)

icd9Explain.numeric(icd9, short_code = icd_guess_short(icd9),
  condense = TRUE, brief = FALSE, warn = FALSE)

icd9ExplainShort(icd9Short, condense = TRUE, brief = FALSE, warn = TRUE)

icd9ExplainDecimal(icd9Decimal, condense = TRUE, brief = FALSE,
  warn = TRUE)
}
\arguments{
\item{icd9}{is
a character vector or factor of ICD-9 codes. If fewer than five characters
is given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010".}

\item{short_code}{single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form. Where reasonable, this
is guessed from the input data.}

\item{condense}{single logical value which indicates whether to condense
the given set of ICD-9 codes by replacing subsets of codes with 'parent'
codes which exactly encompass certain subsets. E.g. If all cholera
diagnoses are provided, only '001 - Cholera' needs to be displayed, not all
subtypes.}

\item{brief}{single logical value, default is \code{FALSE}. If \code{TRUE},
the short description from the canonical CMS descriptions (included in
\code{data-raw}) will be used, otherwise the long description is used.}

\item{warn}{single logical value, default is \code{TRUE}, meaning that codes
which do not correspond to diagnoses, or to three-digit codes, will trigger
a warning.}

\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five
characters is given in a code, then the digits are greedily assigned to
hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}
}
\value{
data frame, or list of data frames, with fields for ICD9 code, name
  and description, derived from datamart lookup table
}
\description{
convert 'decimal' format (123.45 style) ICD9 codes into the name
  and description for human review there are official ICD9-CM data tables,
  not with conversion to decimal notation, but to the textual format.
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: explain Explain all ICD-9 codes in a list of vectors

\item \code{factor}: explain Explain factor of ICD-9 codes

\item \code{character}: explain character vector of ICD-9 codes
}}
\examples{
# by default, just show parent code and ignore children (428.0 not shown
# because 428 is present):
icd9ExplainShort(ahrqComorbid$CHF[1:3])
# same without condensing the list. In this case, 428.0 is shown:
icd9ExplainShort(ahrqComorbid$CHF[1:3], condense = TRUE)
icd9Explain(ahrqComorbid$CHF[1:3], brief = TRUE)
}
\seealso{
package comorbidities
}

