% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explaintable.R
\name{icd_explain_table}
\alias{icd_explain_table}
\alias{icd_explain_table.default}
\alias{icd_explain_table.icd10}
\alias{icd_explain_table.icd10cm}
\alias{icd_explain_table.icd9}
\alias{icd_explain_table.icd9cm}
\title{Explain ICD-9 and ICD-10 codes in English from decimal  (123.45 style),
Tabulates the decimal format alongside converted non-decimal format.}
\usage{
icd_explain_table(...)

\method{icd_explain_table}{default}(x, short_code = icd_guess_short(x),
  condense = FALSE, brief = TRUE, warn = TRUE, ...)

\method{icd_explain_table}{icd9}(...)

\method{icd_explain_table}{icd10}(...)

\method{icd_explain_table}{icd9cm}(x, short_code = icd_guess_short(x),
  condense = FALSE, brief = TRUE, warn = TRUE, ...)

\method{icd_explain_table}{icd10cm}(x, short_code = icd_guess_short(x),
  condense = FALSE, brief = TRUE, warn = TRUE, ...)
}
\arguments{
\item{...}{arguments passed on to other functions}

\item{x}{vector or other structure of ICD codes to explain in human language}

\item{short_code}{single logical value which determines whether the ICD-9
code provided is in short (\code{TRUE}) or decimal (\code{FALSE}) form.
Where reasonable, this is guessed from the input data.}

\item{condense}{single logical value which indicates whether to condense the
given set of ICD-9 codes by replacing subsets of codes with 'parent' codes
which exactly encompass certain subsets. E.g. If all cholera diagnoses are
provided, only '001 - Cholera' needs to be displayed, not all sub-types.}

\item{brief}{single logical value, default is \code{FALSE}. If \code{TRUE},
the short description from the canonical CMS descriptions (included in
\code{data-raw}) will be used, otherwise the long description is used.}

\item{warn}{single logical value, default is \code{TRUE}, meaning that codes
which do not correspond to diagnoses, or to three-digit codes, will trigger
a warning.}
}
\value{
data frame, or list of data frames, with fields for ICD-9 code, name
  and description. The ordering is in the same order as input.
}
\description{
Output is ordered in the same order as the input. A boolean column ismajor
indicates if the code is a parent Category.
}
\details{
A column for source year may be added in the future.

If the input x is of mixed type it will choose to convert by

majority of icd9 vs. icd19
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: explaining ICD codes from a character vector, guessing ICD version

\item \code{icd9}: from vector of ICD-9 codes.

\item \code{icd10}: from vector of ICD-10 codes.

\item \code{icd9cm}: explain character vector of ICD-9-CM codes

\item \code{icd10cm}: explain character vector of ICD1-10-CM codes
}}
\author{
Ed Lee

Ed Lee
}
\keyword{internal}

