% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{icd_filter_valid}
\alias{icd_filter_invalid}
\alias{icd_filter_valid}
\alias{icd_filter_valid.icd9}
\title{Filter ICD-9 codes by validity.}
\usage{
icd_filter_valid(...)

icd_filter_invalid(...)

\method{icd_filter_valid}{icd9}(x, icd_name = NULL, short_code = NULL,
  invert = FALSE)
}
\arguments{
\item{x}{data frame containing columns for visitId (which is the feault
name), icd9 (default for the icd9 code), and  maybe also a POA flag.}

\item{icd_name}{The column in the data frame which contains the ICD codes.
This is a character vector of length one. If it is \code{NULL}, \code{icd9}
will attempt to guess the column name, looking for progressively less
likely possibilities until it matche a single column. Failing this, it will
take the first column in the data frame. Specifying the column using this
argument avoids the guesswork.}

\item{short_code}{single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form. Where reasonable, this
is guessed from the input data.}

\item{invert}{single logical value, if TRUE will return invalid instead of valid rows.}

\item{icd9df}{Deprecated. Use \code{icd_df}}

\item{icd9Field}{Deprecated. Use \code{icd_name}}

\item{isShort}{Deprecated. Single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form. Where reasonable, this
is guessed from the input data.}
}
\description{
Filters a data.frame of patients for valid or invalid ICD-9 codes
}
\section{Methods (by class)}{
\itemize{
\item \code{icd9}: Filter data frame for valid ICD-9 codes
}}

