% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/real.R
\name{icd9GetBillable}
\alias{icd9GetBillable}
\alias{icd9GetBillableDecimal}
\alias{icd9GetBillableShort}
\alias{icd9GetNonBillable}
\alias{icd9GetNonBillableDecimal}
\alias{icd9GetNonBillableShort}
\alias{icd_get_billable}
\alias{icd_get_billable.icd9}
\alias{icd_get_billable.icd9cm}
\title{Get billable ICD codes}
\usage{
icd9GetBillable(icd9, isShort = icd9GuessIsShort(icd9), invert = FALSE,
  version = getLatestBillableVersion())

icd9GetBillableShort(icd9Short, version = getLatestBillableVersion())

icd9GetBillableDecimal(icd9Decimal, version = getLatestBillableVersion())

icd9GetNonBillableShort(icd9Short, version = getLatestBillableVersion())

icd9GetNonBillableDecimal(icd9Decimal, version = getLatestBillableVersion())

icd9GetNonBillable(icd9, isShort = icd9GuessIsShort(icd9),
  version = getLatestBillableVersion())

icd_get_billable(...)

\method{icd_get_billable}{icd9cm}(x, short_code = icd_guess_short(x),
  invert = FALSE, version = getLatestBillableVersion(), ...)

\method{icd_get_billable}{icd9}(...)
}
\description{
Get billable ICD codes, implicitly, this refers to an ICD
  implementation which is specialized for a country, typically for billing,
  e.g. ICD-9-CM in the USA.
}
\section{Methods (by class)}{
\itemize{
\item \code{icd9cm}: Get billable ICD-9-CM codes

\item \code{icd9}: Get billable ICD-9 codes, which is currently implemented assuming ICD-9-CM
}}

