% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/valid.R
\name{icd9GetInvalidDecimal}
\alias{icd9GetInvalidDecimal}
\alias{icd9GetInvalidShort}
\alias{icd_get_invalid}
\title{Returns subset of codes which are not in valid short_code or decimal format.}
\usage{
icd9GetInvalidDecimal(icd9Decimal)

icd9GetInvalidShort(icd9Short)

icd_get_invalid(x, short_code = icd_guess_short(icd9))
}
\description{
Returns subset of codes which are not in valid short_code or decimal format.
}
\section{Functions}{
\itemize{
\item \code{icd9GetInvalidDecimal}: Returns subset of codes which are not in valid decimal format.

\item \code{icd9GetInvalidShort}: Returns subset of codes which are not in valid short_code format.
}}

