% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/deprecated.R, R/valid.R
\name{icd_get_major.icd9}
\alias{icd9GetMajor}
\alias{icd9GetMajorShort}
\alias{icd_get_major}
\alias{icd_get_major.icd9}
\title{Returns major component of code (before decimal)}
\usage{
\method{icd_get_major}{icd9}(x, short_code)

icd9GetMajorShort(icd9Short)

icd9GetMajor(x, isShort)

icd_get_major(x)
}
\description{
Returns major component of code (before decimal)
}
\section{Methods (by class)}{
\itemize{
\item \code{icd9}: Get major part of ICD-9 code, i.e. first three
digits of numeric or V code, or first four digits of E code. This is the part
before the decimal, when a decimal point is used.
}}
\keyword{internal}
\keyword{manip}

