% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/valid.R
\name{icd_get_major.icd9}
\alias{icd_get_major}
\alias{icd_get_major.icd10}
\alias{icd_get_major.icd9}
\title{Get major part of an ICD code}
\usage{
\method{icd_get_major}{icd9}(x, short_code)

icd_get_major(x)

\method{icd_get_major}{icd10}(x)
}
\description{
Returns major component of codes, i.e. the part before the decimal,
or where the decimal would be.
}
\details{
For ICD-10, this is an initial implementation. If speed needed, then
  can re-use C++ ICD-9 version: just grabbing the first three characters,
  after all, and this is much easier in ICD-10 then ICD-9
}
\section{Methods (by class)}{
\itemize{
\item \code{icd9}: Get major part of ICD-9 code, i.e. first three
digits of numeric or V code, or first four digits of E code. This is the part
before the decimal, when a decimal point is used.

\item \code{icd10}: Get major part of an ICD-10 code
}}
\keyword{internal}
\keyword{manip}

