% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/valid.R
\name{icd9GetValid}
\alias{icd9GetInvalid}
\alias{icd9GetValid}
\alias{icd9GetValidDecimal}
\alias{icd9GetValidShort}
\alias{icd_get_valid}
\alias{icd_get_valid.character}
\alias{icd_get_valid.icd10}
\alias{icd_get_valid.icd10cm}
\alias{icd_get_valid.icd9}
\title{invalid subset of decimal or short_code ICD-9 codes}
\usage{
icd9GetValid(icd9, isShort = icd_guess_short(icd9))

icd9GetValidDecimal(icd9Decimal)

icd9GetValidShort(icd9Short)

icd9GetInvalid(icd9, isShort = icd_guess_short(icd9))

icd_get_valid(x, short_code = icd_guess_short(x))

\method{icd_get_valid}{character}(x, short_code = icd_guess_short(x))

\method{icd_get_valid}{icd9}(x, short_code = icd_guess_short(x))

\method{icd_get_valid}{icd10}(x, short_code = icd_guess_short(x))

\method{icd_get_valid}{icd10cm}(x, short_code = icd_guess_short(x))
}
\arguments{
\item{icd9}{is
a character vector or factor of ICD-9 codes. If fewer than five characters
is given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010".}

\item{isShort}{Deprecated. Single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form. Where reasonable, this
is guessed from the input data.}

\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five
characters is given in a code, then the digits are greedily assigned to
hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{x}{input vector of ICD codes}

\item{short_code}{single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form. Where reasonable, this
is guessed from the input data.}
}
\description{
Given vector of short_code or decimal ICD-9 codes, return (in the
  same format) those codes which are valid or invalid. Useful for generating
  error messages with the faulty codes if validation fails.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: get valid ICD codes from character vector, guessing ICD version

\item \code{icd9}: Get valid ICD-9 codes

\item \code{icd10}: Get valid ICD-10 codes

\item \code{icd10cm}: Get valid ICD-10-CM codes
}}
\keyword{internal}
\keyword{manip}

