% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess.R
\name{icd_guess_both}
\alias{icd_guess_both}
\title{Guess both ICD-9 vs ICD-10 or subtype, and whether short of long type.}
\usage{
icd_guess_both(x, short_code = NULL)
}
\description{
I think these are not quite orthogonal problems.
}
\keyword{internal}

