% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess.R
\name{icd_guess_short}
\alias{icd_guess_short}
\title{guess whether codes are \code{short_code} or \code{decimal_code}}
\usage{
icd_guess_short(x, short_code = NULL, test_n = 1000L, icd_name = NULL)
}
\value{
single logical value, \code{TRUE} if input data are predominantly
  \code{short_code} type. If there is some uncertainty, then return
  \code{NA}.
}
\description{
TODO: Partly implemented. Goal is to guess whether codes are
\code{short_code} or \code{decimal_code} form. Currently condense works, but
not with the \code{icd} look-up table currently in use. Of note, validation
is a bit different here, since we don't know the type until after we guess.
We could look for where both short_code and long are invalid, and otherwise
assume valid, even if the bulk are short_code. However, it may be more useful
to check validity after the guess.
}
\keyword{internal}

