% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess.R
\name{icd_guess_short_update}
\alias{icd_guess_short_update}
\title{Guess short vs decimal of ICD and update class}
\usage{
icd_guess_short_update(x, icd_name = get_icd_name(x), short_code = NULL)
}
\value{
the input data with appropriate ICD class set
}
\description{
Guesses whether the given ICD codes are short or long format,
  and set the class of the returned data according to the guess.
}
\keyword{internal}

