% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess.R
\name{icd_guess_version}
\alias{icd_guess_version}
\alias{icd_guess_version.character}
\title{Guess version of ICD}
\usage{
icd_guess_version(...)

\method{icd_guess_version}{character}(icd, short_code)
}
\description{
Guess version of ICD
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Guess version of ICD code from character vector
}}
\keyword{internal}

