% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess.R
\name{icd_guess_version_update}
\alias{icd_guess_version_update}
\title{Guess version of ICD and update class}
\usage{
icd_guess_version_update(x, short_code = icd_guess_short(x))
}
\arguments{
\item{x}{vector of ICD codes of some type}

\item{short_code}{single logical value which determines whether the ICD-9
code provided is in short (\code{TRUE}) or decimal (\code{FALSE}) form.
Where reasonable, this is guessed from the input data.}
}
\value{
the input data with appropriate ICD class set
}
\description{
Guesses the version of given ICD codes and sets the class of the
  returned data according to the guess.
}
\keyword{internal}
