% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess.R
\name{icd_guess_version_update}
\alias{icd_guess_version_update}
\title{Guess version of ICD and update class}
\usage{
icd_guess_version_update(x)
}
\value{
the input data with appropriate ICD class set
}
\description{
Guesses the version of given ICD codes and sets the class of the returned data according to the guess.
}
\keyword{internal}

