% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{icd_is_major}
\alias{icd_is_major}
\alias{icd_is_major.default}
\alias{icd_is_major.icd10}
\alias{icd_is_major.icd10cm}
\alias{icd_is_major.icd9}
\title{Check whether a code is major}
\usage{
icd_is_major(x)

\method{icd_is_major}{default}(x)

\method{icd_is_major}{icd10}(x)

\method{icd_is_major}{icd10cm}(x)

\method{icd_is_major}{icd9}(x)
}
\arguments{
\item{icd}{character vector of ICD codes.}
}
\description{
Check whether a code is major
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method which guesses version

\item \code{icd10}: check whether a code is an ICD-10 major

\item \code{icd10cm}: check whether a code is an ICD-10-CM major.
Currently uses \code{stringr} which uses \code{stringi} which should be
quite fast, but does suffer from handling Unicode, locales, etc.

\item \code{icd9}: check whether a code is an ICD-9 major
}}

\keyword{internal}
