% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/valid.R
\name{icd9IsValidMapping}
\alias{icd9GetInvalidMappingDecimal}
\alias{icd9GetInvalidMappingShort}
\alias{icd9IsValid}
\alias{icd9IsValidDecimal}
\alias{icd9IsValidMapping}
\alias{icd9IsValidMappingDecimal}
\alias{icd9IsValidMappingShort}
\alias{icd9IsValidShort}
\alias{icd_is_valid}
\alias{icd_is_valid.icd10}
\alias{icd_is_valid.icd10who}
\alias{icd_is_valid.icd9}
\title{Check whether ICD-9 codes are syntactically valid}
\usage{
icd9IsValidMapping(icd9Mapping, isShort)

icd9IsValidMappingShort(icd9Mapping)

icd9IsValidMappingDecimal(icd9Mapping)

icd9GetInvalidMappingShort(icd9Mapping)

icd9GetInvalidMappingDecimal(icd9Mapping)

icd9IsValid(icd9, isShort)

icd9IsValid(icd9, isShort)

icd9IsValidShort(icd9)

icd9IsValidDecimal(icd9Decimal)

icd_is_valid(icd, ...)

\method{icd_is_valid}{icd10}(icd, short_code = icd_guess_short(icd))

\method{icd_is_valid}{icd9}(x, short_code)

\method{icd_is_valid}{icd10who}(icd, short_code = icd_guess_short.icd10(icd))
}
\arguments{
\item{icd9Mapping}{named list containing vectors of icd9 codes. E.g. the AHRQ
comorbidities, contains \code{list(OBESE = c("2780", "27800", "27801",
"27803", "V8554", "79391", "64910", "64911", "64912", "64913", "64914",
"V8530", "V8531", "V8532", "V8533", "V8534", "V8535", "V8536", "V8537",
"V8538", "V8539", "V8541", "V8542", "V8543", "V8544", "V8545" ), DEPRESS =
c("3004", "30112", "3090", "3091", "311"))} amongst other longer groups.}

\item{isShort}{Deprecated. Single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form. Where reasonable, this
is guessed from the input data.}

\item{icd9}{is
a character vector or factor of ICD-9 codes. If fewer than five characters
is given in a code, then the digits are greedily assigned to hundreds, then
tens, then units, before the decimal parts. E.g. "10" becomes "010", not
"0010".}

\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five
characters is given in a code, then the digits are greedily assigned to
hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{icd}{An ICD-9 or 10 code. If the class is set to \code{'icd9'},
\code{'icd10'}, \code{'icd10cm'} etc then perform appropriate validation.}

\item{short_code}{single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form. Where reasonable, this
is guessed from the input data.}

\item{icd9Mapping}{named list containing vectors of icd9 codes. E.g. the AHRQ
comorbidities, contains \code{list(OBESE = c("2780", "27800", "27801",
"27803", "V8554", "79391", "64910", "64911", "64912", "64913", "64914",
"V8530", "V8531", "V8532", "V8533", "V8534", "V8535", "V8536", "V8537",
"V8538", "V8539", "V8541", "V8542", "V8543", "V8544", "V8545" ), DEPRESS =
c("3004", "30112", "3090", "3091", "311"))} amongst other longer groups.}

\item{icd9Mapping}{named list containing vectors of icd9 codes. E.g. the AHRQ
comorbidities, contains \code{list(OBESE = c("2780", "27800", "27801",
"27803", "V8554", "79391", "64910", "64911", "64912", "64913", "64914",
"V8530", "V8531", "V8532", "V8533", "V8534", "V8535", "V8536", "V8537",
"V8538", "V8539", "V8541", "V8542", "V8543", "V8544", "V8545" ), DEPRESS =
c("3004", "30112", "3090", "3091", "311"))} amongst other longer groups.}

\item{icd9Mapping}{named list containing vectors of icd9 codes. E.g. the AHRQ
comorbidities, contains \code{list(OBESE = c("2780", "27800", "27801",
"27803", "V8554", "79391", "64910", "64911", "64912", "64913", "64914",
"V8530", "V8531", "V8532", "V8533", "V8534", "V8535", "V8536", "V8537",
"V8538", "V8539", "V8541", "V8542", "V8543", "V8544", "V8545" ), DEPRESS =
c("3004", "30112", "3090", "3091", "311"))} amongst other longer groups.}

\item{icd9Mapping}{named list containing vectors of icd9 codes. E.g. the AHRQ
comorbidities, contains \code{list(OBESE = c("2780", "27800", "27801",
"27803", "V8554", "79391", "64910", "64911", "64912", "64913", "64914",
"V8530", "V8531", "V8532", "V8533", "V8534", "V8535", "V8536", "V8537",
"V8538", "V8539", "V8541", "V8542", "V8543", "V8544", "V8545" ), DEPRESS =
c("3004", "30112", "3090", "3091", "311"))} amongst other longer groups.}

\item{isShort}{Deprecated. Single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form. Where reasonable, this
is guessed from the input data.}

\item{short_code}{single logical value which determines whether the ICD-9 code
provided is in short (TRUE) or decimal (FALSE) form. Where reasonable, this
is guessed from the input data.}

\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{major}{character vector of 'major' part of ICD-9 codes, i.e. that part
which falls before the decimal point, in decimal notation. (In 5 digit
notation, the 'major' part is be the first three characters (with leading
zeroes), and includes V or E prefix. xyz}
}
\value{
logical vector with \code{TRUE} or \code{FALSE} for each icd9 code
  provided according to its validity
}
\description{
This does not check whether the code coresponds to a real ICD-9-CM billing
code, or parent grouping. For that, see \code{\link{icd9IsReal}}.
}
\details{
Factors are accepted, and since the validation is done with \code{grepl}
these are handled correctly.

Currently, there is a limitation on NA values. Calling with \code{NA} (which
is a logical vector of length one by default) fails, because it is not a
string. This is rarely of significance in real life, since the NA will be
part of a character vector of codes, and will therefore be cast already to
\code{NA_character}

\code{NA} values result in a return value of \code{FALSE}.

Leading zeroes in the decimal form are not ambiguous. Although
  integer ICD-9 codes could be intended by the user, there is a difference
  between 100, 100.0, 100.00. Therefore a warning is given if a numeric value
  is provided TODO: add default (when there is no class) which detected icd9
  vs 10 if possible. TODO: use "short_code" or "long" attribute if available to
  tighten validation, or guess if missing.

From WHO ICD-10 manual: "The basic ICD is a single coded list of
  three-character categories, each of which can be further divided into up to
  10 four-character subcategories. In place of the purely numeric coding
  system of previous revisions, the 10th revision uses an alphanumeric code
  with a letter in the first position and a number in the second, third and
  fourth positions. The fourth character follows a decimal point. Possible
  code numbers therefore range from A00.0 to Z99.9. The letter U is not used
  (see Section 2.4.7)."

  "Although not mandatory for reporting at the international level, most of
  the three-character categories are subdivided by means of a fourth, numeric
  character after a decimal point, allowing up to 10 subcategories. Where a
  three-character category is not subdivided, it is recommended that the
  letter ‘X’ be used to fill the fourth position, so that the codes are of a
  standard length for data-processing."

  officially, WHO standard goes to 3+1 digits, but there are officially
  sanctioned extensions which use additional digits, e.g. neurology,
  pediatrics, and of course ICD-10-CM.
}
\section{Methods (by class)}{
\itemize{
\item \code{icd10}: Test whether generic ICD-10 code is valid

\item \code{icd9}: Test whether generic ICD-10 code is valid

\item \code{icd10who}: Check whether ICD-10 WHO codes are valid
}}
\section{Three-digit validation}{
 \code{icd_is_valid_major} validates just the
  'major' three-digit part of an ICD-9 code. This can in fact be provided as
  a numeric, since there is no ambiguity. Numeric-only codes should be one to
  three digitis, V codes are followed by one or two digits, and E codes
  always by three digits between 800 and 999.
}

\section{Class}{
 S3 class of on object in R is just a vector. Attributes are
  lost with manipulation, with the exception of class: therefore, elements of
  the class vector are used to describe features of the data. If these are
  not present, the user may specify (e.g. decimal vs short_code type, ICD-9 vs
  ICD-10 WHO), but if they are, the correct functions are called without any
  guess work. There are overlapping namespaces for short_code vs decimal and ICD-9
  vs ICD-10 codes, so guessing is never going to be perfect.
}
\examples{
  icd_validShort(c("", "1", "22", "333", "4444", "123.45", "V",
                     "V2", "V34", "V567", "E", "E1", "E70", "E"))
  icd_validMajor(c("", "1", "22", "333", "4444", "123.45", "V",
                     "V2", "V34", "V567", "E", "E1", "E70", "E"))
}
\seealso{
\url{http://www.stata.com/users/wgould/icd9/icd9.hlp}
  url{http://www.sascommunity.org/wiki/Validate_the_format_of_ICD-9_codes}

Other ICD9 validation: \code{\link{icd_is_valid.map}}
}

