% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbid.R
\docType{data}
\name{icd_poa_choices}
\alias{icd_poa_choices}
\title{Present-on-admission flags}
\format{An object of class \code{character} of length 4.}
\usage{
icd_poa_choices
}
\description{
See \link{icd_filter_poa} for more details.
}
\examples{
  icd_poa_choices
}
\keyword{character}

