% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/convert.R, R/deprecated.R
\name{icd_short_to_decimal.icd9}
\alias{icd9DecimalToParts}
\alias{icd9DecimalToShort}
\alias{icd9ShortToDecimal}
\alias{icd9ShortToParts}
\alias{icd_short_to_decimal}
\alias{icd_short_to_decimal.icd9}
\title{Convert ICD from short to deimcal forms}
\usage{
\method{icd_short_to_decimal}{icd9}(x)

icd_short_to_decimal(x)

icd9DecimalToParts(icd9Decimal, minorEmpty = "")

icd9DecimalToShort(icd9Short)

icd9ShortToDecimal(icd9Short)

icd9ShortToParts(icd9Short)
}
\arguments{
\item{x}{ICD codes}

\item{icd9Decimal}{character vector of ICD-9 codes. If fewer than five
characters is given in a code, then the digits are greedily assigned to
hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{minorEmpty}{vector of length one, to be used in place of minor part of
zero. Defaults to ""}

\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}

\item{icd9Short}{is a character vector of ICD-9 codes. If fewer than
five characters is given in a code, then the digits are greedily assigned
to hundreds, then tens, then units, before the decimal parts. E.g. "10"
becomes "010", not "0010"}
}
\description{
Convert codes between short and decimal forms
}
\section{Methods (by class)}{
\itemize{
\item \code{icd9}: Convert ICD-9 codes from short to decimal form
}}
\keyword{internal}

