% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{icd_short_to_decimal}
\alias{icd_short_to_decimal}
\alias{icd_short_to_decimal.default}
\alias{icd_short_to_decimal.icd9}
\alias{icd_short_to_decimal.icd10}
\alias{icd_short_to_decimal.icd10cm}
\title{Convert ICD codes from short to decimal forms}
\usage{
icd_short_to_decimal(x)

\method{icd_short_to_decimal}{default}(x)

\method{icd_short_to_decimal}{icd9}(x)

\method{icd_short_to_decimal}{icd10}(x)

\method{icd_short_to_decimal}{icd10cm}(x)
}
\arguments{
\item{x}{ICD codes}
}
\description{
Convert codes between short and decimal forms
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: convert ICD codes of unknown type from short
to decimal format

\item \code{icd9}: convert ICD-9 codes from short to decimal
format

\item \code{icd10}: convert ICD-10 codes from short to decimal
format

\item \code{icd10cm}: convert ICD-10-CM code from short to decimal
format
}}

\keyword{internal}
