% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{is.icd9}
\alias{is.icd10}
\alias{is.icd10cm}
\alias{is.icd10who}
\alias{is.icd9}
\alias{is.icd9cm}
\alias{is.icd_decimal_code}
\alias{is.icd_long_data}
\alias{is.icd_map}
\alias{is.icd_short_code}
\alias{is.icd_wide_data}
\title{test ICD-related classes}
\usage{
is.icd9(x, strict = FALSE)

is.icd10(x, strict = FALSE)

is.icd9cm(x)

is.icd10cm(x)

is.icd10who(x)

is.icd_long_data(x)

is.icd_wide_data(x)

is.icd_short_code(x)

is.icd_decimal_code(x)

is.icd_map(x)
}
\arguments{
\item{x}{Any object which may have ICD-related classes set}

\item{strict}{logical value, if TRUE, will only match the type exactly; if \code{FALSE}}
}
\description{
currently no checks on correctness of the classes for these functions
}
\details{
TODO: could warn or fix if something is icd10cm or icd10who but not icd10

TODO: could warn or fix if something is icd10cm or icd10who but not icd10
}

