% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R, R/z-deprecated.R
\name{is.icd9}
\alias{is.icd9}
\alias{is.icd9cm}
\alias{is.icd9cm_pc}
\alias{is.icd9who}
\alias{is.icd10}
\alias{is.icd10cm}
\alias{is.icd10cm_pc}
\alias{is.icd10who}
\alias{is.icd10fr}
\alias{is.comorbidity_map}
\alias{is.icd_comorbidity_map}
\title{test ICD classes}
\usage{
is.icd9(x)

is.icd9cm(x)

is.icd9cm_pc(x)

is.icd9who(x)

is.icd10(x)

is.icd10cm(x)

is.icd10cm_pc(x)

is.icd10who(x)

is.icd10fr(x)

is.comorbidity_map(x)

is.icd_comorbidity_map(...)
}
\arguments{
\item{x}{Any object which may have ICD-related classes set}

\item{...}{arguments passed on to other functions}
}
\description{
This merely checks whether the given object is a certain type of ICD code, it
does no validation of any kind.
}
\section{Deprecated function names}{
 Future versions of \pkg{icd} will drop
  the \code{icd_} prefix. For example, \code{charlson} should be used in
  favor of \code{icd_charlson}. To distinguish \pkg{icd} function calls,
  consider using the prefix \code{icd::} instead, e.g., \code{icd::charlson}.
  Functions which specifically operate on either ICD-9 or ICD-10 codes or
  their sub-types will retain the prefix. E.g.
  \code{\link{icd9_comorbid_ahrq}}. \pkg{icd} specific classes also retain
  the prefix, e.g., \code{\link{icd_wide_data}}.
}

\examples{
# A character string is not itself an ICD code
is.icd9("100.1")
is_valid("100.1")
is.icd9(as.icd9cm("100.1"))
}
