% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R, R/class.R
\name{is.icd_short_diag}
\alias{is.icd10}
\alias{is.icd10cm}
\alias{is.icd9}
\alias{is.icd9cm}
\alias{is.icd_comorbidity_map}
\alias{is.icd_decimal_diag}
\alias{is.icd_long_data}
\alias{is.icd_short_diag}
\alias{is.icd_wide_data}
\title{test ICD-related classes}
\usage{
is.icd_short_diag(x, must_work = FALSE)

is.icd_decimal_diag(x, must_work = FALSE)

is.icd9(x)

is.icd10(x)

is.icd9cm(x)

is.icd10cm(x)

is.icd_long_data(x)

is.icd_wide_data(x)

is.icd_comorbidity_map(x)
}
\arguments{
\item{x}{Any object which may have ICD-related classes set}

\item{must_work}{single logical value, if \code{FALSE} (the default) this may
return \code{NULL} if the attribute is not present. If \code{TRUE}, then
either \code{TRUE} or \code{FALSE} is returned.}
}
\description{
currently no checks on correctness of the classes for these functions
}
\details{
\code{is.icd_short_diag} tests for presence of an attribute, not
  whether the code is a valid ICD code. If \code{must_work} is \code{TRUE}
  then \code{NULL} (i.e. no attribute set) returns \code{FALSE}, otherwise
  \code{NULL} is returned.
}

