% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{is.icd_long_data}
\alias{is.icd_long_data}
\alias{is.icd_wide_data}
\title{Return \code{TRUE} if \code{x} has the
  \code{icd_long_data} or \code{icd_wide_data} class.}
\usage{
is.icd_long_data(x)

is.icd_wide_data(x)
}
\description{
Return \code{TRUE} if \code{x} has the
  \code{icd_long_data} or \code{icd_wide_data} class.
}
\seealso{
\code{\link{icd_long_data}}
}
