% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{listTrimFlat}
\alias{listTrimFlat}
\title{trim null or empty values from a list}
\usage{
listTrimFlat(x)
}
\arguments{
\item{x}{list}
}
\value{
trimmed list
}
\description{
trim null or empty values from a list
}
\keyword{internal}

