% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{logical_to_binary}
\alias{logical_to_binary}
\title{encode TRUE as 1, and FALSE as 0 (integers)}
\usage{
logical_to_binary(x)
}
\arguments{
\item{x}{\code{data.frame} which may contain logical fields}
}
\value{
\code{data.frame} without logical fields
}
\description{
When saving data as text files for distribution, printing large amounts of
text containing \code{TRUE} and \code{FALSE} is inefficient. Convert to
binary takes more R memory, but allows more compact output
}
\keyword{internal}
\keyword{manip}

