% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{logical_to_binary}
\alias{logical_to_binary}
\title{encode TRUE as 1, and FALSE as 0 (integers)}
\usage{
logical_to_binary(x)
}
\arguments{
\item{x}{dataframe which may contain logical fields}
}
\value{
data frame without logical fields
}
\description{
when saving data as text files for distribution, printing large amounts of
text containing TRUE and FALSE is inefficient. Convert to binary takes more R
memory, but allows more compact output
}
\keyword{internal}
\keyword{manip}

