% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.R
\name{logicalToBinary}
\alias{logicalToBinary}
\title{encode TRUE as 1, and FALSE as 0 (integers)}
\usage{
logicalToBinary(x)
}
\arguments{
\item{x}{dataframe which may contain logical fields}
}
\value{
data frame without logical fields
}
\description{
when saving data as text files for distribution, printing large
  amounts of text containing TRUE and FALSE is inefficient. Convert to binary
  takes more R memory, but allows more compact output TODO: test
}
\details{
Taken from my
  \href{http://cran.r-project.org/web/packages/jwutil/index.html}{\code{jwutil}
  package}.#'
}
\keyword{internal}
\keyword{manip}

