% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/util.R
\name{match_rcpp}
\alias{match_rcpp}
\alias{fin}
\alias{fnin}
\alias{\%fin\%}
\alias{\%fnin\%}
\title{Faster match}
\usage{
match_rcpp(x, table)

fin(x, table)

fnin(x, table)

x \%fin\% table

x \%fnin\% table
}
\description{
Try Rcpp hashing (and simpler logic) compared to R's internal
  do_match and match5 morass. Lose the ability to use \code{incomparables}
  in initial implementation.
}
\section{Functions}{
\itemize{
\item \code{fin}: Use faster Rcpp matching for %in%

\item \code{fnin}: Use faster Rcpp matching for %nin%

\item \code{\%fin\%}: Use faster Rcpp matching for %in%

\item \code{\%fnin\%}: Use faster Rcpp matching for %nin%
}}

\keyword{internal}
