% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{order.icd9}
\alias{order.icd9}
\title{Get order of short-form ICD-9 codes}
\usage{
order.icd9(x)
}
\arguments{
\item{x}{vector or factor of ICD-9 codes}
}
\value{
vector of integers with length of the non-NA values in \code{x}
}
\description{
Puts E codes after V codes. \code{NA} values can't be ordered and are dropped
with a warning if found.
}
