% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-comorbid.R
\name{parse_ahrq_sas}
\alias{parse_ahrq_sas}
\title{parse AHRQ SAS code to get mapping}
\usage{
parse_ahrq_sas(sasPath = system.file("data-raw", "comformat2012-2013.txt",
  package = get_pkg_name()), save = FALSE, path = "data")
}
\arguments{
\item{sasPath}{single character string containing path or URL for some SAS
code. The source SAS code is stored in data-raw or at the location
specified in URL. This function is internal, since it is used to generate
data which ends up in the distributed package. However, the package user
can verify that the code creates the distributed R data. Also, changes to
the original SAS code can be used to regenerate the R data by the user,
without waiting for a package release.}

\item{condense}{Deprecated. The map is not condensed by default. If required,
it can reduced to a smaller set of codes using \code{icd_condense}.}

\item{save_data}{logical whether to save the result in the source tree.
Defaults to \code{FALSE}.}

\item{returnAll}{logical which, if TRUE, will result in the invisible return
of icd9_map_ahrq_all result, otherwise, icd9_map_ahrq is reutrned.}
}
\value{
invisibly returns the list of vectors, where a co-morbidity name is
  associated with a character vector of ICD-9 codes.
}
\description{
Takes the raw data taken directly from the AHRQ web site and
  parses into RData. It is then saved in the development tree data directory,
  so this is an internal function, used in generating the package itself!
}
\keyword{internal}

