% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_leaf_desc_icd9cm_v27}
\alias{parse_leaf_desc_icd9cm_v27}
\title{Parse billable codes for ICD-9-CM version 27}
\usage{
parse_leaf_desc_icd9cm_v27(offline = TRUE)
}
\arguments{
\item{offline}{single logical, if \code{TRUE} then don't pull the file from
internet, only return path and file name if the file already exists in
data-raw. This is helpful for testing without using the internet.}
}
\description{
These have a quirk which needs a different approach
}
\keyword{internal}
