% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/parse-rtf.R
\name{parseRtfLines}
\alias{parseRtfLines}
\alias{parse_rtf_lines}
\title{parse lines of RTF}
\usage{
parseRtfLines(lines, verbose = FALSE)

parse_rtf_lines(lines, verbose = FALSE)
}
\arguments{
\item{lines}{character vector containing RTF. Encoding?}

\item{verbose}{single logical value, defaults to \code{FALSE} in most functions.}
}
\value{
named character vector, with names being the ICD-9 codes, and the
  contents being the descriptions from the RTF source. Elsewhere I do this
  the other way around, but the tests are now wired for this layout. "Tidy"
  data would favour having an un-named two-column data frame.
}
\description{
parse a character vector containing RTF strings
}
\keyword{internal}

