% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parseIcd9Chapters}
\alias{parseIcd9Chapters}
\title{Read higher-level ICD-9 structure from a reliable web site}
\usage{
parseIcd9Chapters(year = NULL, save_data = FALSE, save = NULL)
}
\description{
This is rather slow, queries a web page repeatedly, which is
  both non-reproducible, and perhaps bad form. Aim to deprecate and replace
  with my own RTF parsing of canonical documents, which is now working
  reasonably well, at least for 'major' codes from 2015.
}
\details{
This is not almost obsolete. The only remaining use is the chapter
  and sub-chapter names and ranges, or validation.
}
\keyword{internal}

