% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-rtf.R
\name{parseRtfFifthDigitRanges}
\alias{parseRtfFifthDigitRanges}
\title{parse a row of RTF source data for ranges to apply fifth digit
  sub-classifications}
\usage{
parseRtfFifthDigitRanges(row_str, verbose = FALSE)
}
\description{
returns all the possible 5 digit codes encompassed by the given
  definition. This needs to be whittled down to just those matching fifth
  digits, but we haven't parsed them yet.
}
\examples{
row_str <- "The following fifth-digit subclassification is for use with category 493.0-493.2, 493.9:"
parseRtfFifthDigitRanges(row_str)

row_str <- "The following fifth-digit subclassification is for use with categories 67\\\\hich\\\\af1\\\\dbch\\\\af31505\\\\loch\\\\f1 8-679 to denote the current episode of care:"
parseRtfFifthDigitRanges(row_str)

row_str <- "The following fifth-digit subclassification is for use with category 711; valid digits are in [brackets] under each code. see list at beginning of chapter for definitions:"
parseRtfFifthDigitRanges(row_str)
}
\keyword{internal}

