% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbid.R, R/z-deprecated.R
\docType{data}
\name{poa_choices}
\alias{poa_choices}
\alias{icd_poa_choices}
\title{Present-on-admission flags}
\format{An object of class \code{character} of length 4.}
\usage{
poa_choices

icd_poa_choices
}
\description{
See \link{filter_poa} for more details.
}
\section{Deprecated function names}{
 Future versions of \pkg{icd} will drop
  the \code{icd_} prefix. For example, \code{charlson} should be used in
  favor of \code{icd_charlson}. To distinguish \pkg{icd} function calls,
  consider using the namespace prefix \code{icd::} instead, e.g.,
  \code{icd::charlson}. Functions which specifically operate on either ICD-9
  or ICD-10 codes or their subtypes will retain the prefix. E.g.
  \code{\link{icd9_comorbid_ahrq}}. \pkg{icd} specific classes also retain
  the prefix, e.g., \code{\link{icd_wide_data}}.
}

\examples{
  poa_choices
}
\keyword{character}
\keyword{datasets}
