% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rbind_with_empty}
\alias{rbind_with_empty}
\title{fast expand of logical matrix to add rows filled with false}
\usage{
rbind_with_empty(a, b_rows)
}
\description{
fast expand of logical matrix to add rows filled with false
}
\keyword{internal}
