% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{refactor_worker}
\alias{refactor_worker}
\alias{refactor_narm_worker}
\title{Re-generate a factor with new levels, without doing string matching}
\usage{
refactor_worker(x, new_levels, exclude_na, validate)

refactor_narm_worker(x, new_levels, validate)
}
\description{
This is called by an R wrapper. There is an \code{na.rm} version,
too. Some work simply to mirror behavior of \code{base::factor}, e.g. when a
level is not available, but NA level is available, NA is inserted into the
integer vector, not an index to the NA level.
}
\section{Functions}{
\itemize{
\item \code{refactor_narm_worker}: Drop all `NA` values from levels and values
}}

\keyword{internal}
\keyword{manip}
