% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{refactor_worker}
\alias{refactor_worker}
\title{Re-generate a factor with new levels, without doing string matching}
\usage{
refactor_worker(x, new_levels, na_rm, exclude_na)
}
\keyword{internal}
\keyword{manip}
