% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{regexec32}
\alias{regexec32}
\title{\code{regexec} which accepts \code{perl} argument even in older R}
\usage{
regexec32(pattern, text, ...)
}
\description{
TODO: deprecate this as we won't support ancient R versions indefinitely.
}
\keyword{internal}
