% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-rtf.R
\name{rtf_strip}
\alias{rtf_strip}
\title{Strip RTF}
\usage{
rtf_strip(x, ...)
}
\arguments{
\item{x}{vector of character strings containing RTF}
}
\description{
Take a vector of character strings containing RTF, replace each \\tab with a
space and eradicate all other RTF symbols
}
\details{
just for \\tab, replace with space, otherwise, drop RTF tags entirely
}
\keyword{internal}
\keyword{manip}
