% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sas.R
\name{sasParseAssignments}
\alias{sasParseAssignments}
\title{get assignments from a character string strings.}
\usage{
sasParseAssignments(x, stripWhiteSpace = TRUE, stripQuotes = TRUE)
}
\arguments{
\item{x}{is a character string containing space delimited assignments, in SAS
declaration format.}

\item{stripWhiteSpace}{will strip all whitespace from the returned values}

\item{stripQuotes}{will strip all double quotation marks from the returned
values}
}
\value{
list with each list item containing a matrix of "char ranges",
  "assigned value" pairs
}
\description{
#form is aaa-bbb, ccc-ddd, eee, etc. = "name" abc-def, ghi,
  etc. = "anothername" there is no delimiter between each assignment. '
}
\keyword{internal}
\keyword{list}
\keyword{programming}

