% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-file.R
\name{save_in_data_dir}
\alias{save_in_data_dir}
\title{Save given variable in package data directory}
\usage{
save_in_data_dir(var_name, suffix = "")
}
\arguments{
\item{var_name}{character or symbol, e.g. \code{"myvar"} or \code{myvar},
either of which would find \code{myvar} in the parent environment, and save
it as \code{myvar.RData} in \code{package_root/data}.}

\item{suffix}{character scalar}
}
\value{
invisibly returns the data
}
\description{
File will be named \code{\emph{var_name}.RData} with an optional suffix
before \code{.RData}
}
\keyword{internal}

