% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{save_in_data_dir}
\alias{save_in_data_dir}
\title{Save given variable in package data directory}
\usage{
save_in_data_dir(var, suffix = "")
}
\arguments{
\item{var}{character or symbol, e.g. "myvar" or \code{myvar}, either of which
would find \code{myvar} in the parent environment, and save it as
\code{myvar.RData} in \code{package_root/data}.}

\item{suffix}{character scalar}
}
\description{
File is named varname.RData with an optional suffix before .RData
}
\keyword{internal}

