% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-icd10.R
\name{scrape_icd10_who}
\alias{scrape_icd10_who}
\title{scrape WHO web site for ICD-10 codes}
\usage{
scrape_icd10_who(sleep_secs = 3)
}
\description{
javascript only (at least in recent years), so can't just get the HTML.
Thanks guys.
}
\details{
also requires phatomjs (which is Mac/Win only?) to render the javascript
}
\keyword{internal}

