% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-test.R
\name{set_full_test_options}
\alias{set_full_test_options}
\title{Set test options to do everything}
\usage{
set_full_test_options()
}
\description{
Default without setting options is for slow tests to be skipped. This
function explicitly sets options to do slow tests.
}
\keyword{debugging}
\keyword{internal}

