% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{icd9}
\alias{icd10}
\alias{icd10cm}
\alias{icd10who}
\alias{icd9}
\alias{icd9cm}
\alias{icd_comorbidity_map}
\alias{icd_decimal_code}
\alias{icd_long_data}
\alias{icd_short_code}
\alias{icd_wide_data}
\title{construct ICD-9 data types}
\usage{
icd9(x)

icd9cm(x)

icd10(x)

icd10cm(x)

icd10who(x)

icd_long_data(x)

icd_wide_data(x)

icd_short_code(x, warn = TRUE)

icd_decimal_code(x, warn = TRUE)

icd_comorbidity_map(x)
}
\arguments{
\item{x}{object to set class \code{icd9}}

\item{warn}{single logical value, if \code{TRUE} will gives warning when
converting between types. ICD-9 to ICD-10 will cause an error regardless.}
}
\description{
Takes an R structure and sets class to an ICD type. In the case
  of ICD-9 and ICD-10 codes, if a particular sub-type is set, e.g. ICD-9-CM
  (/code{icd9cm}), then an ICD-9 class (/code{icd9}) is also set.
}
\details{
This, I think, should take priority over ICD-9 vs ICD-10 when doing
  S3 dispatching
}

