% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R, R/z-deprecated.R
\name{short_to_decimal}
\alias{short_to_decimal}
\alias{short_to_decimal.default}
\alias{short_to_decimal.icd9}
\alias{short_to_decimal.icd10}
\alias{short_to_decimal.icd10cm}
\alias{icd_short_to_decimal.default}
\alias{icd_short_to_decimal.icd10}
\alias{icd_short_to_decimal.icd10cm}
\alias{icd_short_to_decimal.icd9}
\alias{icd_sort}
\alias{icd_short_to_decimal}
\title{Convert ICD codes from short to decimal forms}
\usage{
short_to_decimal(x)

\method{short_to_decimal}{default}(x)

\method{short_to_decimal}{icd9}(x)

\method{short_to_decimal}{icd10}(x)

\method{short_to_decimal}{icd10cm}(x)

icd_short_to_decimal.default(...)

icd_short_to_decimal.icd10(...)

icd_short_to_decimal.icd10cm(...)

icd_short_to_decimal.icd9(...)

icd_sort(...)

icd_short_to_decimal(...)
}
\arguments{
\item{x}{ICD codes}

\item{...}{arguments passed on to other functions}

\item{...}{arguments passed on to other functions}

\item{...}{arguments passed on to other functions}
}
\description{
Convert codes between short and decimal forms
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: convert ICD codes of unknown type from short
to decimal format

\item \code{icd9}: convert ICD-9 codes from short to decimal
format

\item \code{icd10}: convert ICD-10 codes from short to decimal
format

\item \code{icd10cm}: convert ICD-10-CM code from short to decimal
format
}}

\section{Deprecated function names}{
 Future versions of \pkg{icd} will drop
  the \code{icd_} prefix. For example, \code{charlson} should be used in
  favor of \code{icd_charlson}. To distinguish \pkg{icd} function calls,
  consider using the prefix \code{icd::} instead, e.g., \code{icd::charlson}.
  Functions which specifically operate on either ICD-9 or ICD-10 codes or
  their sub-types will retain the prefix. E.g.
  \code{\link{icd9_comorbid_ahrq}}. \pkg{icd} specific classes also retain
  the prefix, e.g., \code{\link{icd_wide_data}}.
}

\seealso{
Other ICD data conversion: \code{\link{comorbid_df_to_mat}},
  \code{\link{comorbid_mat_to_df}}, \code{\link{convert}},
  \code{\link{decimal_to_short}},
  \code{\link{long_to_wide}}, \code{\link{wide_to_long}}
}
\concept{ICD data conversion}
\keyword{internal}
\keyword{manip}
