% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sort_std}
\alias{sort_std}
\title{Sort using STL}
\usage{
sort_std(x)
}
\description{
if compiler flags and standard library support is available (only tested on glibc), OpenMP, then this
  will use a parallel sort algorithm which is significantly faster. It doesn't however deal with NA values.
  TODO: handle NA values.
}
\examples{
\dontrun{
pts <- icd9:::randomPatients(1e7)
microbenchmark::microbenchmark(sort_std(pts$icd9), sort(pts$icd9), times = 5)
# four times faster on 4 real core (8 with HT) machine.
}
}

