% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{str_match_all}
\alias{str_match_all}
\title{return all matches for regex}
\usage{
str_match_all(string, pattern, ...)
}
\description{
substitute for removed stringr function. \code{perl} is taken out if not
supported, allows compatibility with older versions of R.
}
\keyword{internal}
