% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{str_match_all}
\alias{str_match_all}
\title{return all matches for regular expression}
\usage{
str_match_all(string, pattern, ...)
}
\description{
return all matches for regular expression
}
\keyword{internal}
\keyword{manip}
