% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{str_pair_match}
\alias{str_pair_match}
\title{Match pairs of strings to get named vector}
\usage{
str_pair_match(string, pattern, pos, swap = FALSE)
}
\arguments{
\item{string}{vector of strings}

\item{pattern}{vector of regular expression which should match exactly two
strings for each element in \code{stringr}. If \code{pos} is specified,
this rule is relaxed.}

\item{pos}{optional pair of integers with positions of the first and second
desired matches, when multiple matches are made by the regular expression}

\item{swap}{logical scalar, whether to swap the names and values. Default is
not to swap, so the first match becomes the name.}
}
\description{
Match a character vector against a regular expression with at least two
parenthetic groupings, returning named vector.
}
\keyword{internal}

