% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{str_pair_match}
\alias{str_pair_match}
\title{Match strings to get named vector}
\usage{
str_pair_match(string, pattern, swap = FALSE, dropEmpty = FALSE,
  pos = c(1, 2), warn_pattern = TRUE, ...)
}
\arguments{
\item{swap}{logical scalar, whether to swap the names and values. Default is
not to swap, so the first match becomes the name.}

\item{warn_pattern}{logical, if the pattern has multiple parentheses,
optinally don't warn if we are forced to choose just a pair of the
sub-matches}
}
\description{
match a character vector against a regex with a pair of parentheses,
returning named vector
}
\keyword{internal}

