% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{strip}
\alias{strip}
\title{Strip character(s) from character vector}
\usage{
strip(x, pattern = " ", use_bytes = TRUE)
}
\arguments{
\item{x}{character vector}

\item{pattern}{passed to \code{gsub} default is " "}

\item{use_bytes}{single logical passed to \code{base::gsub}, default is the
slightly quicker \code{TRUE}}
}
\value{
character vector of same length as input
}
\description{
After benchmarking, \code{gsub} is probably quicker than
\code{stringr}/\code{stringi}. For comorbidity processing.
}
\keyword{internal}
