% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{[[.icd_comorbidity_map}
\alias{[[.icd_comorbidity_map}
\title{extract vector of codes from an ICD comorbidity map}
\usage{
\method{[[}{icd_comorbidity_map}(x, index, ...)
}
\description{
Equivalent to a list, but preserves class of extracted vector
}

