% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{subset_icd}
\alias{[.icd10}
\alias{[.icd9}
\alias{[[.icd10}
\alias{[[.icd9}
\alias{subset_icd}
\title{extract subset from ICD data}
\usage{
\method{[}{icd9}(x, ...)

\method{[[}{icd9}(x, ...)

\method{[}{icd10}(x, ...)

\method{[[}{icd10}(x, ...)
}
\description{
exactly the same as using x[n] or x[[n]] but preserves the ICD
  classes in result
}

