% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{trim}
\alias{trim}
\title{Trim leading and trailing white space}
\usage{
trim(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector
}
\description{
\code{NA} is accepted and returned, probably as \code{NA_character_}
}
\keyword{internal}
\keyword{manip}
