% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip_single.R
\name{unzip_single}
\alias{unzip_single}
\title{unzip a single file from URL}
\usage{
unzip_single(url, file_name, save_path)
}
\arguments{
\item{url}{URL of a zip file}

\item{file_name}{file name of the resource within the zip file}

\item{save_path}{file path to save the first file from the zip}
}
\description{
take a single file from zip located at a given URL, unzip into temporary
directory, and copy to the given \code{save_path}
}
\keyword{internal}
