% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{uranium_pathology}
\alias{uranium_pathology}
\title{United States Transuranium & Uranium Registries}
\description{
This is an ICD-10 data set (not ICD-10-CM) with mortality from the United States
Transuranium & Uranium Registries, published in the public domain.
}
\seealso{
\code{\link[icd.data]{ls_icd_data}}
}
\keyword{datasets}
