% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd-cm-ver.R
\name{with_icd10cm_version}
\alias{with_icd10cm_version}
\title{Evaluate code with a particular version of ICD-10-CM}
\usage{
with_icd10cm_version(ver, code)
}
\arguments{
\item{ver}{Character version of ICD-10-CM to use, corresponding to the suffix
of the \code{data.frame} name, e.g. for 2019 ICD-10-CM, use
\code{"icd10cm2019"} for Dutch 2014 ICD-CM translations, use
'icd10cm2014_nl"}

\item{code}{Code block to execute}
}
\description{
Temporarily sets and restores the option \code{icd.data.icd10cm_active_ver}
}
